% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{add_group}
\alias{add_group}
\title{Add a reserved grouping variable if missing}
\usage{
add_group(x)
}
\arguments{
\item{x}{A data.table}
}
\value{
A data table with a \code{.group} variable
}
\description{
Add a reserved grouping variable if missing
}
\seealso{
Functions used for checking inputs
\code{\link{check_by}()},
\code{\link{check_dates}()},
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
