% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{add_max_observed_delay}
\alias{add_max_observed_delay}
\title{Add maximum observed delay}
\usage{
add_max_observed_delay(new_confirm, observation_indicator = NULL)
}
\arguments{
\item{new_confirm}{A data.table containing the columns: "reference_date",
"delay", ".group", "new_confirm", and "max_obs_delay".
As produced by \code{\link[=enw_preprocess_data]{enw_preprocess_data()}} in the \code{new_confirm} output with the
addition of the "max_obs_delay" column as produced by
\code{\link[=add_max_observed_delay]{add_max_observed_delay()}}.}

\item{observation_indicator}{A character string specifying the column name
in \code{new_confirm} that indicates whether an observation is observed or not.
This column should be a logical vector. If NULL (default), all observations
are considered observed.}
}
\value{
A data.table with the original columns of \code{new_confirm} and an
additional "max_obs_delay" column representing the maximum observed delay
for each group and reference date. If an observation indicator is provided,
unobserved data will have a "max_obs_delay" value of -1.
}
\description{
This function calculates and adds the maximum observed delay for each group
and reference date in the provided dataset. It first checks the validity of
the observation indicator and then computes the maximum delay. If an
observation indicator is provided, it further adjusts the maximum observed
delay for unobserved data to be negative 1 (indicating no maximum observed).
}
\seealso{
Helper functions for model modules:
\code{\link[=add_pmfs]{add_pmfs()}},
\code{\link[=convolution_matrix]{convolution_matrix()}},
\code{\link[=delay_only_ltotal]{delay_only_ltotal()}},
\code{\link[=delay_only_total]{delay_only_total()}},
\code{\link[=enw_dayofweek_structural_reporting]{enw_dayofweek_structural_reporting()}},
\code{\link[=enw_reference_by_report]{enw_reference_by_report()}},
\code{\link[=enw_reps_with_complete_refs]{enw_reps_with_complete_refs()}},
\code{\link[=enw_structural_reporting_metadata]{enw_structural_reporting_metadata()}},
\code{\link[=extract_obs_metadata]{extract_obs_metadata()}},
\code{\link[=extract_sparse_matrix]{extract_sparse_matrix()}},
\code{\link[=latest_obs_as_matrix]{latest_obs_as_matrix()}}
}
\concept{modelmodulehelpers}
