% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_rolling_sum}
\alias{aggregate_rolling_sum}
\title{Internal function to perform rolling sum aggregation}
\usage{
aggregate_rolling_sum(dt, internal_timestep, by = NULL)
}
\arguments{
\item{dt}{A \code{data.table} to be aggregated.}

\item{internal_timestep}{An integer indicating the period over which to
aggregate.}

\item{by}{A character vector specifying the columns to aggregate by.}
}
\value{
A modified data.table with aggregated observations.
}
\description{
This function takes a data.table and applies a rolling sum over a given
timestep,
aggregating by specified columns. It's particularly useful for aggregating
observations over certain periods.
}
\seealso{
Utility functions
\code{\link{coerce_date}()},
\code{\link{coerce_dt}()},
\code{\link{date_to_numeric_modulus}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_unset_cache}()},
\code{\link{get_internal_timestep}()},
\code{\link{is.Date}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
