% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{build_ord_obs}
\alias{build_ord_obs}
\title{Build the ord_obs \code{data.table}.}
\usage{
build_ord_obs(obs, max_delay, internal_timestep, timestep, nowcast = NULL)
}
\arguments{
\item{obs}{Observations as pulled from \code{nowcast$latest[[1]]}.}

\item{max_delay}{Whole number representing the maximum delay
in units of the timestep.}

\item{internal_timestep}{The internal timestep in days.}

\item{timestep}{The timestep to be used. This can be a string
("day", "week", "month") or a numeric whole number representing
the number of days.}

\item{nowcast}{If getting posterior samples, a data frame with
a `.draws`` column to get the draws from, as pulled from the fit
attribute of a nowcast.}
}
\value{
A \code{data.table}.
}
\description{
Build the ord_obs \code{data.table}.
}
\seealso{
Functions used for postprocessing of model fits:
\code{\link[=.check_primarycensored]{.check_primarycensored()}},
\code{\link[=.delay_draw_columns]{.delay_draw_columns()}},
\code{\link[=.discretise_parametric_pmf]{.discretise_parametric_pmf()}},
\code{\link[=enw_add_latest_obs_to_nowcast]{enw_add_latest_obs_to_nowcast()}},
\code{\link[=enw_nowcast_samples]{enw_nowcast_samples()}},
\code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}},
\code{\link[=enw_posterior]{enw_posterior()}},
\code{\link[=enw_posterior_delay]{enw_posterior_delay()}},
\code{\link[=enw_pp_summary]{enw_pp_summary()}},
\code{\link[=enw_quantiles_to_long]{enw_quantiles_to_long()}},
\code{\link[=enw_summarise_samples]{enw_summarise_samples()}},
\code{\link[=subset_obs]{subset_obs()}}
}
\concept{postprocess}
