% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_by}
\alias{check_by}
\title{Check by variables are present in the data}
\usage{
check_by(obs, by = c())
}
\arguments{
\item{obs}{An observation \code{data.frame}.}

\item{by}{A character vector of variables to group by.}
}
\description{
Check by variables are present in the data
}
\seealso{
Functions used for checking inputs
\code{\link{add_group}()},
\code{\link{check_dates}()},
\code{\link{check_group}()},
\code{\link{check_max_delay}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
