% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cmdstanr}
\alias{check_cmdstanr}
\title{Check that cmdstanr is installed}
\usage{
check_cmdstanr()
}
\value{
NULL (called for side effect of raising an error if cmdstanr
is not installed)
}
\description{
Checks that the \code{cmdstanr} package is available and
provides a helpful error message if not.
}
\keyword{internal}
