% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dates}
\alias{check_dates}
\title{Check Report and Reference Dates are present}
\usage{
check_dates(obs)
}
\arguments{
\item{obs}{A data.frame to check for the required date variables}
}
\value{
Returns the input data.frame with dates converted to date format
if not already.
}
\description{
Check Report and Reference Dates are present
}
\seealso{
Functions used for checking inputs
\code{\link{check_quantiles}()}
}
\concept{check}
