% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_environment_setting}
\alias{check_environment_setting}
\title{Check environment setting}
\usage{
check_environment_setting(x)
}
\arguments{
\item{x}{The environment variable to be checked.}
}
\value{
Logical value indicating whether the environment variable is not set
(either null or an empty string).
}
\description{
This internal function checks whether a given environment variable is set or
not. It returns \code{TRUE} if the variable is either null or an empty string,
indicating that the environment variable is not set. Otherwise, it returns
\code{FALSE}.
}
\keyword{internal}
