% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_module}
\alias{check_module}
\title{Check a model module contains the required components}
\usage{
check_module(module)
}
\arguments{
\item{module}{A model module. For example \code{\link[=enw_expectation]{enw_expectation()}}.}
}
\description{
Check a model module contains the required components
}
\seealso{
Functions used for checking inputs:
\code{\link[=check_design_matrix_sparsity]{check_design_matrix_sparsity()}},
\code{\link[=check_group]{check_group()}},
\code{\link[=check_group_date_unique]{check_group_date_unique()}},
\code{\link[=check_max_delay]{check_max_delay()}},
\code{\link[=check_modules_compatible]{check_modules_compatible()}},
\code{\link[=check_numeric_timestep]{check_numeric_timestep()}},
\code{\link[=check_observation_indicator]{check_observation_indicator()}},
\code{\link[=check_quantiles]{check_quantiles()}},
\code{\link[=check_timestep]{check_timestep()}},
\code{\link[=check_timestep_by_date]{check_timestep_by_date()}},
\code{\link[=check_timestep_by_group]{check_timestep_by_group()}}
}
\concept{check}
