% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_modules_compatible}
\alias{check_modules_compatible}
\title{Check that model modules have compatible specifications}
\usage{
check_modules_compatible(modules)
}
\arguments{
\item{modules}{A list of model modules.}
}
\description{
Check that model modules have compatible specifications
}
\seealso{
Functions used for checking inputs:
\code{\link[=check_design_matrix_sparsity]{check_design_matrix_sparsity()}},
\code{\link[=check_group]{check_group()}},
\code{\link[=check_group_date_unique]{check_group_date_unique()}},
\code{\link[=check_max_delay]{check_max_delay()}},
\code{\link[=check_module]{check_module()}},
\code{\link[=check_numeric_timestep]{check_numeric_timestep()}},
\code{\link[=check_observation_indicator]{check_observation_indicator()}},
\code{\link[=check_quantiles]{check_quantiles()}},
\code{\link[=check_timestep]{check_timestep()}},
\code{\link[=check_timestep_by_date]{check_timestep_by_date()}},
\code{\link[=check_timestep_by_group]{check_timestep_by_group()}}
}
\concept{check}
