% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_numeric_timestep}
\alias{check_numeric_timestep}
\title{Check Numeric Timestep}
\usage{
check_numeric_timestep(dates, date_var, timestep, exact = TRUE)
}
\arguments{
\item{dates}{Vector of Date class representing dates.}

\item{date_var}{The variable in \code{obs} representing dates.}

\item{timestep}{Numeric timestep for date difference.}

\item{exact}{Logical, if \code{TRUE}, checks if all differences exactly match the
timestep. If \code{FALSE}, checks if the sum of the differences modulo the
timestep equals zero. Default is \code{TRUE}.}
}
\value{
This function is used for its side effect of stopping if the check
fails. If the check passes, the function returns invisibly.
}
\description{
This function verifies if the difference in numeric dates in the provided
observations corresponds to the provided timestep.
}
\seealso{
Functions used for checking inputs:
\code{\link[=check_design_matrix_sparsity]{check_design_matrix_sparsity()}},
\code{\link[=check_group]{check_group()}},
\code{\link[=check_group_date_unique]{check_group_date_unique()}},
\code{\link[=check_max_delay]{check_max_delay()}},
\code{\link[=check_module]{check_module()}},
\code{\link[=check_modules_compatible]{check_modules_compatible()}},
\code{\link[=check_observation_indicator]{check_observation_indicator()}},
\code{\link[=check_quantiles]{check_quantiles()}},
\code{\link[=check_timestep]{check_timestep()}},
\code{\link[=check_timestep_by_date]{check_timestep_by_date()}},
\code{\link[=check_timestep_by_group]{check_timestep_by_group()}}
}
\concept{check}
