% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_renviron_for_cache}
\alias{check_renviron_for_cache}
\title{Check \code{.Renviron} for cache location setting}
\usage{
check_renviron_for_cache(environ)
}
\arguments{
\item{environ}{A list containing the contents of the \code{.Renviron} file and
its path. This is the output of the \code{\link[=get_renviron_contents]{get_renviron_contents()}} function.}
}
\description{
Check \code{.Renviron} for cache location setting
}
\keyword{internal}
