% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_timestep_by_group}
\alias{check_timestep_by_group}
\title{Check timestep by group}
\usage{
check_timestep_by_group(obs, date_var, timestep = "day", exact = TRUE)
}
\arguments{
\item{obs}{Any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}

\item{date_var}{The variable in \code{obs} representing dates.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week") or a numeric whole number representing the number of days.
Note that "month" is not currently supported in user-facing functions
and will throw an error if used.}

\item{exact}{Logical, if \code{TRUE}, checks if all differences exactly match the
timestep. If \code{FALSE}, checks if the sum of the differences modulo the
timestep equals zero. Default is \code{TRUE}.}
}
\value{
This function is used for its side effect of checking the timestep
by group in \code{obs}. If the check passes for all groups, the function
returns invisibly. Otherwise, it stops and returns an error message.
}
\description{
This function verifies if the difference in dates within each group in the
provided observations corresponds to the provided timestep. This check is
performed for the specified \code{date_var} and for each group in \code{obs}.
}
\seealso{
Functions used for checking inputs:
\code{\link[=check_design_matrix_sparsity]{check_design_matrix_sparsity()}},
\code{\link[=check_group]{check_group()}},
\code{\link[=check_group_date_unique]{check_group_date_unique()}},
\code{\link[=check_max_delay]{check_max_delay()}},
\code{\link[=check_module]{check_module()}},
\code{\link[=check_modules_compatible]{check_modules_compatible()}},
\code{\link[=check_numeric_timestep]{check_numeric_timestep()}},
\code{\link[=check_observation_indicator]{check_observation_indicator()}},
\code{\link[=check_quantiles]{check_quantiles()}},
\code{\link[=check_timestep]{check_timestep()}},
\code{\link[=check_timestep_by_date]{check_timestep_by_date()}}
}
\concept{check}
