% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_date}
\alias{coerce_date}
\title{Coerce Dates}
\usage{
coerce_date(dates = NULL)
}
\arguments{
\item{dates}{A vector-like input, which the function attempts
to coerce via \code{\link[data.table:as.IDate]{data.table::as.IDate()}}. Defaults to NULL.}
}
\value{
An \link[data.table]{IDate} vector.
}
\description{
Provides consistent coercion of inputs to
\link[data.table]{IDate} with error handling
}
\details{
If any of the elements of \code{dates} cannot be coerced,
this function will result in an error, indicating all indices
which cannot be coerced to \link[data.table]{IDate}.

Internal methods of \link{epinowcast} assume dates are represented
as \link[data.table]{IDate}.
}
\examples{
# works
coerce_date(c("2020-05-28", "2020-05-29"))
# does not, indicates index 2 is problem
tryCatch(
  coerce_date(c("2020-05-28", "2020-o5-29")),
  error = function(e) {
    print(e)
  }
)
}
\seealso{
Utility functions:
\code{\link[=coerce_dt]{coerce_dt()}},
\code{\link[=date_to_numeric_modulus]{date_to_numeric_modulus()}},
\code{\link[=enw_get_data]{enw_get_data()}},
\code{\link[=enw_rolling_sum]{enw_rolling_sum()}},
\code{\link[=get_internal_timestep]{get_internal_timestep()}},
\code{\link[=is.Date]{is.Date()}},
\code{\link[=stan_fns_as_string]{stan_fns_as_string()}}
}
\concept{utils}
