% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{construct_re}
\alias{construct_re}
\title{Constructs random effect terms}
\usage{
construct_re(re, data)
}
\arguments{
\item{re}{A random effect as defined using \code{\link[=re]{re()}} which itself takes
random effects specified in a model formula using the \link[lme4]{lme4}
syntax.}

\item{data}{A \code{data.frame} of observations used to define the
random effects. Must contain the variables specified in the
\code{\link[=re]{re()}} term.}
}
\value{
A list containing the transformed data ("data"),
fixed effects terms ("terms") and a  \code{data.frame} specifying
the random effect structure between these terms (\code{effects}). Note
that if the specified random effect was not a factor it will have been
converted into one.
}
\description{
Constructs random effect terms
}
\examples{
# Simple examples
form <- epinowcast:::parse_formula(~ 1 + (1 | day_of_week))
data <- enw_example("prepr")$metareference[[1]]
random_effect <- re(form$random[[1]])
epinowcast:::construct_re(random_effect, data)

# A more complex example
form <- epinowcast:::parse_formula(
  ~ 1 + disp + (1 + gear | cyl) + (0 + wt | am)
)
random_effect <- re(form$random[[1]])
epinowcast:::construct_re(random_effect, mtcars)

random_effect2 <- re(form$random[[2]])
epinowcast:::construct_re(random_effect2, mtcars)
}
\seealso{
Functions used to help convert formulas into model designs:
\code{\link[=ar]{ar()}},
\code{\link[=arima]{arima()}},
\code{\link[=arima_terms]{arima_terms()}},
\code{\link[=arma]{arma()}},
\code{\link[=as_string_formula]{as_string_formula()}},
\code{\link[=construct_arima]{construct_arima()}},
\code{\link[=construct_gp]{construct_gp()}},
\code{\link[=construct_rw]{construct_rw()}},
\code{\link[=enw_formula]{enw_formula()}},
\code{\link[=enw_manual_formula]{enw_manual_formula()}},
\code{\link[=gp]{gp()}},
\code{\link[=gp_terms]{gp_terms()}},
\code{\link[=ma]{ma()}},
\code{\link[=parse_formula]{parse_formula()}},
\code{\link[=re]{re()}},
\code{\link[=remove_arima_terms]{remove_arima_terms()}},
\code{\link[=remove_gp_terms]{remove_gp_terms()}},
\code{\link[=remove_rw_terms]{remove_rw_terms()}},
\code{\link[=rw]{rw()}},
\code{\link[=rw_terms]{rw_terms()}},
\code{\link[=split_formula_to_terms]{split_formula_to_terms()}}
}
\concept{formulatools}
