% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_cmdstan_to_rstan}
\alias{convert_cmdstan_to_rstan}
\title{Convert Cmdstan to Rstan}
\usage{
convert_cmdstan_to_rstan(functions)
}
\arguments{
\item{functions}{A character string of stan functions as produced using
\code{\link[=stan_fns_as_string]{stan_fns_as_string()}}.}
}
\value{
A character string of stan functions converted for use in \code{rstan}.
}
\description{
Convert Cmdstan to Rstan
}
\seealso{
Utility functions
\code{\link{coerce_date}()},
\code{\link{expose_stan_fns}()},
\code{\link{is.Date}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
