% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{convolution_matrix}
\alias{convolution_matrix}
\title{Construct a convolution matrix}
\usage{
convolution_matrix(dist, t, include_partial = FALSE)
}
\arguments{
\item{dist}{A vector of list of vectors describing the distribution to be
convolved as a probability mass function.}

\item{t}{Integer value indicating the number of time steps to convolve over.}

\item{include_partial}{Logical, defaults to FALSE. If TRUE, the convolution
include partially complete secondary events.}
}
\value{
A matrix with each column indicating a primary event and each row
indicating a secondary event.
}
\description{
This function allows the construction of convolution matrices which can be
be combined with a vector of primary events to produce a vector of secondary
events for example in the form of a renewal equation or to simulate
reporting delays. Time-varying delays are supported as well as distribution
padding (to allow for use in renewal equation like approaches).
}
\examples{
# Simple convolution matrix with a static distribution
convolution_matrix(c(1, 2, 3), 10)
# Include partially reported convolutions
convolution_matrix(c(1, 2, 3), 10, include_partial = TRUE)
# Use a list of distributions
convolution_matrix(rep(list(c(1, 2, 3)), 10), 10)
# Use a time-varying list of distributions
convolution_matrix(c(rep(list(c(1, 2, 3)), 10), list(c(4, 5, 6))), 11)
}
\seealso{
Helper functions for model modules:
\code{\link[=add_max_observed_delay]{add_max_observed_delay()}},
\code{\link[=add_pmfs]{add_pmfs()}},
\code{\link[=delay_only_ltotal]{delay_only_ltotal()}},
\code{\link[=delay_only_total]{delay_only_total()}},
\code{\link[=enw_dayofweek_structural_reporting]{enw_dayofweek_structural_reporting()}},
\code{\link[=enw_reference_by_report]{enw_reference_by_report()}},
\code{\link[=enw_reps_with_complete_refs]{enw_reps_with_complete_refs()}},
\code{\link[=enw_structural_reporting_metadata]{enw_structural_reporting_metadata()}},
\code{\link[=extract_obs_metadata]{extract_obs_metadata()}},
\code{\link[=extract_sparse_matrix]{extract_sparse_matrix()}},
\code{\link[=latest_obs_as_matrix]{latest_obs_as_matrix()}}
}
\concept{modelmodulehelpers}
