% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cache_dir}
\alias{create_cache_dir}
\title{Create Stan cache directory}
\usage{
create_cache_dir(path)
}
\arguments{
\item{path}{A valid filepath representing the desired cache location. If
the directory does not exist it will be created.}
}
\value{
\code{NULL}
}
\description{
This function creates a cache directory for Stan models if it does not
already exist. This is useful for users who want to set a persistent
cache location but do not want to create the directory manually.
}
\keyword{internal}
