% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.are_delays_valid}
\alias{.are_delays_valid}
\title{Check if delay values are valid}
\usage{
.are_delays_valid(max_delay, daily_max_delay)
}
\arguments{
\item{max_delay}{Delay value in timestep units}

\item{daily_max_delay}{Delay value in days}
}
\value{
Logical indicating if values are valid
}
\description{
Check if delay values are valid
}
\keyword{internal}
