% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.detect_timestep}
\alias{.detect_timestep}
\title{Detect timestep from date vector}
\usage{
.detect_timestep(dates)
}
\arguments{
\item{dates}{A vector of Date objects.}
}
\value{
A character string describing the detected timestep: "daily",
"weekly", or "monthly".
}
\description{
This internal function attempts to detect the timestep (daily, weekly,
monthly) from a vector of dates by examining the differences between
consecutive dates.
}
\seealso{
Functions used for checking inputs
\code{\link{.check_date_alignment}()},
\code{\link{check_calendar_timestep}()},
\code{\link{check_design_matrix_sparsity}()},
\code{\link{check_group}()},
\code{\link{check_group_date_unique}()},
\code{\link{check_max_delay}()},
\code{\link{check_module}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_numeric_timestep}()},
\code{\link{check_observation_indicator}()},
\code{\link{check_quantiles}()},
\code{\link{check_timestep}()},
\code{\link{check_timestep_by_date}()},
\code{\link{check_timestep_by_group}()}
}
\concept{check}
