% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_daily_delay}
\alias{.format_daily_delay}
\title{Format delay for daily timestep}
\usage{
.format_daily_delay(daily_max_delay)
}
\arguments{
\item{daily_max_delay}{Delay value in days}
}
\value{
Formatted string
}
\description{
Format delay for daily timestep
}
\keyword{internal}
