% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_nondaily_delay}
\alias{.format_nondaily_delay}
\title{Format delay for non-daily timestep}
\usage{
.format_nondaily_delay(max_delay, timestep_unit, daily_max_delay)
}
\arguments{
\item{max_delay}{Delay value in timestep units}

\item{timestep_unit}{Unit name}

\item{daily_max_delay}{Delay value in days}
}
\value{
Formatted string
}
\description{
Format delay for non-daily timestep
}
\keyword{internal}
