% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_timestep_unit}
\alias{.get_timestep_unit}
\title{Get timestep unit name}
\usage{
.get_timestep_unit(timestep)
}
\arguments{
\item{timestep}{Timestep specification (character or numeric)}
}
\value{
Character string with unit name (singular form)
}
\description{
Get timestep unit name
}
\keyword{internal}
