% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pluralise_unit}
\alias{.pluralise_unit}
\title{Pluralise unit name if needed}
\usage{
.pluralise_unit(unit, count)
}
\arguments{
\item{unit}{Character string with unit name}

\item{count}{Numeric value to determine if plural needed}
}
\value{
Character string with potentially pluralised unit
}
\description{
Pluralise unit name if needed
}
\keyword{internal}
