% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.warn_insufficient_coverage_data}
\alias{.warn_insufficient_coverage_data}
\title{Warn about insufficient data for coverage check}
\usage{
.warn_insufficient_coverage_data(
  latest_obs_count,
  max_delay_obs_q,
  timestep,
  internal_timestep,
  warn_internal
)
}
\arguments{
\item{latest_obs_count}{Number of observations in filtered data}

\item{max_delay_obs_q}{Quantile-based maximum delay}

\item{timestep}{Timestep specification}

\item{internal_timestep}{Internal timestep multiplier}

\item{warn_internal}{Whether function is called internally}
}
\description{
Warn about insufficient data for coverage check
}
\keyword{internal}
