% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.warn_max_delay_exceeds_observed}
\alias{.warn_max_delay_exceeds_observed}
\title{Warn about max delay exceeding observed delay}
\usage{
.warn_max_delay_exceeds_observed(
  max_delay,
  timestep,
  daily_max_delay,
  max_delay_obs,
  internal_timestep
)
}
\arguments{
\item{max_delay}{Maximum delay in timestep units}

\item{timestep}{Timestep specification}

\item{daily_max_delay}{Specified maximum delay in daily units}

\item{max_delay_obs}{Maximum observed delay in daily units}

\item{internal_timestep}{Internal timestep multiplier}
}
\description{
Warn about max delay exceeding observed delay
}
\keyword{internal}
