% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_add_delay}
\alias{enw_add_delay}
\title{Add a delay variable to the observations}
\usage{
enw_add_delay(obs, timestep = "day", copy = TRUE)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the
following variables: \verb{reference date} (index date of
interest), \code{report_date} (report date for observations),
and \code{confirm} (cumulative observations by reference and
report date).}

\item{timestep}{The timestep to used. This can be a string ("day",
"week") or a numeric whole number representing the number of days.
Note that "month" is not currently supported in user-facing functions
and will throw an error if used.}

\item{copy}{Should \code{obs} be copied (default) or modified
in place?}
}
\value{
A \code{data.table} of observations with a new column \code{delay}.
}
\description{
This helper function takes a \code{data.frame} or \code{data.table} of
observations and adds the delay (numeric, in days) between \code{reference_date}
and \code{report_date} for each observation.
}
\examples{
obs <- data.frame(report_date = as.Date("2021-01-01") + -2:0)
obs$reference_date <- as.Date("2021-01-01")
enw_add_delay(obs)
}
\seealso{
Preprocessing functions:
\code{\link[=enw_add_max_reported]{enw_add_max_reported()}},
\code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
\code{\link[=enw_assign_group]{enw_assign_group()}},
\code{\link[=enw_complete_dates]{enw_complete_dates()}},
\code{\link[=enw_construct_data]{enw_construct_data()}},
\code{\link[=enw_extend_date]{enw_extend_date()}},
\code{\link[=enw_filter_delay]{enw_filter_delay()}},
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}},
\code{\link[=enw_filter_reference_dates_by_report_start]{enw_filter_reference_dates_by_report_start()}},
\code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}},
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}},
\code{\link[=enw_impute_na_observations]{enw_impute_na_observations()}},
\code{\link[=enw_latest_data]{enw_latest_data()}},
\code{\link[=enw_metadata]{enw_metadata()}},
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}},
\code{\link[=enw_missing_reference]{enw_missing_reference()}},
\code{\link[=enw_obs_at_delay]{enw_obs_at_delay()}},
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}},
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}},
\code{\link[=enw_reporting_triangle_to_long]{enw_reporting_triangle_to_long()}},
\code{\link[=enw_retrospective]{enw_retrospective()}}
}
\concept{preprocess}
