% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_add_max_reported}
\alias{enw_add_max_reported}
\title{Add the maximum number of reported cases for each \code{reference_date}}
\usage{
enw_add_max_reported(obs, copy = TRUE)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the
following variables: \verb{reference date} (index date of
interest), \code{report_date} (report date for observations),
and \code{confirm} (cumulative observations by reference and
report date).}

\item{copy}{Should \code{obs} be copied (default) or modified
in place?}
}
\value{
A data.table with new columns \code{max_confirm} and \code{cum_prop_reported}.
\code{max_confirm} is the maximum number of cases reported for a certain
reference_date. \code{cum_prop_reported} is the proportion of cases for a certain
reference_date that are reported until a given report_day, relative to all
cases so far observed for this reference_date.
}
\description{
This is a helper function which adds the maximum (in the sense
of latest observed) number of reported cases for each reference_date and
computes the proportion of already reported cases for each combination of
reference_date and report_date.
}
\examples{
obs <- data.frame(report_date = as.Date("2021-01-01") + 0:2)
obs$reference_date <- as.Date("2021-01-01")
obs$confirm <- 1:3
enw_add_max_reported(obs)
}
\seealso{
Preprocessing functions:
\code{\link[=enw_add_delay]{enw_add_delay()}},
\code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
\code{\link[=enw_assign_group]{enw_assign_group()}},
\code{\link[=enw_complete_dates]{enw_complete_dates()}},
\code{\link[=enw_construct_data]{enw_construct_data()}},
\code{\link[=enw_extend_date]{enw_extend_date()}},
\code{\link[=enw_filter_delay]{enw_filter_delay()}},
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}},
\code{\link[=enw_filter_reference_dates_by_report_start]{enw_filter_reference_dates_by_report_start()}},
\code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}},
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}},
\code{\link[=enw_impute_na_observations]{enw_impute_na_observations()}},
\code{\link[=enw_latest_data]{enw_latest_data()}},
\code{\link[=enw_metadata]{enw_metadata()}},
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}},
\code{\link[=enw_missing_reference]{enw_missing_reference()}},
\code{\link[=enw_obs_at_delay]{enw_obs_at_delay()}},
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}},
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}},
\code{\link[=enw_reporting_triangle_to_long]{enw_reporting_triangle_to_long()}},
\code{\link[=enw_retrospective]{enw_retrospective()}}
}
\concept{preprocess}
