% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_add_metaobs_features}
\alias{enw_add_metaobs_features}
\title{Add common metadata variables}
\usage{
enw_add_metaobs_features(
  metaobs,
  holidays = NULL,
  holidays_to = "Sunday",
  datecol = "date"
)
}
\arguments{
\item{metaobs}{Raw data, coercible via \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.
Coerced object must have \link{Dates} column corresponding to \code{datecol} name.}

\item{holidays}{a (potentially empty) vector of dates (or input
coercible to such; see \code{\link[=coerce_date]{coerce_date()}}). The \code{day_of_week} column will be
set to \code{holidays_to} for these dates.}

\item{holidays_to}{A character string to assign to holidays, when \code{holidays}
argument non-empty. Replaces the \code{day_of_week} column value}

\item{datecol}{The column in \code{metaobs} corresponding to pertinent dates.}
}
\value{
A copy of the \code{metaobs} input, with additional columns:
\itemize{
\item \code{day_of_week}, a factor of values as output from \code{\link[=weekdays]{weekdays()}} and
possibly as \code{holiday_to} if distinct from weekdays values
\item \code{day}, numeric, 0 based from start of time series
\item \code{week}, numeric, 0 based from start of time series
\item \code{month}, numeric, 0 based from start of time series
}
}
\description{
If not already present, annotates time series data with metadata
commonly used in models: day of week, and days, weeks, and months since start
of time series.
}
\details{
Effects models often need to include covariates for time-based
features, such as day of the week (e.g. to reflect different care-seeking
and/or reporting behaviour).

This function is called from within \code{\link[=enw_preprocess_data]{enw_preprocess_data()}} to systematically
annotate \code{metaobs} with these commonly used metadata, if not already present.

However, it can also be used directly on other data.
}
\examples{

# make some example date
nat_germany_hosp <- subset(
  germany_covid19_hosp,
  location == "DE" & age_group == "80+"
)[1:40]

basemeta <- enw_add_metaobs_features(
  nat_germany_hosp,
  datecol = "report_date"
)
basemeta

# with holidays - n.b.: holidays not found are silently ignored
holidaymeta <- enw_add_metaobs_features(
  nat_germany_hosp,
  datecol = "report_date",
  holidays = c(
    "2021-04-04", "2021-04-05",
    "2021-05-01", "2021-05-13",
    "2021-05-24"
  ),
  holidays_to = "Holiday"
)
holidaymeta
subset(holidaymeta, day_of_week == "Holiday")
}
\seealso{
Preprocessing functions:
\code{\link[=enw_add_delay]{enw_add_delay()}},
\code{\link[=enw_add_max_reported]{enw_add_max_reported()}},
\code{\link[=enw_assign_group]{enw_assign_group()}},
\code{\link[=enw_complete_dates]{enw_complete_dates()}},
\code{\link[=enw_construct_data]{enw_construct_data()}},
\code{\link[=enw_extend_date]{enw_extend_date()}},
\code{\link[=enw_filter_delay]{enw_filter_delay()}},
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}},
\code{\link[=enw_filter_reference_dates_by_report_start]{enw_filter_reference_dates_by_report_start()}},
\code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}},
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}},
\code{\link[=enw_impute_na_observations]{enw_impute_na_observations()}},
\code{\link[=enw_latest_data]{enw_latest_data()}},
\code{\link[=enw_metadata]{enw_metadata()}},
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}},
\code{\link[=enw_missing_reference]{enw_missing_reference()}},
\code{\link[=enw_obs_at_delay]{enw_obs_at_delay()}},
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}},
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}},
\code{\link[=enw_reporting_triangle_to_long]{enw_reporting_triangle_to_long()}},
\code{\link[=enw_retrospective]{enw_retrospective()}}
}
\concept{preprocess}
