% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{enw_apply_truncation_to_cases}
\alias{enw_apply_truncation_to_cases}
\title{FUNCTION_TITLE}
\usage{
enw_apply_truncation_to_cases(report_date, cmf, cases)
}
\arguments{
\item{report_date}{PARAM_DESCRIPTION}

\item{cmf}{PARAM_DESCRIPTION}

\item{cases}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions to generate simulated data
\code{\link{enw_dist_cmf}()},
\code{\link{enw_simulate_lnorm_trunc_obs}()}
}
\concept{generatedata}
