% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_as_data_list}
\alias{enw_as_data_list}
\title{Format data for use with stan}
\usage{
enw_as_data_list(
  pobs,
  reference_effects = epinowcast::enw_manual_formula(pobs$metareference[[1]]),
  report_effects = epinowcast::enw_manual_formula(pobs$metareport[[1]]),
  priors = epinowcast::enw_priors(),
  distribution = "lognormal",
  nowcast = TRUE,
  pp = FALSE,
  likelihood = TRUE,
  debug = FALSE,
  output_loglik = FALSE
)
}
\arguments{
\item{pobs}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{reference_effects}{A list of fixed and random design matrices
defining the date of reference model. Defaults to \code{\link[=enw_manual_formula]{enw_manual_formula()}}
which is an intercept only model.}

\item{report_effects}{A list of fixed and random design matrices
defining the date of reports model. Defaults to \code{\link[=enw_manual_formula]{enw_manual_formula()}}
which is an intercept only model.}

\item{priors}{DESCRIPTION.}

\item{distribution}{Character string indicating the type of distribution to
use for reference date effects. The default is to use a lognormal but other
options available include: gamma distributed ("gamma").}

\item{nowcast}{Logical, defaults to \code{TRUE}. Should a nowcast be made using
posterior predictions of the unobserved future reported notifications.}

\item{pp}{Logical, defaults to \code{FALSE}. Should posterior predictions be made
for observed data. Useful for evaluating the performance of the model.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}
}
\value{
A list as required by stan.
}
\description{
Format data for use with stan
}
\seealso{
Functions used for modelling
\code{\link{enw_inits}()},
\code{\link{enw_model}()},
\code{\link{enw_obs_as_data_list}()},
\code{\link{enw_priors}()},
\code{\link{enw_sample}()}
}
\concept{model}
