% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_cache_location_message}
\alias{enw_cache_location_message}
\title{Cache location message for epinowcast package}
\usage{
enw_cache_location_message()
}
\value{
A character vector containing messages. If \code{enw_cache_location} is
not set, it returns instructions for setting the cache location and where to
find more details. If it is set, the function returns a confirmation message
of the current cache location.
}
\description{
This function generates a message in the \code{\link[=epinowcast]{epinowcast()}} package
regarding the cache location. It checks the environment setting for the
cache location and provides guidance to the user on managing this setting.
}
\details{
\code{\link[=enw_cache_location_message]{enw_cache_location_message()}} examines the \code{enw_cache_location}
environment variable. If this variable is not set, it advises the user to
set the cache location using \code{\link[=enw_set_cache]{enw_set_cache()}} to optimize stan compilation
times. If \code{enw_cache_location} is set, it confirms the current cache
location to the user. Management and setting of the cache location can be
done using \code{\link[=enw_set_cache]{enw_set_cache()}}.
}
\keyword{internal}
