% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_cat_new_confirm}
\alias{enw_cat_new_confirm}
\title{Categorise new confirmations by delay group}
\usage{
enw_cat_new_confirm(pobs, delay_group_thresh)
}
\arguments{
\item{pobs}{A preprocessed data object as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{delay_group_thresh}{A numeric vector defining
left-closed interval thresholds for grouping reporting
delays. The smallest value should be zero and the largest
should correspond to \code{max_delay}.}
}
\value{
A \code{data.table} of notification incidence by reference
date and delay group, including columns \code{prop_reported}
and \code{cum_prop_reported}.
}
\description{
Categorises incidence by delay group with empirical
reporting proportions. Intended for use with the
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}} visualisation types that show
reporting patterns by delay.
}
\examples{
pobs <- enw_example("preprocessed_observations")
enw_cat_new_confirm(pobs, delay_group_thresh = c(0, 2, 5, 10, 20))
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_emp_quant_by_reference}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_reference_dates_by_report_start}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_flag_observed_observations}()},
\code{\link{enw_impute_na_observations}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_obs_at_delay}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_reporting_triangle_to_long}()}
}
\concept{preprocess}
