% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_complete_dates}
\alias{enw_complete_dates}
\title{Complete missing reference and report dates}
\usage{
enw_complete_dates(
  obs,
  by = NULL,
  max_delay,
  min_date = min(obs$reference_date, na.rm = TRUE),
  max_date = max(obs$report_date, na.rm = TRUE),
  timestep = "day",
  missing_reference = TRUE,
  completion_beyond_max_report = FALSE,
  flag_observation = FALSE
)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the
following variables: \verb{reference date} (index date of
interest), \code{report_date} (report date for observations),
and \code{confirm} (cumulative observations by reference and
report date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{max_delay}{The maximum delay to model in the delay
distribution, specified in units of the timestep (e.g., if
\code{timestep = "week"}, then \code{max_delay = 3} means 3 weeks). If not
specified the maximum observed delay is assumed to be the true maximum
delay in the model. Otherwise, an integer greater than or equal to 1
can be specified. Observations with delays larger than the maximum
delay will be dropped. If the specified maximum delay is too short,
nowcasts can be biased as important parts of the true delay
distribution are cut off. At the same time, computational cost scales
non-linearly with this setting, so you want the maximum delay to be
as long as necessary, but not much longer.

Steps to take to determine the maximum delay:
\itemize{
\item Consider what is realistic and relevant for your application.
\item Check the proportion of observations reported (\code{prop_reported})
by delay in the \code{new_confirm} output of \code{enw_preprocess_obs}.
\item Use \code{\link[=check_max_delay]{check_max_delay()}} to check the coverage of a candidate
\code{max_delay}.
\item If in doubt, check if increasing the maximum delay noticeably
changes the delay distribution or nowcasts as estimated by
\code{epinowcast}. If it does, your maximum delay may still be too short.
}

Note that delays are zero indexed and so include the reference date
and \code{max_delay - 1} other intervals (i.e. a \code{max_delay} of 1
corresponds to no delay).}

\item{min_date}{The minimum date to include in the data. Defaults to the
minimum reference date found in the data.}

\item{max_date}{The maximum date to include in the data. Defaults to the
maximum report date found in the data.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week") or a numeric whole number representing the number of days.
Note that "month" is not currently supported in user-facing functions
and will throw an error if used.}

\item{missing_reference}{Logical, should entries for cases with missing
reference date be completed as well?, Default: TRUE}

\item{completion_beyond_max_report}{Logical, should entries be completed
beyond the maximum date found in the data? Default: FALSE}

\item{flag_observation}{Logical, should observations that have been
imputed as missing be flagged as not observed?. Makes use of
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}} to add a \code{.observed} logical vector
which indicates if observations have been imputed. This vector can
then be passed to the \code{observation_indicator} argument of \code{\link[=enw_obs]{enw_obs()}} to
control if these observations are used in the likelihood. Default: FALSE}
}
\value{
A \code{data.table} with completed entries for all combinations of
reference dates, groups and possible report dates.
}
\description{
Ensures that all reference and report dates are present for
all groups based on the maximum and minimum dates found in the data.
This function may be of use to users when preprocessing their data. In
general all features that you may consider using as grouping variables
or as covariates need to be included in the \code{by} variable.
}
\examples{
obs <- data.frame(
  report_date = c("2021-10-01", "2021-10-03"), reference_date = "2021-10-01",
  confirm = 1
)
enw_complete_dates(obs)

# Allow completion beyond the maximum date found in the data
enw_complete_dates(obs, completion_beyond_max_report = TRUE, max_delay = 10)
}
\seealso{
Preprocessing functions:
\code{\link[=enw_add_delay]{enw_add_delay()}},
\code{\link[=enw_add_max_reported]{enw_add_max_reported()}},
\code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
\code{\link[=enw_assign_group]{enw_assign_group()}},
\code{\link[=enw_construct_data]{enw_construct_data()}},
\code{\link[=enw_extend_date]{enw_extend_date()}},
\code{\link[=enw_filter_delay]{enw_filter_delay()}},
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}},
\code{\link[=enw_filter_reference_dates_by_report_start]{enw_filter_reference_dates_by_report_start()}},
\code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}},
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}},
\code{\link[=enw_impute_na_observations]{enw_impute_na_observations()}},
\code{\link[=enw_latest_data]{enw_latest_data()}},
\code{\link[=enw_metadata]{enw_metadata()}},
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}},
\code{\link[=enw_missing_reference]{enw_missing_reference()}},
\code{\link[=enw_obs_at_delay]{enw_obs_at_delay()}},
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}},
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}},
\code{\link[=enw_reporting_triangle_to_long]{enw_reporting_triangle_to_long()}},
\code{\link[=enw_retrospective]{enw_retrospective()}}
}
\concept{preprocess}
