% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_cumulative_day_membership}
\alias{enw_cumulative_day_membership}
\title{FUNCTION_TITLE}
\usage{
enw_cumulative_day_membership(metaobs)
}
\arguments{
\item{metaobs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_day_of_week}()},
\code{\link{enw_add_pooling_effect}()},
\code{\link{enw_add_week}()},
\code{\link{enw_cumulative_week_membership}()},
\code{\link{enw_dates_to_factors}()},
\code{\link{enw_day_model}()},
\code{\link{enw_day_of_week_model}()},
\code{\link{enw_design}()},
\code{\link{enw_effects_metadata}()},
\code{\link{enw_intercept_model}()},
\code{\link{enw_weekly_model}()}
}
\concept{modeldesign}
