% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-designs.R
\name{enw_day_model}
\alias{enw_day_model}
\title{FUNCTION_TITLE}
\usage{
enw_day_model(metaobs, rw = FALSE)
}
\arguments{
\item{metaobs}{PARAM_DESCRIPTION}

\item{rw}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
Functions used to formulate models
\code{\link{enw_day_of_week_model}()},
\code{\link{enw_intercept_model}()},
\code{\link{enw_weekly_model}()}
}
\concept{modeldesign}
