% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_delay_metadata}
\alias{enw_delay_metadata}
\title{Calculate reporting delay metadata for a given maximum delay}
\usage{
enw_delay_metadata(max_delay = 20, breaks = 4)
}
\arguments{
\item{max_delay}{The maximum delay to model in the delay
distribution, specified in units of the timestep (e.g., if
\code{timestep = "week"}, then \code{max_delay = 3} means 3 weeks). If not
specified the maximum observed delay is assumed to be the true maximum
delay in the model. Otherwise, an integer greater than or equal to 1
can be specified. Observations with delays larger than the maximum
delay will be dropped. If the specified maximum delay is too short,
nowcasts can be biased as important parts of the true delay
distribution are cut off. At the same time, computational cost scales
non-linearly with this setting, so you want the maximum delay to be
as long as necessary, but not much longer.

Steps to take to determine the maximum delay:
\itemize{
\item Consider what is realistic and relevant for your application.
\item Check the proportion of observations reported (\code{prop_reported})
by delay in the \code{new_confirm} output of \code{enw_preprocess_obs}.
\item Use \code{\link[=check_max_delay]{check_max_delay()}} to check the coverage of a candidate
\code{max_delay}.
\item If in doubt, check if increasing the maximum delay noticeably
changes the delay distribution or nowcasts as estimated by
\code{epinowcast}. If it does, your maximum delay may still be too short.
}

Note that delays are zero indexed and so include the reference date
and \code{max_delay - 1} other intervals (i.e. a \code{max_delay} of 1
corresponds to no delay).}

\item{breaks}{Numeric, defaults to 4. The number of breaks to use when
constructing a categorised version of numeric delays.}
}
\value{
A  \code{data.frame}  of delay metadata. This includes:
\itemize{
\item \code{delay}: The numeric delay from reference date to report.
\item \code{delay_cat}: The categorised delay. This may be useful for model building.
\item \code{delay_week}: The numeric week since the delay was reported. This again
may be useful for model building.
\item \code{delay_tail}: A logical variable defining if the delay is in the upper
75\% of the potential delays. This may be particularly useful when building
models that assume a parametric distribution in order to increase the
weight of the tail of the reporting distribution in a pragmatic way.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Calculate delay metadata based on the supplied maximum delay and
independent of other metadata or date indexing. These data are meant to be
used in conjunction with metadata on the date of reference. Users can build
additional features this  \code{data.frame}  or regenerate it using this
function in the output of \code{enw_preprocess_data()}.

\code{enw_delay_metadata()} was renamed to \code{\link[=enw_metadata_delay]{enw_metadata_delay()}} for better
consistency.
}
\examples{
enw_delay_metadata(max_delay = 20, breaks = 4)
}
\keyword{internal}
