% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{enw_dist_cmf}
\alias{enw_dist_cmf}
\title{FUNCTION_TITLE}
\usage{
enw_dist_cmf(logmean, logsd, max = 20)
}
\arguments{
\item{logmean}{PARAM_DESCRIPTION}

\item{logsd}{PARAM_DESCRIPTION}

\item{max}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions to generate simulated data
\code{\link{enw_apply_truncation_to_cases}()},
\code{\link{enw_simulate_lnorm_trunc_obs}()}
}
\concept{generatedata}
