% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_filter_obs}
\alias{enw_filter_obs}
\title{Filter observations to restrict the maximum reporting delay}
\usage{
enw_filter_obs(obs, max_delay)
}
\arguments{
\item{obs}{A data frame containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date), and \code{group} (as added by \code{\link[=enw_assign_group]{enw_assign_group()}}).}

\item{max_delay}{Numeric defaults to 20. The maximum number of days to
include in the delay distribution. Computation scales non-linearly with this
setting so consider what maximum makes sense for your data carefully.}
}
\value{
A data frame filtered so that dates by report are less than or equal
the reference date plus the maximum delay.
}
\description{
Filter observations to restrict the maximum reporting delay
}
\examples{
obs <- enw_example("preprocessed")$obs[[1]]
enw_filter_obs(obs, max_delay = 2)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_retrospective_data}()}
}
\concept{preprocess}
