% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_get_environment_contents}
\alias{enw_get_environment_contents}
\title{Identify cache location}
\usage{
enw_get_environment_contents(remove_enw_cache_location = TRUE)
}
\arguments{
\item{remove_enw_cache_location}{Logical indicating whether to remove the
\code{enw_cache_location} entry from the \code{.Renviron} file. Defaults to \code{TRUE}.}
}
\value{
A list containing the contents of the \code{.Renviron} file and its path.
}
\description{
This function retrieves environment variable settings and manages the
\code{.Renviron} file in the user's project or home directory.
The project directory will be examined first, if it exists.
It can optionally remove the entry for \code{enw_cache_location}.
}
\keyword{internal}
