% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_hot_encode_and_bind}
\alias{enw_hot_encode_and_bind}
\title{One-hot encode a variable and column-bind it to the original data.table}
\usage{
enw_hot_encode_and_bind(metaobs, feature, contrasts = FALSE)
}
\arguments{
\item{metaobs}{A data.frame containing the data to be encoded.}

\item{feature}{The name of the categorical variable to one-hot encode as a
character string.}

\item{contrasts}{Logical. If TRUE, create one-hot encoded variables
with contrasts; if FALSE, create them without contrasts. Defaults to FALSE.}
}
\description{
This function takes a data.frame and a categorical variable, performs
one-hot encoding, and column-binds the encoded variables back to the
data.frame.
}
\examples{
metaobs <- data.frame(week = 1:3)
enw_hot_encode_and_bind(metaobs, "week")
enw_hot_encode_and_bind(metaobs, "week", contrasts = TRUE)
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_cumulative_membership}()},
\code{\link{enw_add_pooling_effect}()},
\code{\link{enw_design}()},
\code{\link{enw_effects_metadata}()}
}
\concept{modeldesign}
