% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_impute_na_observations}
\alias{enw_impute_na_observations}
\title{Impute NA observations}
\usage{
enw_impute_na_observations(obs, by = NULL, copy = TRUE)
}
\arguments{
\item{obs}{A \code{data.frame} with at least 'confirm' and 'reference_date'
columns.}

\item{by}{A character vector of column names to group by. Defaults to
an empty vector.}

\item{copy}{A logical; if \code{TRUE} (the default) creates a copy; otherwise,
modifies \code{obs} in place.}
}
\value{
A \code{data.table} with imputed 'confirm' column where NA values have
been replaced with zero.
}
\description{
Imputes NA values in the 'confirm' column.
NA values are replaced with the last available observation or 0.
}
\examples{
dt <- data.frame(
  id = 1:3, confirm = c(NA, 1, 2),
  reference_date = as.Date("2021-01-01")
)
enw_impute_na_observations(dt)
}
\seealso{
Preprocessing functions:
\code{\link[=enw_add_delay]{enw_add_delay()}},
\code{\link[=enw_add_max_reported]{enw_add_max_reported()}},
\code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
\code{\link[=enw_assign_group]{enw_assign_group()}},
\code{\link[=enw_complete_dates]{enw_complete_dates()}},
\code{\link[=enw_construct_data]{enw_construct_data()}},
\code{\link[=enw_extend_date]{enw_extend_date()}},
\code{\link[=enw_filter_delay]{enw_filter_delay()}},
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}},
\code{\link[=enw_filter_reference_dates_by_report_start]{enw_filter_reference_dates_by_report_start()}},
\code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}},
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}},
\code{\link[=enw_latest_data]{enw_latest_data()}},
\code{\link[=enw_metadata]{enw_metadata()}},
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}},
\code{\link[=enw_missing_reference]{enw_missing_reference()}},
\code{\link[=enw_obs_at_delay]{enw_obs_at_delay()}},
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}},
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}},
\code{\link[=enw_reporting_triangle_to_long]{enw_reporting_triangle_to_long()}},
\code{\link[=enw_retrospective]{enw_retrospective()}}
}
\concept{preprocess}
