% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-converters.R
\name{enw_incidence_to_cumulative}
\alias{enw_incidence_to_cumulative}
\title{Calculate cumulative reported cases from incidence of new reports}
\usage{
enw_incidence_to_cumulative(obs, by = NULL)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), and \code{new_confirm} (incident observations by reference and
report date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}
}
\value{
The input \code{data.frame} with a new variable \code{confirm}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
# Default reconstruct incidence
dt <- germany_covid19_hosp[location == "DE"][age_group == "00+"]
dt <- enw_add_incidence(dt)
dt <- dt[, confirm := NULL]
enw_add_cumulative(dt)

# Make use of maximum reported to calculate empirical daily reporting
enw_add_cumulative(dt)
}
\seealso{
Data converters
\code{\link{enw_add_cumulative}()},
\code{\link{enw_add_incidence}()},
\code{\link{enw_aggregate_cumulative}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_incidence_to_linelist}()},
\code{\link{enw_linelist_to_incidence}()}
}
\concept{dataconverters}
\keyword{internal}
