% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-converters.R
\name{enw_incidence_to_linelist}
\alias{enw_incidence_to_linelist}
\title{Convert Aggregate Counts (Incidence) to a Line List}
\usage{
enw_incidence_to_linelist(
  obs,
  reference_date = "reference_date",
  report_date = "report_date"
)
}
\arguments{
\item{obs}{An object coercible to a \code{data.table} (such as a \code{data.frame})
which must have a \code{new_confirm} column.}

\item{reference_date}{A character string of the variable name to use
for the \code{reference_date} in the line list. The default is "reference_date".}

\item{report_date}{A character string of the variable name to use
for the \code{report_date} in the line list. The default is "report_date".}
}
\value{
A \code{data.table} with the following variables: \code{id}, \code{reference_date},
\code{report_date}, and any other variables in the \code{obs} object. Rows in \code{obs}
will be duplicated based on the \code{new_confirm} column. \code{reference_date} and
\code{report_date} may be renamed if \code{reference_date} and \code{report_date} are
supplied.
}
\description{
This function takes a \code{data.table} of aggregate counts or
something coercible to a \code{data.table} (such as a \code{data.frame}) and converts
it to a line list where each row represents a case.
}
\examples{
incidence <- enw_filter_reference_dates_by_report_start(
  germany_covid19_hosp
)
incidence <- enw_add_incidence(incidence)
incidence <- enw_filter_reference_dates(
  incidence[location == "DE"], include_days = 10
)
enw_incidence_to_linelist(incidence, reference_date = "onset_date")
}
\seealso{
Data converters:
\code{\link[=enw_add_cumulative]{enw_add_cumulative()}},
\code{\link[=enw_add_incidence]{enw_add_incidence()}},
\code{\link[=enw_aggregate_cumulative]{enw_aggregate_cumulative()}},
\code{\link[=enw_linelist_to_incidence]{enw_linelist_to_incidence()}}
}
\concept{dataconverters}
