% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace.R
\name{enw_laplace_marginal_model}
\alias{enw_laplace_marginal_model}
\title{Compile the experimental embedded-Laplace nowcast model}
\usage{
enw_laplace_marginal_model(
  model = system.file("stan", "models", "epinowcast_laplace.stan", package =
    "epinowcast"),
  include = system.file("stan", package = "epinowcast"),
  compile = TRUE,
  threads = TRUE,
  profile = FALSE,
  target_dir = epinowcast::enw_get_cache(),
  stanc_options = list(),
  cpp_options = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model}{A character string indicating the path to the model.
If not supplied the package default model is used.}

\item{include}{A character string specifying the path to any stan
files to include in the model. If missing the package default is used.}

\item{compile}{Logical, defaults to \code{TRUE}. Should the model
be loaded and compiled using \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}

\item{threads}{Logical, defaults to \code{TRUE}. Should the model compile with
support for multi-thread support in chain. Note that setting this will
produce a warning that \code{threads_to_chain} is set and ignored. Changing this
to \code{FALSE} is not expected to yield any performance benefits even when
not using multithreading and thus not recommended.}

\item{profile}{Logical, defaults to \code{FALSE}. Should the model be profiled?
For more on profiling see the \href{https://mc-stan.org/cmdstanr/articles/profiling.html}{\code{cmdstanr} documentation}. # nolint}

\item{target_dir}{The path to a directory in which the manipulated .stan
files without profiling statements should be stored. To avoid overriding of
the original .stan files, this should be different from the directory of the
original model and the \code{include_paths}.}

\item{stanc_options}{A list of options to pass to the \code{stanc_options} of
\code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}. By default nothing is passed but potentially
users may wish to pass optimisation flags for example. See the documentation
for \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}} for further details.}

\item{cpp_options}{A list of options to pass to the \code{cpp_options} of
\code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}. By default nothing is passed but potentially
users may wish to pass optimisation flags for example. See the documentation
for \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}} for further details. Note that the \code{threads}
argument replaces \code{stan_threads}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose
messages be shown.}

\item{...}{Additional arguments passed to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}
}
\value{
A \code{cmdstanr} model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Compiles the opt-in, experimental embedded-Laplace nowcast model shipped
in \code{inst/stan/models/epinowcast_laplace.stan}. This model marginalises the
latent log-expected-count field analytically using Stan's embedded Laplace
approximation (\verb{laplace_marginal_*}, available from CmdStan 2.39) rather
than sampling it with NUTS. It is a thin wrapper around \code{\link[=enw_model]{enw_model()}} that
points at the Laplace model and the package Stan include directory.

This inference path is experimental and supports only a subset of the full
\code{\link[=epinowcast]{epinowcast()}} model (see \code{\link[=enw_laplace_marginal]{enw_laplace_marginal()}}).
It requires CmdStan >= 2.39.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
mod <- enw_laplace_marginal_model()
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{check_laplace_supported}()},
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_laplace_marginal}()},
\code{\link{enw_laplace_marginal_data}()},
\code{\link{enw_laplace_marginal_inits}()},
\code{\link{enw_model}()},
\code{\link{enw_pathfinder}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_stan_to_r}()},
\code{\link{enw_unset_cache}()},
\code{\link{remove_profiling}()},
\code{\link{split_laplace_design}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
