% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_latest_data_inner}
\alias{enw_latest_data_inner}
\title{Filter observations to the latest available reported (internal)}
\usage{
enw_latest_data_inner(obs)
}
\arguments{
\item{obs}{An observation data frame containing \code{report_date} and
\code{reference_date} columns.}
}
\description{
Filter observations to the latest available reported (internal)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
