% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-converters.R
\name{enw_linelist_to_incidence}
\alias{enw_linelist_to_incidence}
\title{Convert a Line List to Aggregate Counts (Incidence)}
\usage{
enw_linelist_to_incidence(
  linelist,
  reference_date = "reference_date",
  report_date = "report_date",
  by = NULL,
  max_delay,
  completion_beyond_max_report = FALSE,
  copy = TRUE
)
}
\arguments{
\item{linelist}{An object coercible to a \code{data.table} (such as a
\code{data.frame}) where each row represents a case. Must contain at least
two date variables or variables that can be coerced to dates.}

\item{reference_date}{A date or a variable that can be coerced to a date
that represents the date of interest for the case. For example, if the
\code{reference_date} is the date of symptom onset then the \code{new_confirm} will
be the number of new cases reported (based on \code{report_date}) on each day
that had onset on that day. The default is "reference_date".}

\item{report_date}{A date or a variable that can be coerced to a date
that represents the date the case was reported. The default is "report_date".}

\item{by}{A character vector of variables to also aggregate by (i.e. as well
as using the \code{reference_date} and \code{report_date}). If not supplied
then the function will aggregate by just the \code{reference_date} and
\code{report_date}.}

\item{max_delay}{The maximum delay (in days) between the
\code{reference_date} and the \code{report_date}. If not supplied then the
function will use the maximum observed delay. Note that this function
operates before timestep conversion, so max_delay is always in days
here. If the \code{max_delay} is less than the maximum number of days
between the \code{reference_date} and the \code{report_date} in the \code{linelist}
then the function will use this value instead and inform the user.}

\item{completion_beyond_max_report}{Logical, should entries be completed
beyond the maximum date found in the data? Default: FALSE}

\item{copy}{Should \code{obs} be copied (default) or modified
in place?}
}
\value{
A \code{data.table} with the following variables: \code{reference_date},
\code{report_date}, \code{new_confirm}, \code{confirm}, \code{delay}, and
any variables specified in \code{by}.
}
\description{
This function takes a line list (i.e. tabular data where each
row represents a case) and aggregates to a count (\code{new_confirm}) of cases by
user-specified \code{reference_date}s and \code{report_date}s. This is enables the use
of \code{\link[=enw_preprocess_data]{enw_preprocess_data()}} and other \code{\link[=epinowcast]{epinowcast()}} preprocessing functions.
}
\examples{
linelist <- data.frame(
  onset_date = as.Date(c("2021-01-02", "2021-01-03", "2021-01-02")),
  report_date = as.Date(c("2021-01-03", "2021-01-05", "2021-01-04"))
)
enw_linelist_to_incidence(linelist, reference_date = "onset_date")

# Specify a custom maximum delay and allow completion beyond the maximum
# observed delay
enw_linelist_to_incidence(
 linelist, reference_date = "onset_date", max_delay = 5,
 completion_beyond_max_report = TRUE
)
}
\seealso{
Data converters:
\code{\link[=enw_add_cumulative]{enw_add_cumulative()}},
\code{\link[=enw_add_incidence]{enw_add_incidence()}},
\code{\link[=enw_aggregate_cumulative]{enw_aggregate_cumulative()}},
\code{\link[=enw_incidence_to_linelist]{enw_incidence_to_linelist()}}
}
\concept{dataconverters}
