% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_metadata}
\alias{enw_metadata}
\title{Extract metadata from raw data}
\usage{
enw_metadata(obs, target_date = c("reference_date", "report_date"))
}
\arguments{
\item{obs}{A \code{data.frame} or \code{data.table} with columns: \code{reference_date}
and / or \code{report_date}; at least one must be provided, \code{.group}, a grouping
column and a \code{date}, a \link{Date} column.}

\item{target_date}{A character string, either "reference_date" or
"report_date". The column corresponding to this string will be used
as the target date for metadata extraction.}
}
\value{
A data.table with columns:
\itemize{
\item \code{date}, a \link{Date} column
\item \code{.group}, a grouping column
}

and the first observation for each group and date.
The data.table is sorted by \code{.group} and \code{date}.
}
\description{
Extract metadata from raw data, either
by reference or by report date. For the target date chosen
(reference or report), \code{confirm}, \verb{max_confirm``, and }cum_prop_reported`
are dropped and the first observation for each group and date is retained.
}
\examples{
obs <- data.frame(
  reference_date = as.Date("2021-01-01"),
  report_date = as.Date("2022-01-01"), x = 1:10
)
enw_metadata(obs, target_date = "reference_date")
}
\seealso{
Preprocessing functions:
\code{\link[=enw_add_delay]{enw_add_delay()}},
\code{\link[=enw_add_max_reported]{enw_add_max_reported()}},
\code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
\code{\link[=enw_assign_group]{enw_assign_group()}},
\code{\link[=enw_complete_dates]{enw_complete_dates()}},
\code{\link[=enw_construct_data]{enw_construct_data()}},
\code{\link[=enw_extend_date]{enw_extend_date()}},
\code{\link[=enw_filter_delay]{enw_filter_delay()}},
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}},
\code{\link[=enw_filter_reference_dates_by_report_start]{enw_filter_reference_dates_by_report_start()}},
\code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}},
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}},
\code{\link[=enw_impute_na_observations]{enw_impute_na_observations()}},
\code{\link[=enw_latest_data]{enw_latest_data()}},
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}},
\code{\link[=enw_missing_reference]{enw_missing_reference()}},
\code{\link[=enw_obs_at_delay]{enw_obs_at_delay()}},
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}},
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}},
\code{\link[=enw_reporting_triangle_to_long]{enw_reporting_triangle_to_long()}},
\code{\link[=enw_retrospective]{enw_retrospective()}}
}
\concept{preprocess}
