% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_metadata_maxdelay}
\alias{enw_metadata_maxdelay}
\title{Get the different user-specified, observed, and modelled maximum delays.}
\usage{
enw_metadata_maxdelay(obs, max_delay = 20)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\code{reference_date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{max_delay}{Numeric defaults to 20 and needs to be greater than or equal
to 1 and an integer (internally it will be coerced to one using
\code{\link[=as.integer]{as.integer()}}). The maximum number of days to include in the delay
distribution. Observations with delays larger then the maximum delay will be
dropped. If the specified maximum delay is too short, nowcasts can be biased
as important parts of the true delay distribution are cut off. At the same
time, computational cost scales non-linearly with this setting, so you want
the maximum delay to be as long as necessary, but not much longer. Consider
what delays are realistic for your application, and when in doubt, check if
increasing the maximum delay noticeably changes the delay distribution or
nowcasts as estimated by epinowcast. If it does, your maximum delay may still
be too short. Note that delays are zero indexed and so include the reference
date and \code{max_delay - 1} other days (i.e. a \code{max_delay} of 1 corresponds to
no delay). See \code{\link[=enw_metadata_delay]{enw_metadata_delay()}} for checking the coverage of a delay
distribution for a given maximum delay.}
}
\value{
A \code{list} of maximum delays. These include:
\itemize{
\item \code{spec}: The maximum delay specified by the user.
\item \code{obs}: The maximum delay observed in the data.
\item \code{model}: The maximum delay used by the model.
}
}
\description{
This metadata function records the different types of maximum
delays relevant for modeling. To obtain statistics on individual delays
for a given maximum delay, see \code{enw_metadata_delay} instead. The maximum
reporting delay is used to make the modeling of reporting delays tractable
by right-truncating the delay distribution at a reasonable number of days.
The maximum delay is specified by the user, and can be smaller or larger
than the maximum delay observed in the data. Importantly, epinowcast will
currently always use the smaller one of the two maximum delays for
modeling. This means that observations with a delay larger than the
specified maximum delay will be dropped from the analysis. In some
settings, for example outbreaks where little data is available, this may not
be ideal. If this is an issue for you, please get in touch with the
developers by opening an issue on GitHub.
}
\examples{
obs <- enw_example(type = "preprocessed_observations")$obs[[1]]
enw_metadata_maxdelay(obs, max_delay = 20)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
