% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_new_reports}
\alias{enw_new_reports}
\title{FUNCTION_TITLE}
\usage{
enw_new_reports(obs)
}
\arguments{
\item{obs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
