% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_nowcast_samples}
\alias{enw_nowcast_samples}
\title{Extract posterior samples for the nowcast prediction}
\usage{
enw_nowcast_samples(fit, obs, max_delay = NULL, timestep = "day")
}
\arguments{
\item{fit}{A \code{cmdstanr} fit object.}

\item{obs}{An observation \code{data.frame} containing \code{reference_date}
columns of the same length as the number of rows in the posterior and the
most up to date observation for each date. This is used to align the
posterior with the observations. The easiest source of this data is the
output of latest output of \code{\link[=enw_preprocess_data]{enw_preprocess_data()}} or \code{\link[=enw_latest_data]{enw_latest_data()}}.}

\item{max_delay}{Maximum delay to which nowcasts should be extracted,
in units of the timestep used during preprocessing. Must be equal
(default) or larger than the modelled maximum delay. If it is larger,
then nowcasts for unmodelled dates are added by assuming that case
counts beyond the modelled maximum delay are fully observed.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week") or a numeric whole number representing the number of days.
Note that "month" is not currently supported in user-facing functions
and will throw an error if used.}
}
\value{
A \code{data.frame} of posterior samples for the nowcast prediction.
This uses observed data where available and the posterior prediction
where not.
}
\description{
A generic wrapper around \code{\link[posterior:draws_df]{posterior::draws_df()}} with
opinionated defaults to extract the posterior samples for the
nowcast (\code{"pp_inf_obs"} from the \code{stan} code). The functionality of
this function can be used directly on the output of \code{\link[=epinowcast]{epinowcast()}} using
the supplied \code{\link[=summary.epinowcast]{summary.epinowcast()}} method.
}
\examples{
fit <- enw_example("nowcast")
enw_nowcast_samples(
  fit$fit[[1]],
  fit$latest[[1]],
  fit$max_delay,
  "day"
  )
}
\seealso{
Functions used for postprocessing of model fits:
\code{\link[=.check_primarycensored]{.check_primarycensored()}},
\code{\link[=.delay_draw_columns]{.delay_draw_columns()}},
\code{\link[=.discretise_parametric_pmf]{.discretise_parametric_pmf()}},
\code{\link[=build_ord_obs]{build_ord_obs()}},
\code{\link[=enw_add_latest_obs_to_nowcast]{enw_add_latest_obs_to_nowcast()}},
\code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}},
\code{\link[=enw_posterior]{enw_posterior()}},
\code{\link[=enw_posterior_delay]{enw_posterior_delay()}},
\code{\link[=enw_pp_summary]{enw_pp_summary()}},
\code{\link[=enw_quantiles_to_long]{enw_quantiles_to_long()}},
\code{\link[=enw_summarise_samples]{enw_summarise_samples()}},
\code{\link[=subset_obs]{subset_obs()}}
}
\concept{postprocess}
