% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_obs_as_data_list}
\alias{enw_obs_as_data_list}
\title{Format observed data for use with stan}
\usage{
enw_obs_as_data_list(pobs)
}
\arguments{
\item{pobs}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Format observed data for use with stan
}
\seealso{
Functions used for modelling
\code{\link{enw_as_data_list}()},
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_inits}()},
\code{\link{enw_model}()},
\code{\link{enw_opts_as_data_list}()},
\code{\link{enw_priors}()},
\code{\link{enw_sample}()}
}
\concept{model}
