% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_one_hot_encode_feature}
\alias{enw_one_hot_encode_feature}
\title{One-hot encode a variable and column-bind it to the original data.table}
\usage{
enw_one_hot_encode_feature(metaobs, feature, contrasts = FALSE)
}
\arguments{
\item{metaobs}{A data.frame containing the data to be encoded.}

\item{feature}{The name of the categorical variable to one-hot encode as a
character string.}

\item{contrasts}{Logical. If TRUE, create one-hot encoded variables
with contrasts; if FALSE, create them without contrasts. Defaults to FALSE.}
}
\description{
This function takes a data.frame and a categorical variable, performs
one-hot encoding, and column-binds the encoded variables back to the
data.frame.
}
\examples{
metaobs <- data.frame(week = 1:2)
enw_one_hot_encode_feature(metaobs, "week")
enw_one_hot_encode_feature(metaobs, "week", contrasts = TRUE)

metaobs <- data.frame(week = 1:6)
enw_one_hot_encode_feature(metaobs, "week")
enw_one_hot_encode_feature(metaobs, "week", contrasts = TRUE)
}
\seealso{
Functions used to formulate models:
\code{\link[=enw_add_cumulative_membership]{enw_add_cumulative_membership()}},
\code{\link[=enw_add_pooling_effect]{enw_add_pooling_effect()}},
\code{\link[=enw_design]{enw_design()}},
\code{\link[=enw_effects_metadata]{enw_effects_metadata()}}
}
\concept{modeldesign}
