% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot}
\alias{enw_plot}
\title{Default generic posterior plot}
\usage{
enw_plot(posterior, obs = NULL, log = TRUE, ...)
}
\arguments{
\item{obs}{A data frame of observed data as produced by \code{\link[=latest_obs]{latest_obs()}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:aes]{ggplot2::aes()}}}
}
\value{
A \code{ggplot2} plot.
}
\description{
Default generic posterior plot
}
\seealso{
Plotting functions
\code{\link{enw_plot_theme}()}
}
\concept{plot}
