% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_delay_fraction}
\alias{enw_plot_delay_fraction}
\title{Plot empirical reporting delay heatmap}
\usage{
enw_plot_delay_fraction(pobs, delay_group_thresh, facet = TRUE)
}
\arguments{
\item{pobs}{A preprocessed data object as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{delay_group_thresh}{A numeric vector defining
left-closed interval thresholds for delay groups.}

\item{facet}{Logical. When \code{TRUE} (the default), plots with
more than one \code{.group} are automatically wrapped by group.
Set to \code{FALSE} to disable and add a custom facet layer.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Tile plot showing the fraction reported by delay
group and reference date.
}
\examples{
pobs <- enw_example("preprocessed_observations")
enw_plot_delay_fraction(pobs, c(0, 2, 5, 10, 21))
}
\seealso{
Plotting functions:
\code{\link[=enw_delay_categories]{enw_delay_categories()}},
\code{\link[=enw_delay_quantiles]{enw_delay_quantiles()}},
\code{\link[=enw_plot_delay_counts]{enw_plot_delay_counts()}},
\code{\link[=enw_plot_delay_cumulative]{enw_plot_delay_cumulative()}},
\code{\link[=enw_plot_delay_quantiles]{enw_plot_delay_quantiles()}},
\code{\link[=enw_plot_nowcast_quantiles]{enw_plot_nowcast_quantiles()}},
\code{\link[=enw_plot_obs]{enw_plot_obs()}},
\code{\link[=enw_plot_pp_quantiles]{enw_plot_pp_quantiles()}},
\code{\link[=enw_plot_quantiles]{enw_plot_quantiles()}},
\code{\link[=enw_plot_theme]{enw_plot_theme()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}},
\code{\link[=plot.epinowcast]{plot.epinowcast()}}
}
\concept{plot}
