% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_emprep_frac}
\alias{enw_plot_emprep_frac}
\title{Plot empirical reporting delay heatmap}
\usage{
enw_plot_emprep_frac(pobs, delay_group_thresh)
}
\arguments{
\item{pobs}{A preprocessed data object as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{delay_group_thresh}{A numeric vector defining
left-closed interval thresholds for delay groups.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Tile plot showing the fraction reported by delay
group and reference date.
}
\examples{
pobs <- enw_example("preprocessed_observations")
enw_plot_emprep_frac(pobs, c(0, 2, 5, 10, 21))
}
\seealso{
Plotting functions
\code{\link{enw_plot_emp_ts_del}()},
\code{\link{enw_plot_emprep_cum}()},
\code{\link{enw_plot_emprep_quant}()},
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
