% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_emprep_quant}
\alias{enw_plot_emprep_quant}
\title{Plot quantiles of empirical reporting delay by reference date over time
#'}
\usage{
enw_plot_emprep_quant(pobs, quantiles, ...)
}
\arguments{
\item{pobs}{Pre-processed observations as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{quantiles}{A vector of quantiles used for the "emp_rep_quant" plot.
Default is NULL, corresponding to the 0.10, 0.50, and 0.90 quantiles.}

\item{...}{Additional arguments passed to \code{\link[=enw_plot_emprep_cum]{enw_plot_emprep_cum()}}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot quantiles of empirical reporting delay by reference date over time
#'
}
\seealso{
Plotting functions
\code{\link{enw_plot_emp_ts_del}()},
\code{\link{enw_plot_emprep_cum}()},
\code{\link{enw_plot_emprep_frac}()},
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
