% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_nowcast_quantiles}
\alias{enw_plot_nowcast_quantiles}
\title{Plot nowcast quantiles}
\usage{
enw_plot_nowcast_quantiles(nowcast, latest_obs = NULL, log = FALSE, ...)
}
\arguments{
\item{nowcast}{A \code{data.frame} of summarised posterior nowcast
estimates containing at least a \code{confirm} count column and a
\code{reference_date} date variable.}

\item{latest_obs}{A \code{data.frame} of observed data containing at least a
\code{confirm} count variable and the same date variable as in the main data.frame
used for plotting.}

\item{log}{Logical, defaults to \code{FALSE}. Should counts be plot on the log
scale.}

\item{...}{Additional arguments passed to \code{\link[=enw_plot_pp_quantiles]{enw_plot_pp_quantiles()}}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot nowcast quantiles
}
\examples{
nowcast <- enw_example("nowcast")
nowcast <- summary(nowcast, probs = c(0.05, 0.2, 0.8, 0.95))
enw_plot_nowcast_quantiles(nowcast)
}
\seealso{
Plotting functions:
\code{\link[=enw_delay_categories]{enw_delay_categories()}},
\code{\link[=enw_delay_quantiles]{enw_delay_quantiles()}},
\code{\link[=enw_plot_delay_counts]{enw_plot_delay_counts()}},
\code{\link[=enw_plot_delay_cumulative]{enw_plot_delay_cumulative()}},
\code{\link[=enw_plot_delay_fraction]{enw_plot_delay_fraction()}},
\code{\link[=enw_plot_delay_quantiles]{enw_plot_delay_quantiles()}},
\code{\link[=enw_plot_obs]{enw_plot_obs()}},
\code{\link[=enw_plot_pp_quantiles]{enw_plot_pp_quantiles()}},
\code{\link[=enw_plot_quantiles]{enw_plot_quantiles()}},
\code{\link[=enw_plot_theme]{enw_plot_theme()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}},
\code{\link[=plot.epinowcast]{plot.epinowcast()}}
}
\concept{plot}
