% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_rep_quant}
\alias{enw_plot_rep_quant}
\title{Plot empirical reporting delay quantiles}
\usage{
enw_plot_rep_quant(pobs, quantiles = c(0.1, 0.5, 0.9))
}
\arguments{
\item{pobs}{A preprocessed data object as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{quantiles}{A numeric vector of probabilities.
Defaults to \code{c(0.1, 0.5, 0.9)}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Line plot showing quantiles of the empirical
delay distribution over reference dates.
}
\examples{
pobs <- enw_example("preprocessed_observations")
enw_plot_rep_quant(pobs)
}
\seealso{
Plotting functions
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_rep_cum}()},
\code{\link{enw_plot_rep_frac}()},
\code{\link{enw_plot_theme}()},
\code{\link{enw_plot_ts_delay}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
