% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_theme}
\alias{enw_plot_theme}
\title{Package plot theme}
\usage{
enw_plot_theme(plot)
}
\arguments{
\item{plot}{\code{ggplot2} plot object.}
}
\value{
\code{ggplot2} plot object.
}
\description{
Package plot theme
}
\seealso{
Plotting functions:
\code{\link[=enw_delay_categories]{enw_delay_categories()}},
\code{\link[=enw_delay_quantiles]{enw_delay_quantiles()}},
\code{\link[=enw_plot_delay_counts]{enw_plot_delay_counts()}},
\code{\link[=enw_plot_delay_cumulative]{enw_plot_delay_cumulative()}},
\code{\link[=enw_plot_delay_fraction]{enw_plot_delay_fraction()}},
\code{\link[=enw_plot_delay_quantiles]{enw_plot_delay_quantiles()}},
\code{\link[=enw_plot_nowcast_quantiles]{enw_plot_nowcast_quantiles()}},
\code{\link[=enw_plot_obs]{enw_plot_obs()}},
\code{\link[=enw_plot_pp_quantiles]{enw_plot_pp_quantiles()}},
\code{\link[=enw_plot_quantiles]{enw_plot_quantiles()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}},
\code{\link[=plot.epinowcast]{plot.epinowcast()}}
}
\concept{plot}
