% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_posterior_as_prior}
\alias{enw_posterior_as_prior}
\title{FUNCTION_TITLE}
\usage{
enw_posterior_as_prior(
  nowcast,
  priors = epinowcast::enw_priors(),
  variables = c(),
  scale = 5
)
}
\arguments{
\item{nowcast}{DESCRIPTION}

\item{priors}{DESCRIPTION.}

\item{variables}{A character vector of variables both in the
posterior and in the default priors.}

\item{scale}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_opts_as_data_list}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{remove_profiling}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
