% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_preprocess_data}
\alias{enw_preprocess_data}
\title{Preprocess observations}
\usage{
enw_preprocess_data(
  obs,
  by = NULL,
  max_delay,
  timestep = "day",
  set_negatives_to_zero = TRUE,
  ...,
  copy = TRUE
)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\code{reference_date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{max_delay}{The maximum delay to model in the delay
distribution, specified in units of the timestep (e.g., if
\code{timestep = "week"}, then \code{max_delay = 3} means 3 weeks). If not
specified the maximum observed delay is assumed to be the true maximum
delay in the model. Otherwise, an integer greater than or equal to 1
can be specified. Observations with delays larger than the maximum
delay will be dropped. If the specified maximum delay is too short,
nowcasts can be biased as important parts of the true delay
distribution are cut off. At the same time, computational cost scales
non-linearly with this setting, so you want the maximum delay to be
as long as necessary, but not much longer.

Steps to take to determine the maximum delay:
\itemize{
\item Consider what is realistic and relevant for your application.
\item Check the proportion of observations reported (\code{prop_reported})
by delay in the \code{new_confirm} output of \code{enw_preprocess_obs}.
\item Use \code{\link[=check_max_delay]{check_max_delay()}} to check the coverage of a candidate
\code{max_delay}.
\item If in doubt, check if increasing the maximum delay noticeably
changes the delay distribution or nowcasts as estimated by
\code{epinowcast}. If it does, your maximum delay may still be too short.
}

Note that delays are zero indexed and so include the reference date
and \code{max_delay - 1} other intervals (i.e. a \code{max_delay} of 1
corresponds to no delay).}

\item{timestep}{The timestep to used in the process model (i.e. the
reference date model). This can be a string ("day", "week", "month") or a
numeric whole number representing the number of days. If your data does not
have this timestep then you may wish to make use of
\code{\link[=enw_aggregate_cumulative]{enw_aggregate_cumulative()}} to aggregate your data to the desired timestep.}

\item{set_negatives_to_zero}{Logical, defaults to TRUE.
Should negative counts (for calculated incidence of
observations) be set to zero? Currently downstream
modelling does not support negative counts and so
setting must be TRUE if intending to use
\code{\link[=epinowcast]{epinowcast()}}.}

\item{...}{Other arguments to \code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
e.g. \code{holidays}, which sets commonly used metadata
(e.g. day of week, days since start of time series)}

\item{copy}{A logical; if \code{TRUE} (the default) creates a copy; otherwise,
modifies \code{obs} in place.}
}
\value{
A data.table containing processed observations as a series of nested
data.frames as well as variables containing metadata. These are:
\itemize{
\item \code{obs}: (observations with the addition of empirical reporting proportions
and restricted to the specified maximum delay).
\item \code{new_confirm}: Incidence of notifications by reference and report date.
Empirical reporting distributions are also added.
\item \code{latest}: The latest available observations.
\item \code{missing_reference}: Observations missing reference dates.
\item \code{reporting_triangle}: Incident observations by report and reference date in
the standard reporting triangle matrix format.
\item \code{metareference}: Metadata reference dates derived from observations.
\item \code{metrareport}: Metadata for report dates.
\item \code{metadelay}: Metadata for reporting delays produced using
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}}.
\item \code{max_delay}: Maximum delay to be modelled by epinowcast.
\item \code{time}: Numeric, number of timepoints in the data.
\item \code{snapshots}: Numeric, number of available data snapshots to use for
nowcasting.
\item \code{groups}: Numeric, Number of groups/strata in the supplied observations
(set using \code{by}).
\item \code{max_date}: The maximum available report date.
}
}
\description{
This function preprocesses raw observations under the
assumption they are reported as cumulative counts by a reference and
report date and is used to assign groups. It also constructs data objects
used by visualisation and modelling functions including the
observed empirical probability of a report on a given day, the cumulative
probability of report, the latest available observations, incidence of
observations, and metadata about the date of reference and report (used to
construct models). This function wraps other preprocessing functions that may
be instead used individually if required. Note that internally reports
beyond the user specified delay are dropped for modelling purposes with the
\code{cum_prop_reported} and \code{max_confirm} variables allowing the user to check
the impact this may have (if \code{cum_prop_reported} is significantly below 1 a
longer \code{max_delay} may be appropriate). Also note that if missing reference
or report dates are suspected to occur in your data then these need to be
completed with \code{\link[=enw_complete_dates]{enw_complete_dates()}}.
}
\details{
If \code{max_delay} is numeric, it will be internally coerced to integer
using \code{\link[=as.integer]{as.integer()}}).
}
\examples{
library(data.table)

# Filter example hospitalisation data to be national and over all ages
nat_germany_hosp <- germany_covid19_hosp[location == "DE"]
nat_germany_hosp <- nat_germany_hosp[age_group == "00+"]

# Preprocess with default settings
pobs <- enw_preprocess_data(nat_germany_hosp)
pobs
}
\seealso{
Preprocessing functions:
\code{\link[=enw_add_delay]{enw_add_delay()}},
\code{\link[=enw_add_max_reported]{enw_add_max_reported()}},
\code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
\code{\link[=enw_assign_group]{enw_assign_group()}},
\code{\link[=enw_complete_dates]{enw_complete_dates()}},
\code{\link[=enw_construct_data]{enw_construct_data()}},
\code{\link[=enw_extend_date]{enw_extend_date()}},
\code{\link[=enw_filter_delay]{enw_filter_delay()}},
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}},
\code{\link[=enw_filter_reference_dates_by_report_start]{enw_filter_reference_dates_by_report_start()}},
\code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}},
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}},
\code{\link[=enw_impute_na_observations]{enw_impute_na_observations()}},
\code{\link[=enw_latest_data]{enw_latest_data()}},
\code{\link[=enw_metadata]{enw_metadata()}},
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}},
\code{\link[=enw_missing_reference]{enw_missing_reference()}},
\code{\link[=enw_obs_at_delay]{enw_obs_at_delay()}},
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}},
\code{\link[=enw_reporting_triangle_to_long]{enw_reporting_triangle_to_long()}},
\code{\link[=enw_retrospective]{enw_retrospective()}}
}
\concept{preprocess}
