% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_priors}
\alias{enw_priors}
\title{FUNCTION_TITLE}
\usage{
enw_priors()
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
enw_priors()
}
\seealso{
Functions used for modelling
\code{\link{enw_inits}()},
\code{\link{enw_model}()},
\code{\link{enw_obs}()},
\code{\link{enw_sample}()}
}
\concept{model}
