% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{enw_random_intercept_scenario}
\alias{enw_random_intercept_scenario}
\title{FUNCTION_TITLE}
\usage{
enw_random_intercept_scenario(
  obs,
  snapshots = c(30:0),
  logmean = 1.6,
  logmean_sd = 0.1,
  logsd = 1,
  logsd_sd = 0.1
)
}
\arguments{
\item{obs}{PARAM_DESCRIPTION}

\item{snapshots}{PARAM_DESCRIPTION, Default: c(30:0)}

\item{logmean}{PARAM_DESCRIPTION, Default: 1.6}

\item{logmean_sd}{PARAM_DESCRIPTION, Default: 0.1}

\item{logsd}{PARAM_DESCRIPTION, Default: 1}

\item{logsd_sd}{PARAM_DESCRIPTION, Default: 0.1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\concept{scenarios}
