% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_retrospective_data}
\alias{enw_retrospective_data}
\title{FUNCTION_TITLE}
\usage{
enw_retrospective_data(obs, rep_date, rep_days, ref_date, ref_days)
}
\arguments{
\item{obs}{PARAM_DESCRIPTION}

\item{rep_date}{PARAM_DESCRIPTION}

\item{rep_days}{PARAM_DESCRIPTION}

\item{ref_date}{PARAM_DESCRIPTION}

\item{ref_days}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_obs}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
