% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_rolling_sum}
\alias{enw_rolling_sum}
\title{Perform rolling sum aggregation}
\usage{
enw_rolling_sum(dt, internal_timestep, by = NULL, value_col = "confirm")
}
\arguments{
\item{dt}{A \code{data.table} to be aggregated.}

\item{internal_timestep}{An integer indicating the period over which to
aggregate.}

\item{by}{A character vector specifying the columns to aggregate by.}

\item{value_col}{A character string specifying the column to aggregate.
Defaults to "confirm".}
}
\value{
A modified data.table with aggregated observations.
}
\description{
This function takes a data.table and applies a rolling sum over a given
timestep, aggregating by specified columns. It's particularly useful for
aggregating observations over certain periods.
}
\seealso{
Utility functions:
\code{\link[=coerce_date]{coerce_date()}},
\code{\link[=coerce_dt]{coerce_dt()}},
\code{\link[=date_to_numeric_modulus]{date_to_numeric_modulus()}},
\code{\link[=enw_get_data]{enw_get_data()}},
\code{\link[=get_internal_timestep]{get_internal_timestep()}},
\code{\link[=is.Date]{is.Date()}},
\code{\link[=stan_fns_as_string]{stan_fns_as_string()}}
}
\concept{utils}
