% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{enw_simulate_lnorm_trunc_obs}
\alias{enw_simulate_lnorm_trunc_obs}
\title{FUNCTION_TITLE}
\usage{
enw_simulate_lnorm_trunc_obs(scenarios, cases, truncation_max = 20)
}
\arguments{
\item{scenarios}{PARAM_DESCRIPTION}

\item{cases}{PARAM_DESCRIPTION}

\item{truncation_max}{PARAM_DESCRIPTION, Default: 20}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions to generate simulated data
\code{\link{enw_apply_truncation_to_cases}()},
\code{\link{enw_dist_cmf}()}
}
\concept{generatedata}
