% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_stan_to_r}
\alias{enw_stan_to_r}
\title{Expose \code{epinowcast} stan functions in R}
\usage{
enw_stan_to_r(
  files = list.files(include, pattern = "\\\\.stan$"),
  include = system.file("stan", "functions", package = "epinowcast"),
  global = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{files}{A character vector specifying the names of Stan files to be
exposed. These must be in the \code{include} directory. Defaults to all Stan
files in the \code{include} directory. Note that the following files contain
overloaded functions and cannot be exposed: "delay_lpmf.stan",
"allocate_observed_obs.stan", "obs_lpmf.stan", and "effects_priors_lp.stan".
Functions implemented in C++ (currently "logit_hazard_to_log_prob.stan",
see \code{\link[=enw_model]{enw_model()}}'s \code{use_cpp} argument) are exposed via their retained
pure-Stan fallback instead, since \code{expose_functions()} cannot link the
package's C++ header. The two agree to floating-point precision except
where a hazard rounds to 0 or 1, where the C++ version stays finite.}

\item{include}{A character string specifying the directory containing Stan
files. Defaults to the 'stan/functions' directory of the \code{\link[=epinowcast]{epinowcast()}}
package.}

\item{global}{A logical value indicating whether to expose the functions
globally. Defaults to \code{TRUE}. Passed to the \link[cmdstanr]{expose_functions}
method of \link[cmdstanr]{CmdStanModel}.}

\item{verbose}{Logical, defaults to \code{TRUE}. Should verbose
messages be shown.}

\item{...}{Additional arguments passed to \url{enw_model}.}
}
\value{
An object of class \code{CmdStanModel} with functions from the model
exposed for use in R.
}
\description{
This function facilitates the exposure of Stan functions from
the \url{epinowcast} package in R. It utilizes the
\link[cmdstanr]{expose_functions} method of \link[cmdstanr:CmdStanModel]{cmdstanr::CmdStanModel} for
this purpose. This function is useful for
developers and contributors to the \link{epinowcast} package, as well as for
users interested in exploring and prototyping with model functionalities.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Compile functions in stan/functions/hazard.stan
stan_functions <- enw_stan_to_r("hazard.stan")
# These functions can now be used in R
stan_functions$functions$prob_to_hazard(c(0.5, 0.1, 0.1))
# or exposed globally and used directly
prob_to_hazard(c(0.5, 0.1, 0.1))
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_model}()},
\code{\link{enw_pathfinder}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_unset_cache}()},
\code{\link{epinowcast_stan_header}()},
\code{\link{remove_profiling}()},
\code{\link{stage_stan_header}()},
\code{\link{stan_cpp_adjoint_files}()},
\code{\link{stan_cpp_fallback_body}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
