% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{enw_startup_message}
\alias{enw_startup_message}
\title{Startup message for epinowcast package}
\usage{
enw_startup_message()
}
\value{
A character vector containing messages. If \code{enw_cache_location} is
not set, the function returns instructions on how to set it and where to
find more details. If it is set, it returns a confirmation message of the
current cache location.
}
\description{
This function generates a startup message for the \code{\link[=epinowcast]{epinowcast()}} package.
It checks the environment setting for the cache location and provides
appropriate guidance to the user based on the setting.
}
\details{
\code{enw_startup_message} checks if the \code{enw_cache_location}
environment variable is set. If it is not set, the function advises the
user to set this variable using the \code{enw_set_cache} function. This helps in
reducing future Stan compilation times. If the \code{enw_cache_location} is set,
the function confirms the current cache location to the user.
\code{enw_cache_location} can be set using \code{\link[=enw_set_cache]{enw_set_cache()}}.
}
\keyword{internal}
