% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_summarise_samples}
\alias{enw_summarise_samples}
\title{Summarise posterior samples}
\usage{
enw_summarise_samples(
  samples,
  probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95),
  by = c("reference_date", ".group"),
  link_with_obs = TRUE
)
}
\arguments{
\item{samples}{A \code{data.frame} of posterior samples with at least a numeric
sample variable.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work.}

\item{by}{A character vector of variables to summarise by. Defaults to
\code{c("reference_date", ".group")}.}

\item{link_with_obs}{Logical, should the observed data be linked to the
posterior summary? This is useful for plotting the posterior against the
observed data. Defaults to \code{TRUE}.}
}
\value{
A \code{data.frame} summarising the posterior samples.
}
\description{
This function summarises posterior samples for arbitrary
strata. It optionally holds out the observed data (variables that are not
".draw", ".iteration", ".sample", ".chain" ) joins this to the summarised
posterior.
}
\examples{
fit <- enw_example("nowcast")
samples <- summary(fit, type = "nowcast_sample")
enw_summarise_samples(samples, probs = c(0.05, 0.5, 0.95))
}
\seealso{
Functions used for postprocessing of model fits:
\code{\link[=.check_primarycensored]{.check_primarycensored()}},
\code{\link[=.delay_draw_columns]{.delay_draw_columns()}},
\code{\link[=.discretise_parametric_pmf]{.discretise_parametric_pmf()}},
\code{\link[=build_ord_obs]{build_ord_obs()}},
\code{\link[=enw_add_latest_obs_to_nowcast]{enw_add_latest_obs_to_nowcast()}},
\code{\link[=enw_nowcast_samples]{enw_nowcast_samples()}},
\code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}},
\code{\link[=enw_posterior]{enw_posterior()}},
\code{\link[=enw_posterior_delay]{enw_posterior_delay()}},
\code{\link[=enw_pp_summary]{enw_pp_summary()}},
\code{\link[=enw_quantiles_to_long]{enw_quantiles_to_long()}},
\code{\link[=subset_obs]{subset_obs()}}
}
\concept{postprocess}
