% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_unset_cache}
\alias{enw_unset_cache}
\title{Unset Stan cache location}
\usage{
enw_unset_cache(type = c("session", "persistent", "all"))
}
\arguments{
\item{type}{A character string specifying the type of cache to unset.
It can be one of "session", "persistent", or "all". Default is "session".
"session" unsets the cache for the current session, "persistent" removes the
cache location from the user's \code{.Renviron} file,and "all" does all options.}
}
\value{
The prior cache location, if it existed otherwise \code{NULL}.
}
\description{
Optionally removes the \code{enw_cache_location} environment variable from
the user .Renviron file and/or removes it from the local
environment. If you unset the local cache and want to switch
back to using the persistent cache, you can reload the
\code{.Renviron} file using \code{readRenviron("~/.Renviron")}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
enw_unset_cache()
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_model}()},
\code{\link{enw_pathfinder}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_stan_to_r}()},
\code{\link{epinowcast_stan_header}()},
\code{\link{remove_profiling}()},
\code{\link{stage_stan_header}()},
\code{\link{stan_cpp_adjoint_files}()},
\code{\link{stan_cpp_fallback_body}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
