% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-designs.R
\name{enw_weekly_model}
\alias{enw_weekly_model}
\title{FUNCTION_TITLE}
\usage{
enw_weekly_model(metaobs, rw = FALSE, day_of_week = FALSE, holidays = c())
}
\arguments{
\item{metaobs}{PARAM_DESCRIPTION}

\item{rw}{PARAM_DESCRIPTION, Default: FALSE}

\item{day_of_week}{PARAM_DESCRIPTION, Default: FALSE}

\item{holidays}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
Functions used to formulate models
\code{\link{enw_day_model}()},
\code{\link{enw_day_of_week_model}()},
\code{\link{enw_intercept_model}()}
}
\concept{modeldesign}
