% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expose_stan_fns}
\alias{expose_stan_fns}
\title{FUNCTION_TITLE}
\usage{
expose_stan_fns(files, target_dir, ...)
}
\arguments{
\item{files}{PARAM_DESCRIPTION}

\item{target_dir}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Utility functions
\code{\link{inv_logit}()},
\code{\link{is.Date}()},
\code{\link{logit}()}
}
\concept{utils}
