% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_environment_contents}
\alias{get_environment_contents}
\title{Identify cache location}
\usage{
get_environment_contents()
}
\value{
A list containing the contents of the \code{.Renviron} file and its path.
}
\description{
This function retrieves environment variable settings and manages the
\code{.Renviron} file in the user's project or home directory.
The project directory will be examined first, if it exists.
}
\keyword{internal}
