% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_internal_timestep}
\alias{get_internal_timestep}
\title{Get internal timestep}
\usage{
get_internal_timestep(timestep)
}
\arguments{
\item{timestep}{The timestep to used. This can be a string ("day",
"week") or a numeric whole number representing the number of days.
Note that "month" is not currently supported in user-facing functions
and will throw an error if used.}
}
\value{
A numeric value representing the number of days for "day" and
"week", or the input value if it is a numeric whole number.
}
\description{
This function converts the string representation of the timestep to its
corresponding numeric value or returns the numeric input (if it is a whole
number). For "day", "week", it returns 1 and 7 respectively.
"month" is not supported and will throw an error.
If the input is a numeric whole number, it is returned as is.
}
\seealso{
Utility functions:
\code{\link[=coerce_date]{coerce_date()}},
\code{\link[=coerce_dt]{coerce_dt()}},
\code{\link[=date_to_numeric_modulus]{date_to_numeric_modulus()}},
\code{\link[=enw_get_data]{enw_get_data()}},
\code{\link[=enw_rolling_sum]{enw_rolling_sum()}},
\code{\link[=is.Date]{is.Date()}},
\code{\link[=stan_fns_as_string]{stan_fns_as_string()}}
}
\concept{utils}
