% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_logit}
\alias{inv_logit}
\title{FUNCTION_TITLE}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
inv_logit(c(-10, 1, 0, 100))
}
\seealso{
Utility functions
\code{\link{expose_stan_fns}()},
\code{\link{logit}()}
}
\concept{utils}
