% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{linelist}
\alias{linelist}
\title{A fake linelist to test data conversion functions}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 310 rows and 6 columns.
}
\usage{
linelist
}
\value{
A \code{data.table}
}
\description{
A fake linelist to test data conversion functions
}
\seealso{
Package data sets
\code{\link{enw_example}()},
\code{\link{germany_covid19_hosp}}
}
\concept{data}
\keyword{datasets}
