% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{FUNCTION_TITLE}
\usage{
logit(p)
}
\arguments{
\item{p}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
logit(c(0.01, 0.1, 0.5, 0.9, 1))
}
\seealso{
Utility functions
\code{\link{expose_stan_fns}()},
\code{\link{inv_logit}()},
\code{\link{is.Date}()}
}
\concept{utils}
