% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.epinowcast}
\alias{plot.epinowcast}
\title{Plot method for epinowcast}
\usage{
\method{plot}{epinowcast}(
  x,
  latest_obs = NULL,
  type = c("nowcast", "posterior_prediction"),
  log = FALSE,
  ...
)
}
\arguments{
\item{x}{A \code{data.table} of output as produced by \code{\link[=epinowcast]{epinowcast()}}.}

\item{latest_obs}{A \code{data.frame} of observed data which may be passed to
lower level methods.}

\item{type}{Character string indicating the plot required; enforced by
\code{\link[base:match.arg]{base::match.arg()}}. Currently supported options:
\itemize{
\item "nowcast" which plots the nowcast for each dataset along with latest
available observed data using \code{\link[=enw_plot_nowcast_quantiles]{enw_plot_nowcast_quantiles()}},
\item "posterior_prediction" which plots observations reported at the time
against simulated observations from the model using
\code{\link[=enw_plot_pp_quantiles]{enw_plot_pp_quantiles()}}.
}}

\item{log}{Logical, defaults to \code{FALSE}. Should counts be plot on the log
scale.}

\item{...}{Additional arguments to the plot function specified by \code{type}.}
}
\value{
\code{ggplot2} object
}
\description{
\code{plot} method for class "epinowcast".
}
\examples{
nowcast <- enw_example("nowcast")
latest_obs <- enw_example("obs")

# Plot nowcast
plot(nowcast, latest_obs = latest_obs, type = "nowcast")

# Plot posterior predictions by reference date
plot(nowcast, type = "posterior_prediction") +
 ggplot2::facet_wrap(ggplot2::vars(reference_date), scales = "free")
}
\seealso{
Other epinowcast:
\code{\link[=epinowcast]{epinowcast()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}},
\code{\link[=print.enw_preprocess_data]{print.enw_preprocess_data()}},
\code{\link[=print.epinowcast]{print.epinowcast()}},
\code{\link[=print.summary.enw_preprocess_data]{print.summary.enw_preprocess_data()}},
\code{\link[=summary.enw_preprocess_data]{summary.enw_preprocess_data()}},
\code{\link[=summary.epinowcast]{summary.epinowcast()}}

Plotting functions:
\code{\link[=enw_delay_categories]{enw_delay_categories()}},
\code{\link[=enw_delay_quantiles]{enw_delay_quantiles()}},
\code{\link[=enw_plot_delay_counts]{enw_plot_delay_counts()}},
\code{\link[=enw_plot_delay_cumulative]{enw_plot_delay_cumulative()}},
\code{\link[=enw_plot_delay_fraction]{enw_plot_delay_fraction()}},
\code{\link[=enw_plot_delay_quantiles]{enw_plot_delay_quantiles()}},
\code{\link[=enw_plot_nowcast_quantiles]{enw_plot_nowcast_quantiles()}},
\code{\link[=enw_plot_obs]{enw_plot_obs()}},
\code{\link[=enw_plot_pp_quantiles]{enw_plot_pp_quantiles()}},
\code{\link[=enw_plot_quantiles]{enw_plot_quantiles()}},
\code{\link[=enw_plot_theme]{enw_plot_theme()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}}
}
\concept{epinowcast}
\concept{plot}
