% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_theme}
\alias{plot_theme}
\title{FUNCTION_TITLE}
\usage{
plot_theme(plot)
}
\arguments{
\item{plot}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\concept{plot}
