% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.enw_preprocess_data}
\alias{print.enw_preprocess_data}
\title{Print method for enw_preprocess_data}
\usage{
\method{print}{enw_preprocess_data}(x, ...)
}
\arguments{
\item{x}{A \code{data.table} output from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}} or \code{\link[=enw_construct_data]{enw_construct_data()}}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
\code{print} method for class
\code{"enw_preprocess_data"}.
}
\examples{
pobs <- enw_example("preprocessed_observations")
pobs
}
\seealso{
Other epinowcast:
\code{\link[=epinowcast]{epinowcast()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}},
\code{\link[=plot.epinowcast]{plot.epinowcast()}},
\code{\link[=print.epinowcast]{print.epinowcast()}},
\code{\link[=print.summary.enw_preprocess_data]{print.summary.enw_preprocess_data()}},
\code{\link[=summary.enw_preprocess_data]{summary.enw_preprocess_data()}},
\code{\link[=summary.epinowcast]{summary.epinowcast()}}
}
\concept{epinowcast}
