% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{re}
\alias{re}
\title{Defines random effect terms using the lme4 syntax}
\usage{
re(formula)
}
\arguments{
\item{formula}{A random effect as returned by \link[reformulas]{findbars}
when a random effect is defined using the \link[lme4]{lme4} syntax in
formula. Currently only simplified random effects (i.e
LHS | RHS) are supported.}
}
\value{
A list defining the fixed and random effects of the specified
random effect
}
\description{
Defines random effect terms using the lme4 syntax
}
\examples{
form <- epinowcast:::parse_formula(~ 1 + (1 | age_group))
re(form$random[[1]])

form <- epinowcast:::parse_formula(~ 1 + (location | age_group))
re(form$random[[1]])
}
\seealso{
Functions used to help convert formulas into model designs:
\code{\link[=ar]{ar()}},
\code{\link[=arima]{arima()}},
\code{\link[=arima_terms]{arima_terms()}},
\code{\link[=arma]{arma()}},
\code{\link[=as_string_formula]{as_string_formula()}},
\code{\link[=construct_arima]{construct_arima()}},
\code{\link[=construct_gp]{construct_gp()}},
\code{\link[=construct_re]{construct_re()}},
\code{\link[=construct_rw]{construct_rw()}},
\code{\link[=enw_formula]{enw_formula()}},
\code{\link[=enw_manual_formula]{enw_manual_formula()}},
\code{\link[=gp]{gp()}},
\code{\link[=gp_terms]{gp_terms()}},
\code{\link[=ma]{ma()}},
\code{\link[=parse_formula]{parse_formula()}},
\code{\link[=remove_arima_terms]{remove_arima_terms()}},
\code{\link[=remove_gp_terms]{remove_gp_terms()}},
\code{\link[=remove_rw_terms]{remove_rw_terms()}},
\code{\link[=rw]{rw()}},
\code{\link[=rw_terms]{rw_terms()}},
\code{\link[=split_formula_to_terms]{split_formula_to_terms()}}
}
\concept{formulatools}
