% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-validation.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_forecast_sample}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{scoringutils}{\code{\link[scoringutils]{as_forecast_sample}}}
}}

