% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{remove_arima_terms}
\alias{remove_arima_terms}
\title{Remove ARIMA terms from a formula object}
\usage{
remove_arima_terms(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link[lme4]{lme4} syntax (see \code{\link[=re]{re()}}), and
random walks defined using the \code{\link[=rw]{rw()}} helper function. See the Details
section below for a comprehensive explanation of the supported syntax.}
}
\value{
A formula object with the ARIMA terms removed.
}
\description{
This function removes ARIMA terms — \code{arima()}, \code{ar()},
\code{ma()}, and \code{arma()} — from a formula so they can be processed on
their own.
}
\examples{
epinowcast:::remove_arima_terms(~ 1 + age_group + arima(week))
epinowcast:::remove_arima_terms(~ 1 + age_group + ar(week, p = 2))
}
\seealso{
Functions used to help convert formulas into model designs:
\code{\link[=ar]{ar()}},
\code{\link[=arima]{arima()}},
\code{\link[=arima_terms]{arima_terms()}},
\code{\link[=arma]{arma()}},
\code{\link[=as_string_formula]{as_string_formula()}},
\code{\link[=construct_arima]{construct_arima()}},
\code{\link[=construct_gp]{construct_gp()}},
\code{\link[=construct_re]{construct_re()}},
\code{\link[=construct_rw]{construct_rw()}},
\code{\link[=enw_formula]{enw_formula()}},
\code{\link[=enw_manual_formula]{enw_manual_formula()}},
\code{\link[=gp]{gp()}},
\code{\link[=gp_terms]{gp_terms()}},
\code{\link[=ma]{ma()}},
\code{\link[=parse_formula]{parse_formula()}},
\code{\link[=re]{re()}},
\code{\link[=remove_gp_terms]{remove_gp_terms()}},
\code{\link[=remove_rw_terms]{remove_rw_terms()}},
\code{\link[=rw]{rw()}},
\code{\link[=rw_terms]{rw_terms()}},
\code{\link[=split_formula_to_terms]{split_formula_to_terms()}}
}
\concept{formulatools}
