% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{remove_rw_terms}
\alias{remove_rw_terms}
\title{Remove random walk terms from a formula object}
\usage{
remove_rw_terms(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link[lme4]{lme4} syntax (see \code{\link[=re]{re()}}), and
random walks defined using the \code{\link[=rw]{rw()}} helper function. See the Details
section below for a comprehensive explanation of the supported syntax.}
}
\value{
A formula object with the random walk terms removed.
}
\description{
This function removes random walk terms
denoted using \code{\link[=rw]{rw()}} from a formula so that they can be
processed on their own.
}
\section{Reference}{

This function was adapted from code written
by J Scott (under an MIT license) as part of
the \code{epidemia} package (https://github.com/ImperialCollegeLondon/epidemia/).
}

\examples{
epinowcast:::remove_rw_terms(~ 1 + age_group + location)

epinowcast:::remove_rw_terms(~ 1 + age_group + location + rw(week, location))
}
\seealso{
Functions used to help convert formulas into model designs:
\code{\link[=ar]{ar()}},
\code{\link[=arima]{arima()}},
\code{\link[=arima_terms]{arima_terms()}},
\code{\link[=arma]{arma()}},
\code{\link[=as_string_formula]{as_string_formula()}},
\code{\link[=construct_arima]{construct_arima()}},
\code{\link[=construct_gp]{construct_gp()}},
\code{\link[=construct_re]{construct_re()}},
\code{\link[=construct_rw]{construct_rw()}},
\code{\link[=enw_formula]{enw_formula()}},
\code{\link[=enw_manual_formula]{enw_manual_formula()}},
\code{\link[=gp]{gp()}},
\code{\link[=gp_terms]{gp_terms()}},
\code{\link[=ma]{ma()}},
\code{\link[=parse_formula]{parse_formula()}},
\code{\link[=re]{re()}},
\code{\link[=remove_arima_terms]{remove_arima_terms()}},
\code{\link[=remove_gp_terms]{remove_gp_terms()}},
\code{\link[=rw]{rw()}},
\code{\link[=rw_terms]{rw_terms()}},
\code{\link[=split_formula_to_terms]{split_formula_to_terms()}}
}
\concept{formulatools}
