% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{remove_secondary_terms}
\alias{remove_secondary_terms}
\title{Remove secondary stratum terms from a formula object}
\usage{
remove_secondary_terms(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link[lme4]{lme4} syntax (see \code{\link[=re]{re()}}), and
random walks defined using the \code{\link[=rw]{rw()}} helper function. See the Details
section below for a comprehensive explanation of the supported syntax.}
}
\value{
A formula object with the secondary stratum terms removed.
}
\description{
This function removes secondary stratum terms denoted
using \code{\link[=secondary]{secondary()}} from a formula so they can be processed on their
own.
}
\examples{
epinowcast:::remove_secondary_terms(~ 1 + secondary(cases))
epinowcast:::remove_secondary_terms(
  ~ 1 + secondary(cases, ascertainment = ~ 1 + (1 | region))
)
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{ar}()},
\code{\link{arima}()},
\code{\link{arima_terms}()},
\code{\link{arma}()},
\code{\link{as_string_formula}()},
\code{\link{construct_arima}()},
\code{\link{construct_gp}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{construct_secondary}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()},
\code{\link{gp}()},
\code{\link{gp_terms}()},
\code{\link{ma}()},
\code{\link{parse_formula}()},
\code{\link{re}()},
\code{\link{remove_arima_terms}()},
\code{\link{remove_gp_terms}()},
\code{\link{remove_rw_terms}()},
\code{\link{rw}()},
\code{\link{rw_terms}()},
\code{\link{secondary}()},
\code{\link{secondary_terms}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
