% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{sample_random_init}
\alias{sample_random_init}
\title{Initialize and sample using random initialization}
\usage{
sample_random_init(data, model, init, ...)
}
\arguments{
\item{data}{A list of data as produced by model modules (for example
\code{\link[=enw_expectation]{enw_expectation()}}, \code{\link[=enw_obs]{enw_obs()}}, etc.) and as required for use the
\code{model} being used.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}} or as
supplied by the user.}

\item{init}{A list of initial values or a function to generate initial
values. If not provided, the model will attempt to generate initial values}

\item{...}{Additional arguments passed to the \code{sample} method of \code{cmdstanr}.}
}
\value{
A data.table containing the fit, data, and fit arguments
}
\description{
This function samples from the model using random initialization.
}
