% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{simulate_double_censored_pmf}
\alias{simulate_double_censored_pmf}
\title{Simulate daily double censored PMF}
\usage{
simulate_double_censored_pmf(
  max,
  fun_primary = stats::runif,
  primary_args = list(),
  fun_dist = stats::rlnorm,
  dist_args = list(...),
  n = 1e+06,
  ...
)
}
\arguments{
\item{max}{Maximum value for the computed CDF. If not specified, the maximum
value is the maximum simulated delay.}

\item{fun_primary}{Primary distribution function (default is \code{runif}).}

\item{primary_args}{List of additional arguments to be passed to the primary
distribution function.}

\item{fun_dist}{Distribution function to be added to the primary (default is
\code{rlnorm}).}

\item{dist_args}{List of additional arguments to be passed to the
distribution function.}

\item{n}{Number of simulations (default is 1e6).}

\item{...}{Additional arguments to be passed to the distribution function.
This is an alternative to \code{dist_args}.}
}
\value{
A numeric vector representing the PMF.
}
\description{
This function simulates the probability mass function of a  daily
double-censored process. The process involves two distributions: a primary
distribution which represents the censoring process for the primary event
and another distribution (which is offset by the primary).
}
\examples{
simulate_double_censored_pmf(10, meanlog = 0, sdlog = 1)
}
\seealso{
Helper functions for model modules
\code{\link{add_pmfs}()},
\code{\link{convolution_matrix}()},
\code{\link{enw_reference_by_report}()},
\code{\link{enw_reps_with_complete_refs}()},
\code{\link{extract_obs_metadata}()},
\code{\link{extract_sparse_matrix}()},
\code{\link{latest_obs_as_matrix}()}
}
\concept{modelmodulehelpers}
