% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{subset_obs}
\alias{subset_obs}
\title{Subset observations data table for either modelled dates
or not-modelled earlier dates.}
\usage{
subset_obs(ord_obs, max_delay, internal_timestep, reference_subset)
}
\arguments{
\item{ord_obs}{The observations \code{data.table} to be subset,
as pulled from the result of calling \code{\link[=epinowcast]{epinowcast()}} and
coerced to a data table.}

\item{max_delay}{Whole number representing the maximum delay
in units of the timestep.}

\item{internal_timestep}{A numeric value representing the number
of days in the timestep, e.g. 7 when the timesteps are weeks.}

\item{reference_subset}{String giving a relational operator
to subset ord_obs by reference date; e.g. \code{>} to keep the
modelled reference dates from after the max_delay.}
}
\value{
A \code{data.frame} subset for the desired observations
}
\description{
Subset observations data table for either modelled dates
or not-modelled earlier dates.
}
\seealso{
Functions used for postprocessing of model fits:
\code{\link[=.check_primarycensored]{.check_primarycensored()}},
\code{\link[=.delay_draw_columns]{.delay_draw_columns()}},
\code{\link[=.discretise_parametric_pmf]{.discretise_parametric_pmf()}},
\code{\link[=build_ord_obs]{build_ord_obs()}},
\code{\link[=enw_add_latest_obs_to_nowcast]{enw_add_latest_obs_to_nowcast()}},
\code{\link[=enw_nowcast_samples]{enw_nowcast_samples()}},
\code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}},
\code{\link[=enw_posterior]{enw_posterior()}},
\code{\link[=enw_posterior_delay]{enw_posterior_delay()}},
\code{\link[=enw_pp_summary]{enw_pp_summary()}},
\code{\link[=enw_quantiles_to_long]{enw_quantiles_to_long()}},
\code{\link[=enw_summarise_samples]{enw_summarise_samples()}}
}
\concept{postprocess}
