% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.enw_preprocess_data}
\alias{summary.enw_preprocess_data}
\title{Summary method for enw_preprocess_data}
\usage{
\method{summary}{enw_preprocess_data}(object, n = 6, ...)
}
\arguments{
\item{object}{A \code{data.table} output from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}} or \code{\link[=enw_construct_data]{enw_construct_data()}}.}

\item{n}{Integer number of rows to show in previews.
Defaults to 6.}

\item{...}{Additional arguments (not used).}
}
\value{
A list of class \code{"summary.enw_preprocess_data"}
containing the preprocessed data object and preview
parameters, printed via
\code{\link[=print.summary.enw_preprocess_data]{print.summary.enw_preprocess_data()}}.
}
\description{
\code{summary} method for class
\code{"enw_preprocess_data"}. Returns a structured overview of
the preprocessed data including a preview of the latest
observations and a corner of the reporting triangle.
}
\examples{
pobs <- enw_example("preprocessed_observations")
summary(pobs)
}
\seealso{
Other epinowcast:
\code{\link[=epinowcast]{epinowcast()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}},
\code{\link[=plot.epinowcast]{plot.epinowcast()}},
\code{\link[=print.enw_preprocess_data]{print.enw_preprocess_data()}},
\code{\link[=print.epinowcast]{print.epinowcast()}},
\code{\link[=print.summary.enw_preprocess_data]{print.summary.enw_preprocess_data()}},
\code{\link[=summary.epinowcast]{summary.epinowcast()}}
}
\concept{epinowcast}
