% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.epinowcast}
\alias{summary.epinowcast}
\title{Summary method for epinowcast}
\usage{
\method{summary}{epinowcast}(
  object,
  type = c("nowcast", "nowcast_samples", "fit", "posterior_prediction"),
  max_delay = object$max_delay,
  ...
)
}
\arguments{
\item{object}{A \code{data.table} output from \code{\link[=epinowcast]{epinowcast()}}.}

\item{type}{Character string indicating the summary to return; enforced by
\code{\link[base:match.arg]{base::match.arg()}}. Supported options are:
\itemize{
\item "nowcast" which summarises nowcast posterior with \code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}},
\item "nowcast_samples" which samples latest with \code{\link[=enw_nowcast_samples]{enw_nowcast_samples()}},
\item "fit" which returns the summarised \code{cmdstanr} fit with \code{\link[=enw_posterior]{enw_posterior()}},
\item "posterior_prediction" which returns summarised posterior predictions for
the observations after fitting using \code{\link[=enw_pp_summary]{enw_pp_summary()}}.
}}

\item{max_delay}{Maximum delay to which nowcasts should be
summarised, in units of the timestep used during preprocessing. Must
be equal (default) or larger than the modelled maximum delay. If it is
larger, then nowcasts for unmodelled dates are added by assuming that
case counts beyond the modelled maximum delay are fully observed.}

\item{...}{Additional arguments passed to summary specified by \code{type}.}
}
\value{
A summary data.frame
}
\description{
\code{summary} method for class "epinowcast".
}
\examples{
nowcast <- enw_example("nowcast")

# Summarise nowcast posterior
summary(nowcast, type = "nowcast")

# Nowcast posterior samples
summary(nowcast, type = "nowcast_samples")

# Nowcast model fit
summary(nowcast, type = "fit")

# Posterior predictions
summary(nowcast, type = "posterior_prediction")
}
\seealso{
summary epinowcast

Other epinowcast:
\code{\link[=epinowcast]{epinowcast()}},
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}},
\code{\link[=plot.epinowcast]{plot.epinowcast()}},
\code{\link[=print.enw_preprocess_data]{print.enw_preprocess_data()}},
\code{\link[=print.epinowcast]{print.epinowcast()}},
\code{\link[=print.summary.enw_preprocess_data]{print.summary.enw_preprocess_data()}},
\code{\link[=summary.enw_preprocess_data]{summary.enw_preprocess_data()}}
}
\concept{epinowcast}
