% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{terms_rw}
\alias{terms_rw}
\title{Finds random walk terms in a formula object}
\usage{
terms_rw(x)
}
\arguments{
\item{x}{An object of class "formula"}
}
\value{
The value of attributes. See \code{\link[base]{attr}} for more
details.
}
\description{
Adapted from code in \code{epidemia} written by J. Scott.
}
