% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{update_inits}
\alias{update_inits}
\title{Update initial values for model fitting}
\usage{
update_inits(
  data,
  model,
  init,
  init_method = c("prior", "pathfinder"),
  init_method_args = list(),
  ...
)
}
\arguments{
\item{data}{A list of data as produced by model modules (for example
\code{\link[=enw_expectation]{enw_expectation()}}, \code{\link[=enw_obs]{enw_obs()}}, etc.) and as required for use the
\code{model} being used.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}} or as
supplied by the user.}

\item{init}{A list of initial values or a function to generate initial
values. If not provided, the model will attempt to generate initial values}

\item{init_method}{The method to use for initializing the model. Defaults to
"prior" which samples initial values from the prior. "pathfinder", which uses
the pathfinder algorithm (\code{\link[=enw_pathfinder]{enw_pathfinder()}}) to initialize the model.}

\item{init_method_args}{A list of additional arguments to pass to the
initialization method.}

\item{...}{Additional arguments passed to initialization methods.}
}
\value{
A list containing updated initial values and method-specific output.
}
\description{
This function updates the initial values for model fitting based on the
specified initialization method.
}
\keyword{internal}
