% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{add_pmfs}
\alias{add_pmfs}
\title{Add probability mass functions}
\usage{
add_pmfs(pmfs)
}
\arguments{
\item{pmfs}{A list of vectors describing the probability mass functions to}
}
\value{
A vector describing the probability mass function of the sum of the
}
\description{
This function allows the addition of probability mass functions (PMFs) to
produce a new PMF. This is useful for example in the context of reporting
delays where the PMF of the sum of two Poisson distributions is the
convolution of the PMFs.
}
\examples{
# Sample and analytical PMFs for two Poisson distributions
x <- rpois(10000, 5)
xpmf <- dpois(0:20, 5)
y <- rpois(10000, 7)
ypmf <- dpois(0:20, 7)
# Add sampled Poisson distributions up to get combined distribution
z <- x + y
# Analytical convolution of PMFs
conv_pmf <- add_pmfs(list(xpmf, ypmf))
conv_cdf <- cumsum(conv_pmf)
# Empirical convolution of PMFs
cdf <- ecdf(z)(0:42)
# Compare sampled and analytical CDFs
plot(conv_cdf)
lines(cdf, col = "black")
}
\seealso{
Helper functions for model modules
\code{\link{convolution_matrix}()},
\code{\link{enw_reference_by_report}()},
\code{\link{enw_reps_with_complete_refs}()},
\code{\link{extract_obs_metadata}()},
\code{\link{extract_sparse_matrix}()},
\code{\link{latest_obs_as_matrix}()},
\code{\link{simulate_double_censored_pmf}()}
}
\concept{modelmodulehelpers}
