% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{arima_terms}
\alias{arima_terms}
\title{Finds ARIMA terms in a formula object}
\usage{
arima_terms(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link[lme4]{lme4} syntax (see \code{\link[=re]{re()}}), and
random walks defined using the \code{\link[=rw]{rw()}} helper function. See the Details
section below for a comprehensive explanation of the supported syntax.}
}
\value{
A character vector containing the ARIMA terms identified in
the supplied formula.
}
\description{
This function extracts ARIMA terms denoted using
\code{\link[=arima]{arima()}} from a formula so that they can be processed on their own.
}
\examples{
epinowcast:::arima_terms(~ 1 + age_group + arima(week))

epinowcast:::arima_terms(
  ~ 1 + age_group + arima(week, location, p = 1, d = 1, q = 1)
)
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{arima}()},
\code{\link{as_string_formula}()},
\code{\link{construct_arima}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()},
\code{\link{parse_formula}()},
\code{\link{re}()},
\code{\link{remove_arima_terms}()},
\code{\link{remove_rw_terms}()},
\code{\link{rw}()},
\code{\link{rw_terms}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
