% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-validation.R
\name{as_forecast_sample.epinowcast}
\alias{as_forecast_sample.epinowcast}
\title{Convert an epinowcast object to a forecast_sample object}
\usage{
as_forecast_sample.epinowcast(nowcast, latest_obs, ...)
}
\arguments{
\item{nowcast}{An \code{epinowcast} nowcast object as returned by
\code{\link[=epinowcast]{epinowcast()}}.}

\item{latest_obs}{Latest observations to use for the true values must
contain \code{confirm} and \code{observed} variables.}

\item{...}{Additional arguments passed to
\code{\link[scoringutils:as_forecast_sample]{scoringutils::as_forecast_sample()}}}
}
\value{
A \code{forecast_sample} object as returned by
\code{\link[scoringutils:as_forecast_sample]{scoringutils::as_forecast_sample()}}
}
\description{
This function is used to convert an \code{epinowcast} as returned by
\code{\link[=epinowcast]{epinowcast()}} object to a \code{forecast_sample} object which can be used for
scoring using the \code{scoringutils} package.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(scoringutils)

nowcast <- enw_example("nowcast")
latest_obs <- enw_example("observations")
as_forecast_sample(nowcast, latest_obs)
\dontshow{\}) # examplesIf}
}
\seealso{
Other modelvalidation: 
\code{\link{enw_score_nowcast}()}
}
\concept{modelvalidation}
