% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{as_string_formula}
\alias{as_string_formula}
\title{FUNCTION_TITLE}
\usage{
as_string_formula(x)
}
\arguments{
\item{x}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()},
\code{\link{terms_rw}()}
}
\concept{formulatools}
