% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_calendar_timestep}
\alias{check_calendar_timestep}
\title{Check calendar timestep}
\usage{
check_calendar_timestep(dates, date_var, exact = TRUE)
}
\arguments{
\item{dates}{Vector of Date class representing dates.}

\item{date_var}{The variable in \code{obs} representing dates.}

\item{exact}{Logical, if \verb{TRUE``, checks if all differences exactly match the timestep. If }FALSE``, checks if the sum of the differences modulo the
timestep equals zero. Default is \code{TRUE}.}
}
\value{
This function is used for its side effect of stopping if the check
fails. If the check passes, the function returns invisibly.
}
\description{
This function verifies if the difference in calendar dates in the provided
observations corresponds to the provided timestep of "month".
}
\seealso{
Functions used for checking inputs
\code{\link{check_group}()},
\code{\link{check_group_date_unique}()},
\code{\link{check_max_delay}()},
\code{\link{check_module}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_numeric_timestep}()},
\code{\link{check_observation_indicator}()},
\code{\link{check_quantiles}()},
\code{\link{check_timestep}()},
\code{\link{check_timestep_by_date}()},
\code{\link{check_timestep_by_group}()}
}
\concept{check}
