% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dates}
\alias{check_dates}
\title{Check report and reference dates are present in the observations.}
\usage{
check_dates(obs)
}
\arguments{
\item{obs}{An observation \code{data.frame} containing \code{report_date} and
\code{reference_date} columns.}
}
\value{
Returns the input \code{data.frame} with dates converted to date format
if not already.
}
\description{
Check report and reference dates are present in the observations.
}
\seealso{
Functions used for checking inputs
\code{\link{add_group}()},
\code{\link{check_by}()},
\code{\link{check_group}()},
\code{\link{check_max_delay}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
