% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dates}
\alias{check_dates}
\title{Check for and Format Report and Reference Dates}
\usage{
check_dates(obs)
}
\arguments{
\item{obs}{An object that will be \code{\link[=coerce_dt]{coerce_dt()}}d in place; must have
\code{report_date} and \code{reference_date} columns.}
}
\value{
The \code{obs} object, modified in place (if necessary), where
\code{report_date} and \code{reference_date} are cast via \code{\link[data.table:IDateTime]{data.table::as.IDate()}}.
}
\description{
Check for and Format Report and Reference Dates
}
\seealso{
Functions used for checking inputs
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
