% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dates}
\alias{check_dates}
\title{Check for and Format Report and Reference Dates}
\usage{
check_dates(obs)
}
\arguments{
\item{obs}{A \code{data.frame} containing \code{report_date} and \code{reference_date}
columns.}
}
\value{
The \code{obs} object, potentially modified, where \code{report_date} and
\code{reference_date} cast via \link{IDateTime} format.
}
\description{
Check for and Format Report and Reference Dates
}
\seealso{
Functions used for checking inputs
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
