% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_design_matrix_sparsity}
\alias{check_design_matrix_sparsity}
\title{Check design matrix sparsity}
\usage{
check_design_matrix_sparsity(
  matrix,
  sparsity_threshold = 0.9,
  min_matrix_size = 50,
  name = "checked"
)
}
\arguments{
\item{matrix}{A numeric matrix to be checked for sparsity.}

\item{sparsity_threshold}{A numeric value between 0 and 1 indicating the
threshold for considering a matrix sparse. Default is 0.9.}

\item{min_matrix_size}{An integer indicating the minimum size of the matrix
for which to perform the sparsity check. Default is 50.}

\item{name}{A character string specifying the name of the design
matrix. Default is "checked".}
}
\value{
This function is used for its side effect of providing an
informational message if the matrix is sparse. It returns NULL invisibly.
}
\description{
This function checks the sparsity of a design matrix and provides a
recommendation if the matrix is considered sparse.
}
\seealso{
Functions used for checking inputs
\code{\link{check_calendar_timestep}()},
\code{\link{check_group}()},
\code{\link{check_group_date_unique}()},
\code{\link{check_max_delay}()},
\code{\link{check_module}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_numeric_timestep}()},
\code{\link{check_observation_indicator}()},
\code{\link{check_quantiles}()},
\code{\link{check_timestep}()},
\code{\link{check_timestep_by_date}()},
\code{\link{check_timestep_by_group}()}
}
\concept{check}
