% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_group}
\alias{check_group}
\title{Check Observations for reserved grouping variables}
\usage{
check_group(obs)
}
\arguments{
\item{obs}{An observation data frame that does not contain \code{.group},
\code{.old_group}, or \code{.new_group} as these are reserved variables.}
}
\description{
Check Observations for reserved grouping variables
}
\seealso{
Functions used for checking inputs
\code{\link{check_dates}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
