% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_group_date_unique}
\alias{check_group_date_unique}
\title{Check Observations for Uniqueness of Grouping Variables with respect
to reference_date and report_date}
\usage{
check_group_date_unique(obs)
}
\arguments{
\item{obs}{An object that will be \code{coerce_dt}d in place, that contains
\code{.group}, \code{reference_date}, and \code{report_date} columns.}
}
\description{
This function checks that the input data is stratified by
reference_date, report_date, and .group. It does this by counting the
number of observations for each combination of these variables, and
throwing a warning if any combination has more than one observation.
}
\seealso{
Functions used for checking inputs
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
