% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_max_delay}
\alias{check_max_delay}
\title{Compare maximum delays specified by the user vs. observed in the data, and
raise potential warnings.}
\usage{
check_max_delay(latest_obs, max_delay)
}
\arguments{
\item{latest_obs}{The latest available observations.}

\item{max_delay}{Metadata for the maximum delay produced using
\code{\link[=enw_metadata_maxdelay]{enw_metadata_maxdelay()}}.}
}
\description{
Compare maximum delays specified by the user vs. observed in the data, and
raise potential warnings.
}
\seealso{
Functions used for checking inputs
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
