% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_modules_compatible}
\alias{check_modules_compatible}
\title{Check that model modules have compatible specifications}
\usage{
check_modules_compatible(modules)
}
\arguments{
\item{modules}{A list of model modules.}
}
\description{
Check that model modules have compatible specifications
}
\seealso{
Functions used for checking inputs
\code{\link{check_dates}()},
\code{\link{check_group}()},
\code{\link{check_module}()},
\code{\link{check_quantiles}()}
}
\concept{check}
