% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_observation_indicator}
\alias{check_observation_indicator}
\title{Check observation indicator}
\usage{
check_observation_indicator(new_confirm, observation_indicator = NULL)
}
\arguments{
\item{new_confirm}{A data frame containing the observations to be checked.}

\item{observation_indicator}{A character string specifying the column name
in \code{new_confirm} that represents the observation indicator. This column
should be of logical type. If NULL, no check is performed.}
}
\value{
This function is used for its side effect of checking the observation
indicator in \code{new_confirm}. If the check passes, the function returns
invisibly. Otherwise, it stops and returns an error message.
}
\description{
This function verifies if the \code{observation_indicator} within the provided
\code{new_confirm} observations is logical. The check is performed to ensure
that the \code{observation_indicator} is of the correct type.
}
\seealso{
Functions used for checking inputs
\code{\link{check_design_matrix_sparsity}()},
\code{\link{check_group}()},
\code{\link{check_group_date_unique}()},
\code{\link{check_max_delay}()},
\code{\link{check_module}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_numeric_timestep}()},
\code{\link{check_quantiles}()},
\code{\link{check_timestep}()},
\code{\link{check_timestep_by_date}()},
\code{\link{check_timestep_by_group}()}
}
\concept{check}
