% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_quantiles}
\alias{check_quantiles}
\title{Check Quantiles Required are Present}
\usage{
check_quantiles(posterior, req_probs = c(0.5, 0.95, 0.2, 0.8))
}
\arguments{
\item{posterior}{A dataframe containing quantiles identified using
the \code{q5} naming scheme. Default: No default.}

\item{req_probs}{A numeric vector of required probabilities. Default:
c(0.5, 0.95, 0.2, 0.8).}
}
\description{
Check Quantiles Required are Present
}
\seealso{
Functions used for checking inputs
\code{\link{check_dates}()},
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()}
}
\concept{check}
