% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_timestep}
\alias{check_timestep}
\title{Check Timestep}
\usage{
check_timestep(
  obs,
  date_var,
  timestep = "day",
  exact = TRUE,
  check_nrow = TRUE
)
}
\arguments{
\item{obs}{Any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}

\item{date_var}{The variable in \code{obs} representing dates.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week", "month") or a numeric whole number representing the number of days.}

\item{exact}{Logical, if \verb{TRUE``, checks if all differences exactly match the timestep. If }FALSE``, checks if the sum of the differences modulo the
timestep equals zero. Default is \code{TRUE}.}

\item{check_nrow}{Logical, if \code{TRUE}, checks if there are at least two
observations. Default is \code{TRUE}. If \code{FALSE}, the function returns invisibly
if there is only one observation.}
}
\value{
This function is used for its side effect of stopping if the check
fails. If the check passes, the function returns invisibly.
}
\description{
This function verifies if the difference in dates in the provided
observations corresponds to the provided timestep. If the \code{exact} argument
is set to TRUE, the function checks if all differences exactly match the
timestep; otherwise, it checks if the sum of the differences modulo the
timestep equals zero. If the check fails, the function stops and returns an
error message.
}
\seealso{
Functions used for checking inputs
\code{\link{check_calendar_timestep}()},
\code{\link{check_group_date_unique}()},
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_numeric_timestep}()},
\code{\link{check_quantiles}()},
\code{\link{check_timestep_by_date}()}
}
\concept{check}
