% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_timestep_by_date}
\alias{check_timestep_by_date}
\title{Check Timestep By Date}
\usage{
check_timestep_by_date(obs, timestep = "day", exact = TRUE)
}
\value{
This function is used for its side effect of checking the timestep
by date in \code{obs}. If the check passes for all dates, the function
returns invisibly. Otherwise, it stops and returns an error message.
}
\description{
This function verifies if the difference in dates within each date in the
provided observations corresponds to the provided timestep. This check is
performed for both \code{report_date} and \code{reference_date} and for each group in
\code{obs}.
}
\seealso{
Functions used for checking inputs
\code{\link{check_calendar_timestep}()},
\code{\link{check_group_date_unique}()},
\code{\link{check_group}()},
\code{\link{check_modules_compatible}()},
\code{\link{check_module}()},
\code{\link{check_numeric_timestep}()},
\code{\link{check_quantiles}()},
\code{\link{check_timestep}()}
}
\concept{check}
