% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_timestep_by_group}
\alias{check_timestep_by_group}
\title{Check Timestep By Group}
\usage{
check_timestep_by_group(obs, date_var, timestep = "day", exact = TRUE)
}
\arguments{
\item{obs}{Any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}

\item{date_var}{The variable in \code{obs} representing dates.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week", "month") or a numeric whole number representing the number of days.}

\item{exact}{Logical, if \verb{TRUE``, checks if all differences exactly match the timestep. If }FALSE``, checks if the sum of the differences modulo the
timestep equals zero. Default is \code{TRUE}.}
}
\value{
This function is used for its side effect of checking the timestep
for each group in \code{obs}. If the check passes for all groups, the function
returns invisibly. Otherwise, it stops and returns an error message.
}
\description{
This function verifies if the difference in dates within each group in the
provided observations corresponds to the provided timestep.
}
