% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_date}
\alias{coerce_date}
\title{Coerce Dates}
\usage{
coerce_date(dates)
}
\arguments{
\item{dates}{a vector-like input, which the function attempts
to coerce via \code{\link[data.table:IDateTime]{data.table::as.IDate()}}}
}
\value{
an \link{IDate} vector
}
\description{
Provides consistent coercion of inputs to \link{IDate}
with error handling
}
\details{
If any of the elements of \code{dates} cannot be coerced,
this function will result in an error, indicating all indices
which cannot be coerced to \link{IDate}.

Internal methods of \link{epinowcast} assume dates are represented
as \link{IDate}.
}
\examples{
# works
epinowcast::coerce_date(c("2020-05-28", "2020-05-29"))
# does not, indicates index 2 is problem
tryCatch(
 epinowcast::coerce_date(c("2020-05-28", "2020-o5-29")),
 error = function(e) print(e)
)
}
