% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{coerce_dt}
\alias{coerce_dt}
\title{Coerce \code{data.table}s}
\usage{
coerce_dt(data, required_cols, forbidden_cols, new = TRUE)
}
\arguments{
\item{data}{any of the types supported by \code{\link[data.table:as.data.table]{data.table::as.data.table()}}}

\item{required_cols}{character vector of required columns}

\item{forbidden_cols}{character vector of forbidden columns}

\item{new}{logical; if \code{TRUE} (default), a new \code{data.table} is returned}
}
\value{
a \code{data.table}; if \code{data} is a \code{data.table}, the returned object
will have a new address, unless \code{new = FALSE}.
i.e. be distinct from the original and not cause any side effects with
changes.
}
\description{
Provides consistent coercion of inputs to \link{data.table}
with error handling
}
\details{
TODO
}
\seealso{
Utility functions
\code{\link{coerce_date}()},
\code{\link{convert_cmdstan_to_rstan}()},
\code{\link{expose_stan_fns}()},
\code{\link{is.Date}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
