% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{construct_re}
\alias{construct_re}
\title{Constructs random effect terms}
\usage{
construct_re(re, data)
}
\arguments{
\item{re}{}

\item{data}{}
}
\value{
RETURN
}
\description{
Constructs random effect terms
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
