% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{construct_rw}
\alias{construct_rw}
\title{Constructs random walk terms}
\usage{
construct_rw(rw, data)
}
\arguments{
\item{rw}{A random walk term as defined by \code{\link[=rw]{rw()}}.}

\item{data}{A data.frame of observations used to define the
random walk term. Must contain the time and grouping variables
defined in the \code{\link[=rw]{rw()}} term specified.}
}
\value{
A list containing the following:
\itemize{
\item \code{data}: The input data.frame with the addition of the new variables
required by the specified random walk. These are added using
\code{\link[=enw_add_cumulative_membership]{enw_add_cumulative_membership()}}.
-\code{terms}: A character vector of new fixed effects terms to add to a model
formula.
\item \code{effects}: A data.frame describing the random effect structure of the new
effects.
}
}
\description{
This function takes random walks as defined
by \code{\link[=rw]{rw()}}, produces the required additional variables
(denoted using a "c" prefix and constructed using
\code{\link[=enw_add_cumulative_membership]{enw_add_cumulative_membership()}}), and then returns the
extended data.frame along with the new fixed effects and the
random effect structure.
}
\examples{
data <- enw_example("preproc")$metareference[[1]]

epinowcast:::construct_rw(rw(week), data)

epinowcast:::construct_rw(rw(week, day_of_week), data)
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
