% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{construct_rw}
\alias{construct_rw}
\title{Constructs random walk terms}
\usage{
construct_rw(rw, data)
}
\arguments{
\item{rw}{}

\item{data}{}
}
\value{
RETURN
}
\description{
Constructs random walk terms
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
