% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{.check_primarycensored}
\alias{.check_primarycensored}
\title{Error if the primarycensored package is not installed}
\usage{
.check_primarycensored(caller)
}
\arguments{
\item{caller}{A string naming the calling function for the error message.}
}
\value{
Invisibly \code{TRUE}; called for its side effect.
}
\description{
Error if the primarycensored package is not installed
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{.discretise_parametric_pmf}()},
\code{\link{build_ord_obs}()},
\code{\link{enw_add_latest_obs_to_nowcast}()},
\code{\link{enw_nowcast_samples}()},
\code{\link{enw_nowcast_summary}()},
\code{\link{enw_posterior}()},
\code{\link{enw_posterior_delay}()},
\code{\link{enw_pp_summary}()},
\code{\link{enw_quantiles_to_long}()},
\code{\link{enw_summarise_samples}()},
\code{\link{subset_obs}()}
}
\concept{postprocess}
