% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{.enw_prior_families}
\alias{.enw_prior_families}
\title{Prior families supported for each model prior distribution}
\usage{
.enw_prior_families(distribution)
}
\arguments{
\item{distribution}{A character string naming the prior family the Stan
model applies (see \code{\link[=.enw_prior_table]{.enw_prior_table()}}), or \code{NA}.}
}
\value{
A character vector of the \verb{<dist_spec>} distribution types (as
returned by \code{\link[distspec:get_distribution]{distspec::get_distribution()}}) that can be used to specify
the prior.
}
\description{
Prior families supported for each model prior distribution
}
\keyword{internal}
