% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{.validate_check_max_delay_args}
\alias{.validate_check_max_delay_args}
\title{Validate check_max_delay arguments}
\usage{
.validate_check_max_delay_args(
  max_delay,
  cum_coverage,
  maxdelay_quantile_outlier
)
}
\arguments{
\item{max_delay}{Maximum delay}

\item{cum_coverage}{Cumulative coverage threshold}

\item{maxdelay_quantile_outlier}{Quantile for outlier detection}
}
\value{
Integer max_delay value
}
\description{
Validate check_max_delay arguments
}
\keyword{internal}
