% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_add_latest_obs_to_nowcast}
\alias{enw_add_latest_obs_to_nowcast}
\title{FUNCTION_TITLE}
\usage{
enw_add_latest_obs_to_nowcast(nowcast, obs)
}
\arguments{
\item{nowcast}{PARAM_DESCRIPTION}

\item{obs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{enw_nowcast_summary}()},
\code{\link{enw_posterior}()},
\code{\link{enw_pp_summary}()}
}
\concept{postprocess}
