% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_add_max_reported}
\alias{enw_add_max_reported}
\title{Add maximum number of reported cases for each reference_date}
\usage{
enw_add_max_reported(obs)
}
\arguments{
\item{obs}{A data frame containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), and \code{confirm} (cumulative observations by reference and report
date).}
}
\value{
A data.table with new columns \code{max_confirm} and \code{cum_prop_reported}
}
\description{
This is a helper function which adds the maximum (in the sense
of latest observed) number of reported cases for each reference_date and
computes the proportion of already reported cases for each combination of
reference_date and report_date.
}
\examples{
Add `max_confirm` and `cum_prop_reported` to data
enw_add_max_reported(germany_covid19_hosp)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_incidence_to_cumulative}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_metadata_maxdelay}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
