% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_add_metaobs_features}
\alias{enw_add_metaobs_features}
\title{add common metadata variables}
\usage{
enw_add_metaobs_features(metaobs, holidays_to = "Sunday", holidays, ...)
}
\arguments{
\item{metaobs}{raw meta data for observations, coerceable via \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.
must have a \code{date} column}

\item{holidays_to}{A character string to assign to holidays when present.
Replaces the day of the week and defaults to Sunday.}

\item{holidays}{an optional vector of items coerceable via \code{\link[data.table:IDateTime]{data.table::as.IDate()}}}
}
\value{
a copy of the \code{metaobs} input, with additional columns:
\itemize{
\item \code{day_of_week}, a factor of values as output from \code{\link[=weekday]{weekday()}} and possibly as \code{holiday_to} if distinct from weekdays values
\item \code{day}, numeric, 0 based from start of time series
\item \code{week}, numeric, 0 based from start of time series
\item \code{month}, numeric, 0 based from start of time series
}
}
\description{
annotates time series data with commonly required data
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
