% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_add_metaobs_features}
\alias{enw_add_metaobs_features}
\title{FUNCTION_TITLE}
\usage{
enw_add_metaobs_features(metaobs, holidays = c(), holidays_to = "Sunday")
}
\arguments{
\item{metaobs}{PARAM_DESCRIPTION}

\item{holidays}{PARAM_DESCRIPTION}

\item{holidays_to}{A character string to assign to holidays when present.
Replaces the day of the week and defaults to Sunday.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Preprocessing functions
\code{\link{enw_assign_group}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_retrospective_data}()}
}
\concept{preprocess}
