% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_add_pooling_effect}
\alias{enw_add_pooling_effect}
\title{FUNCTION_TITLE}
\usage{
enw_add_pooling_effect(effects, string, var_name = "sd")
}
\arguments{
\item{effects}{PARAM_DESCRIPTION}

\item{string}{PARAM_DESCRIPTION}

\item{var_name}{PARAM_DESCRIPTION, Default: 'sd'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_cumulative_membership}()},
\code{\link{enw_dates_to_factors}()},
\code{\link{enw_design}()},
\code{\link{enw_effects_metadata}()},
\code{\link{enw_formula}()}
}
\concept{modeldesign}
