% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-converters.R
\name{enw_aggregate_cumulative}
\alias{enw_aggregate_cumulative}
\title{Aggregate observations over a given timestep for both report and reference
dates.}
\usage{
enw_aggregate_cumulative(
  obs,
  timestep = "day",
  by = NULL,
  min_date = min(obs$reference_date, na.rm = TRUE),
  copy = TRUE
)
}
\arguments{
\item{obs}{An object coercible to a \code{data.table} (such as a \code{data.frame})
which must have a \code{new_confirm} numeric column, and \code{report_date} and
\code{reference_date} date columns. The input must have a timestep of a day
and be complete. See \code{\link[=enw_complete_dates]{enw_complete_dates()}} for more information. If
NA values are present in the \code{confirm} column then these will be set to
zero before aggregation this may not be desirable if this missingness
is meaningful.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week", "month") or a numeric whole number representing the number of days.}

\item{by}{A character vector of variables to also aggregate by (i.e. as well
as using the \code{reference_date} and \code{report_date}). If not supplied
then the function will aggregate by just the \code{reference_date} and
\code{report_date}.}

\item{copy}{Should \code{obs} be copied (default) or modified in place?}

\item{min_reference_date}{The minimum reference date to start the
aggregation from. Note that the timestep will start from the minimum
reference date + a single time step (i.e. the first timestep will be
"2022-10-23" if the minimum reference date is "2022-10-16"). The default
is the minimum reference date in the \code{obs} object. Other sensible values
would be the minimum report date in the \code{obs} object + 1 day if reporting
is already weekly and you wish to ensure that the timestep of the output
matches the reporting timestep.}
}
\value{
A data.table with aggregated observations.
}
\description{
This function aggregates observations over a specified timestep,
ensuring alignment on the same day of week for report and reference dates.
It is  useful for aggregating data to a weekly timestep, for example which
may be desirable if testing using a weekly timestep or if you are very
concerned about runtime. Note that the start of the timestep will be
determined by \code{min_date} + a single timestep (i.e. the
first timestep will be "2022-10-23" if the minimum reference date is
"2022-10-16").
}
\examples{
nat_hosp <- germany_covid19_hosp[location == "DE"][age_group \%in\% "00+"]
enw_aggregate_cumulative(nat_hosp, timestep = "week")
}
\seealso{
Data converters
\code{\link{enw_add_cumulative}()},
\code{\link{enw_add_incidence}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_incidence_to_cumulative}()},
\code{\link{enw_incidence_to_linelist}()},
\code{\link{enw_linelist_to_incidence}()}
}
\concept{dataconverters}
