% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_as_data_list}
\alias{enw_as_data_list}
\title{Format data for use with stan}
\usage{
enw_as_data_list(
  pobs,
  reference_effects = enw_intercept_model(pobs$metareference[[1]]),
  report_effects = enw_intercept_model(pobs$metareport[[1]]),
  dist = "lognormal",
  nowcast = TRUE,
  pp = FALSE,
  likelihood = TRUE,
  debug = FALSE,
  output_loglik = FALSE
)
}
\arguments{
\item{pobs}{}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}
}
\value{
A list as required by stan.
}
\description{
Format data for use with stan
}
\seealso{
Functions used for modelling
\code{\link{enw_inits}()},
\code{\link{enw_model}()},
\code{\link{enw_sample}()}
}
\concept{model}
