% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_assign_group}
\alias{enw_assign_group}
\title{FUNCTION_TITLE}
\usage{
enw_assign_group(obs, by = c())
}
\arguments{
\item{obs}{PARAM_DESCRIPTION}

\item{by}{PARAM_DESCRIPTION, Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_retrospective_data}()}
}
\concept{preprocess}
