% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_cat_new_confirm}
\alias{enw_cat_new_confirm}
\title{Calculate categorised version of new_confirm}
\usage{
enw_cat_new_confirm(pobs, delay_group_thresh)
}
\arguments{
\item{pobs}{Pre-processed observations as resulting from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}}

\item{delay_group_thresh}{A vector defining thresholds for the grouping of
reporting delays based on left-closed intervals. Smallest value should be
zero, largest should correspond to max_delay.}
}
\value{
A \code{data.table} of delay notification incidence by reference date and
delay group. Empirical reporting distributions are also added.
}
\description{
Calculate categorised version of new_confirm for different categories of
reporting delays. These data are meant to be used in plotting.
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_emp_quant_by_reference}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
