% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_complete_dates}
\alias{enw_complete_dates}
\title{Complete missing reference and report dates}
\usage{
enw_complete_dates(
  obs,
  by = NULL,
  max_delay,
  min_date = min(obs$reference_date, na.rm = TRUE),
  max_date = max(obs$report_date, na.rm = TRUE),
  timestep = "day",
  missing_reference = TRUE,
  completion_beyond_max_report = FALSE,
  flag_observation = FALSE
)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\code{reference_date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{max_delay}{Numeric defaults to 20 and needs to be greater than or equal
to 1 and an integer (internally it will be coerced to one using
\code{\link[=as.integer]{as.integer()}}). The maximum number of days to include in the delay
distribution. Computation scales non-linearly with this setting so consider
what maximum makes sense for your data carefully. Note that this is zero
indexed and so includes the reference date and \code{max_delay - 1} other days
(i.e. a \code{max_delay} of 1 corresponds with no delay).}

\item{min_date}{The minimum date to include in the data. Defaults to the
minimum reference date found in the data.}

\item{max_date}{The maximum date to include in the data. Defaults to the
maximum report date found in the data.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week", "month") or a numeric whole number representing the number of days.}

\item{missing_reference}{Logical, should entries for cases with missing
reference date be completed as well?, Default: TRUE}

\item{completion_beyond_max_report}{Logical, should entries be completed
beyond the maximum date found in the data? Default: FALSE}

\item{flag_observation}{Logical, should observations that have been
imputed as missing be flagged as not observed?. Makes use of
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}} to add a \code{.observed} logical vector
which indicates if observations have been imputed. This vector can
then be passed to the \code{observation_indicator} argument of \code{\link[=enw_obs]{enw_obs()}} to
control if these observations are used in the likelihood. Default: FALSE}
}
\value{
A \code{data.table} with completed entries for all combinations of
reference dates, groups and possible report dates.
}
\description{
Ensures that all reference and report dates are present for
all groups based on the maximum and minimum dates found in the data.
This function may be of use to users when preprocessing their data. In
general all features that you may consider using as grouping variables
or as covariates need to be included in the \code{by} variable.
}
\examples{
obs <- data.frame(
  report_date = c("2021-10-01", "2021-10-03"), reference_date = "2021-10-01",
  confirm = 1
)
enw_complete_dates(obs)

# Allow completion beyond the maximum date found in the data
enw_complete_dates(obs, completion_beyond_max_report = TRUE, max_delay = 10)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_flag_observed_observations}()},
\code{\link{enw_impute_na_observations}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
