% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_complete_dates}
\alias{enw_complete_dates}
\title{Complete missing reference and report dates}
\usage{
enw_complete_dates(
  obs,
  by = NULL,
  max_delay,
  missing_reference = TRUE,
  completion_beyond_max_report = FALSE
)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\code{reference_date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{max_delay}{Numeric, defaults to 20 and needs to be greater than or
equal to 1 and an integer (internally it will be coerced to one using
\code{\link[=as.integer]{as.integer()}}). The maximum number of days to include in the delay
distribution. Observations with delays larger then the maximum delay will be
dropped. If the specified maximum delay is too short, nowcasts can be biased
as important parts of the true delay distribution are cut off. At the same
time, computational cost scales non-linearly with this setting, so you want
the maximum delay to be as long as necessary, but not much longer. Consider
what delays are realistic for your application, and when in doubt, check if
increasing the maximum delay noticeably changes the delay distribution or
nowcasts as estimated by epinowcast. If it does, your maximum delay may still
be too short. Note that delays are zero indexed and so include the reference
date and \code{max_delay - 1} other days (i.e. a \code{max_delay} of 1 corresponds to
no delay). See \code{\link[=enw_metadata_delay]{enw_metadata_delay()}} for checking the coverage of a delay
distribution for a given maximum delay.}

\item{missing_reference}{Logical, should entries for cases with missing
reference date be completed as well?, Default: TRUE}

\item{completion_beyond_max_report}{Logical, should entries be completed
beyond the maximum date found in the data? Default: FALSE}
}
\value{
A \code{data.table} with completed entries for all combinations of
reference dates, groups and possible report dates.
}
\description{
Ensures that all reference and report dates are present for
all groups based on the maximum and minimum dates found in the data.
This function may be of use to users when preprocessing their data. In
general all features that you may consider using as grouping variables
or as covariates need to be included in the \code{by} variable.
}
\examples{
obs <- data.frame(
  report_date = c("2021-10-01", "2021-10-03"), reference_date = "2021-10-01",
  confirm = 1
)
enw_complete_dates(obs)

# Allow completion beyond the maximum date found in the data
enw_complete_dates(obs, completion_beyond_max_report = TRUE, max_delay = 10)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
