% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_complete_dates}
\alias{enw_complete_dates}
\title{Complete missing reference and report dates}
\usage{
enw_complete_dates(obs, by = c(), max_delay, missing_reference = TRUE)
}
\arguments{
\item{obs}{A data frame containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{max_delay}{Numeric defaults to 20. The maximum number of days to
include in the delay distribution. Computation scales non-linearly with this
setting so consider what maximum makes sense for your data carefully. Note
that this is zero indexed and so includes the reference date and
\code{max_delay - 1} other days.}

\item{missing_reference}{Should entries for cases with missing reference date
be completed as well?, Default: TRUE}
}
\value{
A \code{data.table} with completed entries for all combinations of
reference dates, groups and possible report dates.
}
\description{
Ensures that all reference and report dates are present for
all groups based on the maximum and minimum dates found in the data.
This function may be of use to users when preprocessing their data. In
general all features that you may consider using as grouping variables
or as covariates need to be included in the \code{by} variable.
}
\examples{
obs <- data.frame(
  report_date = c("2021-10-01", "2021-10-03"), reference_date = "2021-10-01",
  confirm = 1
)
enw_complete_dates(obs)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_incidence_to_cumulative}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
