% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_construct_data}
\alias{enw_construct_data}
\title{Construct preprocessed data}
\usage{
enw_construct_data(
  obs,
  new_confirm,
  latest,
  missing_reference,
  reporting_triangle,
  metareport,
  metareference,
  metadelay,
  metamaxdelay,
  by
)
}
\arguments{
\item{obs}{Observations with the addition of empirical reporting proportions
and and restricted to the specified maximum delay.}

\item{new_confirm}{Incidence of notifications by reference and report date.
Empirical reporting distributions are also added.}

\item{latest}{The latest available observations.}

\item{missing_reference}{A \code{data.frame} of reported observations that are
missing the reference date.}

\item{reporting_triangle}{Incident observations by report and reference
date in the standard reporting triangle matrix format.}

\item{metareport}{Metadata for report dates.}

\item{metareference}{Metadata reference dates derived from observations.}

\item{metadelay}{Metadata for reporting delays produced using
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}}.}

\item{metamaxdelay}{Metadata for the maximum delay produced using
\code{\link[=enw_metadata_maxdelay]{enw_metadata_maxdelay()}}.}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}
}
\value{
A data.table containing processed observations as a series of nested
data.frames as well as variables containing metadata. These are:
\itemize{
\item \code{obs}: (observations with the addition of empirical reporting proportions
and and restricted to the specified maximum delay).
\item \code{new_confirm}: Incidence of notifications by reference and report date.
Empirical reporting distributions are also added.
\item \code{latest}: The latest available observations.
\item \code{missing_reference}: Observations missing reference dates.
\item \code{reporting_triangle}: Incident observations by report and reference date in
the standard reporting triangle matrix format.
\item \code{metareference}: Metadata reference dates derived from observations.
\item \code{metrareport}: Metadata for report dates.
\item \code{metadelay}: Metadata for reporting delays produced using
\code{\link[=enw_metadata_delay]{enw_metadata_delay()}}.
\item \code{metamaxdelay}: Metadata for reporting delays produced using
\code{\link[=enw_metadata_maxdelay]{enw_metadata_maxdelay()}}.
\item \code{time}: Numeric, number of timepoints in the data.
\item \code{snapshots}: Numeric, number of available data snapshots to use for
nowcasting.
\item \code{groups}: Numeric, Number of groups/strata in the supplied observations
(set using \code{by}).
\item \code{max_date}: The maximum available report date.
}
}
\description{
This function is used internally by \code{\link[=enw_preprocess_data]{enw_preprocess_data()}} to combine
various pieces of processed observed data into a single object. It
is exposed to the user in order to allow for modular data preprocessing
though this is not currently recommended. See documentation and code
of \code{\link[=enw_preprocess_data]{enw_preprocess_data()}} for more on the expected inputs.
}
\examples{
pobs <- enw_example("preprocessed")
enw_construct_data(
  obs = pobs$obs[[1]],
  new_confirm = pobs$new_confirm[[1]],
  latest = pobs$latest[[1]],
  missing_reference = pobs$missing_reference[[1]],
  reporting_triangle = pobs$reporting_triangle[[1]],
  metareport = pobs$metareport[[1]],
  metareference = pobs$metareference[[1]],
  metadelay = pobs$metadelay[[1]],
  metamaxdelay = pobs$metamaxdelay[[1]],
  by = c()
)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_metadata_maxdelay}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
