% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_dates_to_factors}
\alias{enw_dates_to_factors}
\title{FUNCTION_TITLE}
\usage{
enw_dates_to_factors(data)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_cumulative_membership}()},
\code{\link{enw_add_pooling_effect}()},
\code{\link{enw_design}()},
\code{\link{enw_effects_metadata}()},
\code{\link{enw_formula}()}
}
\concept{modeldesign}
