% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_dates_to_factors}
\alias{enw_dates_to_factors}
\title{Convert all \code{Date}s to \code{Factor}s}
\usage{
enw_dates_to_factors(data)
}
\arguments{
\item{data}{A \code{data.frame}.}
}
\value{
A \code{data.frame} with all \code{Date} variables converted to \code{Factor}s and
all other columns dropped.
}
\description{
Converts all \code{Date} columns to \code{Factor}s.
in a \code{data.frame}.
}
\examples{
data <- data.frame(date = as.Date("2019-01-01") + 0:2, x = 1:3)
enw_dates_to_factors(data)
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_cumulative_membership}()},
\code{\link{enw_add_pooling_effect}()},
\code{\link{enw_design}()},
\code{\link{enw_effects_metadata}()}
}
\concept{modeldesign}
