% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_delay_filter}
\alias{enw_delay_filter}
\title{Filter observations to restrict the maximum reporting delay}
\usage{
enw_delay_filter(obs, max_delay)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), and \code{confirm} (cumulative observations by reference and report
date).}

\item{max_delay}{Numeric, defaults to 20 and needs to be greater than or
equal to 1 and an integer (internally it will be coerced to one using
\code{\link[=as.integer]{as.integer()}}). The maximum number of days to include in the delay
distribution. Observations with delays larger then the maximum delay will be
dropped. If the specified maximum delay is too short, nowcasts can be biased
as important parts of the true delay distribution are cut off. At the same
time, computational cost scales non-linearly with this setting, so you want
the maximum delay to be as long as necessary, but not much longer. Consider
what delays are realistic for your application, and when in doubt, check if
increasing the maximum delay noticeably changes the delay distribution or
nowcasts as estimated by epinowcast. If it does, your maximum delay may still
be too short. Note that delays are zero indexed and so include the reference
date and \code{max_delay - 1} other days (i.e. a \code{max_delay} of 1 corresponds to
no delay). See \code{\link[=enw_metadata_delay]{enw_metadata_delay()}} for checking the coverage of a delay
distribution for a given maximum delay.}
}
\value{
A \code{data.frame} filtered so that dates by report are less than or
equal the reference date plus the maximum delay.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{enw_delay_filter()} was renamed to \code{enw_filter_delay()} for better consistency.
}
\examples{
obs <- enw_example("preprocessed")$obs[[1]]
enw_delay_filter(obs, max_delay = 2)
}
\keyword{internal}
