% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_delay_metadata}
\alias{enw_delay_metadata}
\title{Calculate reporting delay metadata}
\usage{
enw_delay_metadata(max_delay = 20, breaks = 4)
}
\arguments{
\item{max_delay}{Numeric defaults to 20. The maximum number of days to
include in the delay distribution. Computation scales non-linearly with this
setting so consider what maximum makes sense for your data carefully. Note
that this is zero indexed and so includes the reference date and
\code{max_delay - 1} other days.}

\item{breaks}{Numeric, defaults to 4. The number of breaks to use when
constructing a categorised version of numeric delays.}
}
\value{
A \code{data.frame} of delay metadata. This includes:
\itemize{
\item \code{delay}: The numeric delay from reference date to report.
\item \code{delay_cat}: The categorised delay. This may be useful for model building.
\item \code{delay_week}: The numeric week since the delay was reported. This again
may be useful for model building.
\item \code{delay_tail}: A logical variable defining if the delay is in the upper
75\% of the potential delays. This may be particularly useful when building
models that assume a parametric distribution in order to increase the weight
of the tail of the reporting distribution in a pragmatic way.
}
}
\description{
Calculate delay metadata based on the supplied maximum delay and independent
of other metadata or date indexing. These data are meant to be used in
conjunction with metadata on the date of reference. Users can build
additional features this \code{data.frame} or regenerate it using this function
in the output of \code{enw_preprocess_data()}.
}
\examples{
enw_delay_metadata(20, breaks = 4)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_incidence_to_cumulative}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
