% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_delay_quantiles}
\alias{enw_delay_quantiles}
\title{Empirical delay quantiles by reference date}
\usage{
enw_delay_quantiles(pobs, quantiles = c(0.1, 0.5, 0.9))
}
\arguments{
\item{pobs}{A preprocessed data object as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{quantiles}{A numeric vector of probabilities for which
quantiles are computed. Defaults to \code{c(0.1, 0.5, 0.9)}.}
}
\value{
A \code{data.table} with columns for each quantile by
reference date.
}
\description{
Computes empirical quantiles of the reporting
delay distribution for each reference date. Intended for
use with the \code{"delay_quantiles"} plot type in
\code{\link[=plot.enw_preprocess_data]{plot.enw_preprocess_data()}}.
}
\examples{
pobs <- enw_example("preprocessed_observations")
enw_delay_quantiles(pobs)
}
\seealso{
Plotting functions
\code{\link{enw_delay_categories}()},
\code{\link{enw_plot_delay_counts}()},
\code{\link{enw_plot_delay_cumulative}()},
\code{\link{enw_plot_delay_fraction}()},
\code{\link{enw_plot_delay_quantiles}()},
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
