% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_design}
\alias{enw_design}
\title{FUNCTION_TITLE}
\usage{
enw_design(formula, data, no_contrasts = FALSE, sparse = TRUE, ...)
}
\arguments{
\item{formula}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{no_contrasts}{PARAM_DESCRIPTION, Default: FALSE}

\item{sparse}{PARAM_DESCRIPTION, Default: TRUE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_cumulative_membership}()},
\code{\link{enw_add_pooling_effect}()},
\code{\link{enw_dates_to_factors}()},
\code{\link{enw_effects_metadata}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()}
}
\concept{modeldesign}
