% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_emp_quant_by_reference}
\alias{enw_emp_quant_by_reference}
\title{Calculate quantiles of empirical delay distribution by reference date}
\usage{
enw_emp_quant_by_reference(pobs, quantiles)
}
\arguments{
\item{pobs}{Pre-processed observations as resulting from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}}

\item{quantiles}{A vector of quantiles used for the "emp_rep_quant" plot.
Default is NULL, corresponding to the 0.10, 0.50, and 0.90 quantiles.}
}
\value{
A \code{data.table} of delay notification incidence by reference date and
delay group. Empirical reporting distributions are also added.
}
\description{
Calculate empirical quantiles of  delay distribution by reference date.
These data are meant to be used in plotting.
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_cat_new_confirm}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
