% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_example}
\alias{enw_example}
\title{Load a package example}
\usage{
enw_example(type = "nowcast")
}
\arguments{
\item{type}{A character string indicating the example to load.
Supported options are "nowcast, "preprocessed_observations", "observations",
and "script" which are the output of epinowcast()], \code{\link[=enw_preprocess_data]{enw_preprocess_data()}},
and \code{\link[=enw_latest_data]{enw_latest_data()}} applied to the \link{germany_covid19_hosp} package
dataset), and the script used to generate these examples respectively.}
}
\value{
A \code{data.table} of summarised output
}
\description{
Loads examples of nowcasts produce using example scripts. Used to streamline
examples, in package tests and to enable users to explore package
functionality without needing to install \code{cmdstanr}.
}
\examples{
# Load the nowcast
enw_example(type = "nowcast")

# Load the preprocessed observations
enw_example(type = "preprocessed_observations")

# Load the latest observations
enw_example(type = "observations")

# Load the script used to generate these examples
# Optionally source this script to regenerate the example
readLines(enw_example(type = "script"))
}
\seealso{
Package data sets
\code{\link{germany_covid19_hosp}}
}
\concept{data}
