% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_expectation}
\alias{enw_expectation}
\title{Expectation model module}
\usage{
enw_expectation(
  r = ~0 + (1 | day:.group),
  generation_time = 1,
  observation = ~1,
  latent_reporting_delay = 1,
  data,
  ...
)
}
\arguments{
\item{r}{A formula (as implemented in \code{\link[=enw_formula]{enw_formula()}}) describing
the generative process used for expected incidence. This can use features
defined by reference date as defined in \code{metareference} as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. By default this is set to use a daily random effect
by group. This parameterisation is highly flexible and may not be the
most appropriate choice when data is sparsely reported or reporting delays
are substantial. In these settings an alternative could be a group-specific
weekly random walk (specified as \code{rw(week, by = .group)}). Setting to
\code{~0} will produce an error as an expectation model is required. See
\code{\link[=enw_formula]{enw_formula()}} for details on formula syntax.}

\item{generation_time}{A numeric vector that sums to 1 and defaults to 1.
Describes the weighting to apply to previous generations (i.e as part of a
renewal equation), with the first entry the weight of the previous day.
When set to 1 (the default) this corresponds to modelling
the daily growth rate. Alternatively a bounded \verb{<dist_spec>} from the
\code{distspec} package (e.g. \code{distspec::Gamma(mean = 4, sd = 3, max = 15)}),
which is discretised to a daily probability mass function (PMF) using
\code{\link[distspec:discretise]{distspec::discretise()}}. As the renewal equation has no weight for the
current day, the probability of a delay of zero days is dropped and the
PMF renormalised.}

\item{observation}{A formula (as implemented in \code{\link[=enw_formula]{enw_formula()}}) describing
the modifiers used to adjust expected observations. This can use features
defined by reference date as defined in \code{metareference} as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. By default no modifiers are used but a common choice
might be to adjust for the day of the week. Note that as the baseline is
no modification, an intercept is always used and is set to 0. Set to \code{~0}
to disable observation modifiers (internally converted to \code{~1} and flagged
as inactive). See \code{\link[=enw_formula]{enw_formula()}} for details on formula syntax.}

\item{latent_reporting_delay}{A numeric vector that defaults to 1.
Describes the weighting to apply to past and current latent expected
observations (from most recent to least). This can be used both to convolve
based on some assumed reporting delay and to rescale observations (by
multiplying a probability mass function by some fraction) to account
ascertainment etc. A list of PMFs can be provided to allow for time-varying
PMFs. This should be the same length as the modelled time period plus the
length of the generation time if supplied. A bounded \verb{<dist_spec>} from
the \code{distspec} package (e.g. \code{distspec::LogNormal(mean = 5, sd = 2, max = 15)}) can be used in place of a PMF (including within a list) and
is discretised using \code{\link[distspec:discretise]{distspec::discretise()}} with the first entry the
probability of a delay of zero days.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{...}{Additional parameters passed to \code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}}. The
same arguments as passed to \code{enw_preprocess_data()} should be used here.}
}
\value{
A list containing the supplied formulas, data passed into a list
describing the models, a \code{data.table} describing the priors used (with a
\code{prior} column of \verb{<dist_spec>} objects from the \code{distspec} package, see
\code{\link[=enw_replace_priors]{enw_replace_priors()}}), and a function that takes the output data and
priors and returns a function that can be used to sample from a tightened
version of the prior distribution.
}
\description{
Expectation model module
}
\examples{
enw_expectation(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{.expectation_is_minimal}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
