% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_expectation}
\alias{enw_expectation}
\title{Expectation model module}
\usage{
enw_expectation(formula = ~rw(day, .group), order = 1, data)
}
\arguments{
\item{formula}{A formula (as implemented in \code{\link[=enw_formula]{enw_formula()}}) describing
the generative process used for expected incidence. This can use features
defined by reference date as defined in \code{metareference} as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. By default this is set to use a daily group
specific random walk. Note that the daily group specific random walk is
currently the only option supported by \code{\link[=epinowcast]{epinowcast()}}.}

\item{order}{Integer, defaults to 1. The order of the expectation process
with 1 being a simple log scale generative process, 2 being a log scale
generative process where each time point is offset by the log count from
previous time points.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list containing the supplied formulas, data passed into a list
describing the models, a \code{data.frame} describing the priors used, and a
function that takes the output data and priors and returns a function that
can be used to sample from a tightened version of the prior distribution.
}
\description{
Expectation model module
}
\examples{
enw_expectation(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
