% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_expectation}
\alias{enw_expectation}
\title{Expectation model module}
\usage{
enw_expectation(formula = ~rw(day, by = .group), generation_time = 1, data)
}
\arguments{
\item{formula}{A formula (as implemented in \code{\link[=enw_formula]{enw_formula()}}) describing
the generative process used for expected incidence. This can use features
defined by reference date as defined in \code{metareference} as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. By default this is set to use a daily group
specific random walk. Note that the daily group specific random walk is
currently the only option supported by \code{\link[=epinowcast]{epinowcast()}}.}

\item{generation_time}{A numeric vector that sums to 1 and defaults to 1.
Describes the weighting to apply to previous generations (i.e as part of a
renewal equation). When set to 1 (the default) this corresponds to modelling
the daily growth rate.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list containing the supplied formulas, data passed into a list
describing the models, a \code{data.frame} describing the priors used, and a
function that takes the output data and priors and returns a function that
can be used to sample from a tightened version of the prior distribution.
}
\description{
Expectation model module
}
\examples{
enw_expectation(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
