% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_extend_date}
\alias{enw_extend_date}
\title{Extend a time series with additional dates}
\usage{
enw_extend_date(metaobs, days = 20, direction = "end")
}
\arguments{
\item{metaobs}{A data.table with a \code{date} column.}

\item{days}{PARAM_DESCRIPTION}

\item{direction}{Should new dates be added at the beginning or end of
the data. Default is "end" with "start" also available}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Extend a time series with additional dates. This is useful
when extending the report dates of a time series to include future dates
for nowcasting purposes or to include additional dates for backcasting
when using a renewal process as the expectation model.
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_incidence_to_cumulative}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
