% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_extend_date}
\alias{enw_extend_date}
\title{Extend a time series with additional dates}
\usage{
enw_extend_date(
  metaobs,
  days = 20,
  direction = c("end", "start"),
  timestep = "day"
)
}
\arguments{
\item{metaobs}{A \code{data.frame} with a \code{date} column.}

\item{days}{Number of days to add to the time series. Defaults to 20.}

\item{direction}{Should new dates be added at the beginning or end of
the data. Default is "end" with "start" also available.}

\item{timestep}{The timestep to used. This can be a string ("day",
"week") or a numeric whole number representing the number of days.
Note that "month" is not currently supported in user-facing functions
and will throw an error if used.}
}
\value{
A data.table with the same columns as \code{metaobs} but with
additional rows for each date in the range of \code{date} to \code{date + days}
(or \code{date - days} if \code{direction = "start"}). An additional variable
observed is added with a value of FALSE for all new dates and TRUE
for all existing dates.
}
\description{
Extend a time series with additional dates. This is useful
when extending the report dates of a time series to include future dates
for nowcasting purposes or to include additional dates for backcasting
when using a renewal process as the expectation model.
}
\examples{
metaobs <- data.frame(date = as.Date("2021-01-01") + 0:4)
enw_extend_date(metaobs, days = 2)
enw_extend_date(metaobs, days = 2, direction = "start")
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_filter_report_dates_by_min}()},
\code{\link{enw_flag_observed_observations}()},
\code{\link{enw_impute_na_observations}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_reporting_triangle_to_long}()}
}
\concept{preprocess}
