% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_filter_reference_dates}
\alias{enw_filter_reference_dates}
\title{Filter by reference dates}
\usage{
enw_filter_reference_dates(
  obs,
  earliest_date,
  include_days,
  latest_date,
  remove_days
)
}
\arguments{
\item{obs}{A \code{data.frame}; must have \code{report_date} and \code{reference_date}
columns.}

\item{earliest_date}{earliest reference date to include in the data set}

\item{include_days}{if \code{earliest_date} is not given, the number
of reference dates to include, ending with the latest reference
date included (determined by \code{latest_date} or \code{remove_days}).}

\item{latest_date}{Date, the latest reference date to include in the
returned dataset.}

\item{remove_days}{Integer, if \code{latest_date} is not given, the number
of reference dates to remove, starting from the latest date included.}
}
\value{
A \code{data.table} filtered by report date
}
\description{
This is a helper function which allows users to filter datasets
by reference date. This is useful, for example, when evaluating nowcast
performance against fully observed data. Users may wish to combine this
function with \code{\link[=enw_filter_report_dates]{enw_filter_report_dates()}}. Note that by definition it is
assumed that report dates must be equal or greater than the corresponding
reference date (i.e a report cannot happen before the event being reported
occurs). This means that this function will also filter out any report dates
that are earlier than their corresponding reference date.
}
\examples{
# Filter by date
enw_filter_reference_dates(
  germany_covid19_hosp,
  earliest_date = "2021-09-01",
  latest_date = "2021-10-01"
)
#
# Filter by days
enw_filter_reference_dates(
  germany_covid19_hosp,
  include_days = 10, remove_days = 10
)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_flag_observed_observations}()},
\code{\link{enw_impute_na_observations}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
