% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_filter_report_dates}
\alias{enw_filter_report_dates}
\title{Filter by report dates}
\usage{
enw_filter_report_dates(obs, latest_date, remove_days)
}
\arguments{
\item{obs}{An observation data frame containing \code{report_date} and
\code{reference_date} columns.}

\item{latest_date}{Date, the latest report date to include in the
returned dataset.}

\item{remove_days}{Integer, if \code{latest_date} is not given, the number
of report dates to remove, starting from the latest date included.}
}
\value{
A data.table  filtered by report date
}
\description{
This is a helper function which allows users to create
truncated data sets at past time points from a given larger data set.
This is useful when evaluating nowcast performance against fully
observed data. Users may wish to combine this function with
\code{\link[=enw_filter_reference_dates]{enw_filter_reference_dates()}}.
}
\examples{
# Filter by date
enw_filter_report_dates(germany_covid19_hosp, latest_date = "2021-09-01")
#
# Filter by days
enw_filter_report_dates(germany_covid19_hosp, remove_days = 10)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_incidence_to_cumulative}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
