% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_fit_opts}
\alias{enw_fit_opts}
\title{Format model fitting options for use with stan}
\usage{
enw_fit_opts(
  sampler = epinowcast::enw_sample,
  nowcast = TRUE,
  pp = FALSE,
  likelihood = TRUE,
  likelihood_aggregation = c("snapshots", "groups"),
  threads_per_chain = 1L,
  debug = FALSE,
  output_loglik = FALSE,
  sparse_design = FALSE,
  use_batched_likelihood = FALSE,
  ...
)
}
\arguments{
\item{sampler}{A function that creates an object that be used to extract
posterior samples from the specified model. By default this is \code{\link[=enw_sample]{enw_sample()}}
which makes use of \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{nowcast}{Logical, defaults to \code{TRUE}. Should a nowcast be made using
posterior predictions of the unobserved future reported notifications.}

\item{pp}{Logical, defaults to \code{FALSE}. Should posterior predictions be made
for observed data. Useful for evaluating the performance of the model.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{likelihood_aggregation}{Character string, aggregation over which
stratify the likelihood when \code{threads_per_chain} is greater than 1; enforced
by \code{\link[base:match.arg]{base::match.arg()}}. Currently supported options:
\itemize{
\item "snapshots" which aggregates over report dates and groups (i.e the lowest
level that observations are reported at),
\item "groups" which aggregates across user defined groups.
}

Note that some model modules override this setting depending on model
requirements. For example, the \code{\link[=enw_missing]{enw_missing()}} module model forces
"groups" option. Generally, Users should typically want the default
"snapshots" aggregation.}

\item{threads_per_chain}{Integer, defaults to \code{1}. The number of threads to
use within each MCMC chain. If this is greater than \code{1} then components of
the likelihood will be calculated in parallel within each chain.}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}

\item{sparse_design}{Logical, defaults to \code{FALSE}. Should a sparse design
matrices be used for all design matrices. This reduces memory requirements
and may reduce computation time when fitting models with very sparse
design matrices (90\% or more zeros).}

\item{use_batched_likelihood}{Logical, defaults to \code{FALSE}. Should the
per-snapshot observation likelihood be computed with a single batched
pass over all snapshots in a work item, rather than one
\code{inv_logit()}/\code{log1m()}/\code{cumulative_sum()} call per snapshot. This is a
pure Stan, mathematically identical reformulation (see
\code{expected_obs_from_snaps_batched()} in \code{inst/stan/functions}) that
reduces per-snapshot call overhead; it does not change the model or the
likelihood aggregation scheme. It speeds up models fitted without
within-chain threading (\code{threads_per_chain = 1}) and gives little or no
gain with threading.}

\item{...}{Additional arguments to pass to the fitting function being used
by \code{\link[=epinowcast]{epinowcast()}}. By default this will be \code{\link[=enw_sample]{enw_sample()}} and so \code{cmdstanr}
options should be used.}
}
\value{
A list containing the specified sampler function, data as a list
specifying the fitting options to use, and additional arguments to pass
to the sampler function when it is called.
}
\description{
Format model fitting options for use with stan
}
\examples{
# Default options along with settings to pass to enw_sample
enw_fit_opts(iter_sampling = 1000, iter_warmup = 1000)
}
\seealso{
Model modules
\code{\link{.expectation_is_minimal}()},
\code{\link{enw_expectation}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
