% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_fit_opts}
\alias{enw_fit_opts}
\title{Format model fitting options for use with stan}
\usage{
enw_fit_opts(
  sampler = epinowcast::enw_sample,
  nowcast = TRUE,
  pp = FALSE,
  likelihood = TRUE,
  debug = FALSE,
  output_loglik = FALSE,
  ...
)
}
\arguments{
\item{sampler}{A function that creates an object that be used to extract
posterior samples from the specfied model. By default this is \code{\link[=enw_sample]{enw_sample()}}
which makes use of \code{\link[cmdstanr:model-method-sample]{cmdstanr::sample()}}.}

\item{nowcast}{Logical, defaults to \code{TRUE}. Should a nowcast be made using
posterior predictions of the unobserved future reported notifications.}

\item{pp}{Logical, defaults to \code{FALSE}. Should posterior predictions be made
for observed data. Useful for evaluating the performance of the model.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}

\item{...}{Additional arguments to pass to the fitting function being used
by \code{\link[=epinowcast]{epinowcast()}}. By default this will be \code{\link[=enw_sample]{enw_sample()}} and so \code{cmdstanr}
options should be used.}
}
\value{
A list as required by stan.
}
\description{
Format model fitting options for use with stan
}
\examples{
# Default options along with settings to pass to enw_sample
enw_fit_opts(iter_sampling = 1000, iter_warmup = 1000)
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_posterior_as_prior}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{remove_profiling}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
