% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_forecast}
\alias{enw_forecast}
\title{Forcasting model module}
\usage{
enw_forecast(
  formula = ~rw(day, by = .group),
  N_days = 0,
  continue_dynamics = FALSE,
  data = data
)
}
\arguments{
\item{N_days}{Number of days to forecast}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{trend_estimation_method}{Which trend to use for forecasting. A value of N uses an average over the last N days of trend estimates, while a value of "dynamic" propagates the trend using the specified expectation model, for example continuing the random walk.}
}
\value{
A list containing the supplied formulas, data passed into a list
describing the models, a \code{data.frame} describing the priors used, and a
function that takes the output data and priors and returns a function that
can be used to sample from a tightened version of the prior distribution.
}
\description{
Forcasting model module
}
\examples{
enw_forcast(N_days=14)
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
