% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{enw_formula}
\alias{enw_formula}
\title{FUNCTION_TITLE}
\usage{
enw_formula(formula, data)
}
\arguments{
\item{formula}{DESCRIPTION}

\item{data}{DESCRIPTION}
}
\value{
A design matrix and metadata
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_cumulative_membership}()},
\code{\link{enw_add_pooling_effect}()},
\code{\link{enw_dates_to_factors}()},
\code{\link{enw_design}()},
\code{\link{enw_effects_metadata}()},
\code{\link{enw_manual_formula}()}
}
\concept{modeldesign}
