% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{enw_formula}
\alias{enw_formula}
\title{FUNCTION_TITLE}
\usage{
enw_formula(metaobs, fixed = c(), random = c(), custom_random = c())
}
\arguments{
\item{metaobs}{PARAM_DESCRIPTION}

\item{fixed}{PARAM_DESCRIPTION. Default: c()}

\item{random}{PARAM_DESCRIPTION. Default: c()}

\item{custom_random}{PARAM_DESCRIPTION. Default: c()}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to formulate models
\code{\link{enw_add_cumulative_membership}()},
\code{\link{enw_add_pooling_effect}()},
\code{\link{enw_dates_to_factors}()},
\code{\link{enw_design}()},
\code{\link{enw_effects_metadata}()}
}
\concept{modeldesign}
