% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{enw_formula}
\alias{enw_formula}
\title{Define a model using a formula interface}
\usage{
enw_formula(formula, data, sparse = TRUE)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link{lme4} syntax (see \code{\link[=re]{re()}}), and random walks
defined using the \code{\link[=rw]{rw()}} helper function.}

\item{data}{A \code{data.frame} of observations. It must include all
variables used in the supplied formula.}

\item{sparse}{Logical, defaults to  \code{TRUE}. Should the fixed effects design
matrix be sparely defined.}
}
\value{
A list containing the following:
\itemize{
\item \code{formula}: The user supplied formula
\item \code{parsed_formula}: The formula as parsed by \code{\link[=parse_formula]{parse_formula()}}
\item \code{extended_formula}: The flattened version of the formula with
both user supplied terms and terms added for the user supplied
complex model components.
\item \code{fixed}:  A list containing the fixed effect formula, sparse design
matrix, and the index linking the design matrix with observations.
\item \code{random}: A list containing the random effect formula, sparse design
matrix, and the index linking the design matrix with random effects.
}
}
\description{
This function allows models to be defined using a
flexible formula interface that supports fixed effects, random effects
(using \link{lme4} syntax). Note that the returned fixed effects design matrix is
sparse and so the index supplied is required to link observations to the
appropriate design matrix row.
}
\examples{
# Use meta data for references dates from the Germany COVID-19
# hospitalisation data.
obs <- enw_filter_report_dates(
  germany_covid19_hosp[location == "DE"],
  remove_days = 40
)
obs <- enw_filter_reference_dates(obs, include_days = 40)
pobs <- enw_preprocess_data(obs, by = c("age_group", "location"))
data <- pobs$metareference[[1]]

# Model with fixed effects for age group
enw_formula(~ 1 + age_group, data)

# Model with random effects for age group
enw_formula(~ 1 + (1 | age_group), data)

# Model with a random effect for age group and a random walk
enw_formula(~ 1 + (1 | age_group) + rw(week), data)

# Model defined without a sparse fixed effects design matrix
enw_formula(~1, data[1:20, ])
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{enw_manual_formula}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
