% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_formula_as_data_list}
\alias{enw_formula_as_data_list}
\title{Format formula data for use with stan}
\usage{
enw_formula_as_data_list(data, reference_effects, report_effects)
}
\arguments{
\item{data}{A list of stan observation data as produced by
\code{\link[=enw_obs_as_data_list]{enw_obs_as_data_list()}}.}

\item{reference_effects}{A list of fixed and random design matrices
defining the date of reference model. Defaults to \code{\link[=enw_formula]{enw_formula()}}
which is an intercept only model.}

\item{report_effects}{A list of fixed and random design matrices
defining the date of reports model. Defaults to \code{\link[=enw_formula]{enw_formula()}}
which is an intercept only model.}
}
\value{
A list as required by stan.
}
\description{
Format formula data for use with stan
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_opts_as_data_list}()},
\code{\link{enw_posterior_as_prior}()},
\code{\link{enw_priors_as_data_list}()}
}
\concept{modeltools}
