% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_get_cache}
\alias{enw_get_cache}
\title{Retrieve Stan cache location}
\usage{
enw_get_cache()
}
\value{
The string of the filepath

something
}
\description{
Retrieves the user set cache location for Stan models. This
path can be set through the \code{enw_cache_location} function call.
}
\seealso{
Utility functions
\code{\link{aggregate_rolling_sum}()},
\code{\link{coerce_date}()},
\code{\link{coerce_dt}()},
\code{\link{date_to_numeric_modulus}()},
\code{\link{enw_unset_cache}()},
\code{\link{get_internal_timestep}()},
\code{\link{is.Date}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
