% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_get_cache}
\alias{enw_get_cache}
\title{Retrieve Stan cache location}
\usage{
enw_get_cache()
}
\value{
A string representing the file path for the cache location
}
\description{
Retrieves the user set cache location for Stan models. This
path can be set through the \code{enw_cache_location} function call.
If no environmental variable is available the output from
\code{\link[=tempdir]{tempdir()}} will be returned.
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_model}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_stan_to_r}()},
\code{\link{enw_unset_cache}()},
\code{\link{remove_profiling}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
