% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_incidence_to_cumulative}
\alias{enw_incidence_to_cumulative}
\title{Calculate cumulative reported cases from incidence of new reports}
\usage{
enw_incidence_to_cumulative(obs, by = c())
}
\arguments{
\item{obs}{A data frame containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), and \code{new_confirm} (incident observations by reference and
report date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}
}
\value{
The input data frame with a new variable \code{confirm}.
}
\description{
Calculate cumulative reported cases from incidence of new reports
}
\examples{
# Default reconstruct incidence
dt <- germany_covid19_hosp[location == "DE"][age_group == "00+"]
enw_cumulative_to_incidence(dt)

# Make use of maximum reported to calculate empirical daily reporting
dt <- enw_assign_group(dt)
dt <- enw_add_max_reported(dt)
enw_cumulative_to_incidence(dt)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_delay_filter}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
