% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-converters.R
\name{enw_incidence_to_linelist}
\alias{enw_incidence_to_linelist}
\title{FUNCTION_TITLE}
\usage{
enw_incidence_to_linelist(obs, by = NULL)
}
\arguments{
\item{obs}{DESCRIPTION.}

\item{by}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
