% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_inits}
\alias{enw_inits}
\title{Set up initial conditions for model}
\usage{
enw_inits(data)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=enw_as_data_list]{enw_as_data_list()}}.}
}
\value{
A function that when called returns a list of initial conditions
for the package stan models.
}
\description{
Set up initial conditions for model
}
\seealso{
Functions used for modelling
\code{\link{enw_as_data_list}()},
\code{\link{enw_model}()},
\code{\link{enw_sample}()}
}
\concept{model}
