% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_latest_data}
\alias{enw_latest_data}
\title{Filter observations to the latest available}
\usage{
enw_latest_data(obs, ref_window)
}
\arguments{
\item{obs}{A data frame containing at least the following variables:
\verb{reference date} (index date of interest), and \code{report_date} (report date for
observations).}

\item{ref_window}{Numeric vector of length 2. Can optionally be used to
restrict the reference date to a window referenced to the maximum available
reference date.}
}
\value{
A filtered data frame with the same variables as \code{obs}.
}
\description{
Filter observations to the latest available
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_obs}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_retrospective_data}()}
}
\concept{preprocess}
