% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_latest_data}
\alias{enw_latest_data}
\title{FUNCTION_TITLE}
\usage{
enw_latest_data(obs)
}
\arguments{
\item{obs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Preprocessing functions
\code{\link{enw_assign_group}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_metadata}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
