% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_linelist_to_incidence}
\alias{enw_linelist_to_incidence}
\title{Convert linelist data into a format usable by epinowcast}
\usage{
enw_linelist_to_incidence(
  obs,
  reference_col,
  report_col,
  by = c(),
  remove_reference_date_NA = TRUE
)
}
\arguments{
\item{obs}{}

\item{reference_col}{}

\item{by}{}
}
\value{
stuff
}
\description{
Convert linelist data into a format usable by epinowcast
}
\examples{
# Transform fake linelist into cumulative
data("linelist", package = "epinowcast")

enw_linelist_to_incidence(obs = linelist, 
reference_col = "date_onset", 
by = "age_group")

}
