% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_linelist_to_incidence}
\alias{enw_linelist_to_incidence}
\title{Convert linelist data into a format usable by epinowcast}
\usage{
enw_linelist_to_incidence(obs, reference_col, by = c())
}
\arguments{
\item{obs}{}

\item{reference_col}{}

\item{by}{}
}
\value{
stuff
}
\description{
Convert linelist data into a format usable by epinowcast
}
\examples{
# Will sort this when the function works properly
}
