% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-converters.R
\name{enw_linelist_to_incidence}
\alias{enw_linelist_to_incidence}
\title{FUNCTION_TITLE}
\usage{
enw_linelist_to_incidence(linelist, by = c(), max_delay)
}
\arguments{
\item{linelist}{DESCRIPTION.}

\item{by}{DESCRIPTION.}

\item{max_delay}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
linelist <- data.frame(
  reference_date = as.Date(c("2021-01-02", "2021-01-03", "2021-01-02")),
  report_date = as.Date(c("2021-01-03", "2021-01-05", "2021-01-04"))
)
enw_linelist_to_incidence(linelist)
}
\seealso{
Data converters
\code{\link{enw_add_cumulative}()},
\code{\link{enw_add_incidence}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_incidence_to_cumulative}()}
}
\concept{dataconverters}
