% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{enw_manual_formula}
\alias{enw_manual_formula}
\title{Define a model manually using fixed and random effects}
\usage{
enw_manual_formula(
  data,
  fixed = c(),
  random = c(),
  custom_random = c(),
  no_contrasts = FALSE,
  add_intercept = TRUE
)
}
\arguments{
\item{data}{A \code{data.frame} of observations. It must include all
variables used in the supplied formula.}

\item{fixed}{A character vector of fixed effects.}

\item{random}{A character vector of random effects. Random effects
specified here will be added to the fixed effects.}

\item{custom_random}{A vector of random effects. Random effects added here
will not be added to the vector of fixed effects. This can be used to random
effects for fixed effects that only have a partial name match.}

\item{no_contrasts}{Logical, defaults to \code{FALSE}. \code{TRUE} means that no
variable uses contrast. Alternatively a character vector of variables can be
supplied indicating which variables should  not have contrasts.}

\item{add_intercept}{Logical, defaults to \code{FALSE}. Should an intercept be
added to the fixed effects.}
}
\value{
A list specifying the fixed effects (formula, design matrix, and
design matrix index), and random effects (formula and design matrix).
}
\description{
For most typical use cases \code{\link[=enw_formula]{enw_formula()}} should
provide sufficient flexibility to allow models to be defined. However,
there may be some instances where more manual model specification is
required. This function supports this by allowing the user to supply
vectors of fixed, random, and customised random effects (where they are
not first treated as fixed effect terms). Prior to \verb{1.0.0} this was the
main interface for specifying models and it is still used internally to
handle some parts of the model specification process.
}
\examples{
data <- enw_example("prep")$metareference[[1]]
enw_manual_formula(data, fixed = "week", random = "day_of_week")
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{enw_formula}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
