% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_metadata}
\alias{enw_metadata}
\title{FUNCTION_TITLE}
\usage{
enw_metadata(obs, target_date = "reference_date")
}
\arguments{
\item{obs}{PARAM_DESCRIPTION}

\item{target_date}{PARAM_DESCRIPTION, Default: 'reference_date'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Preprocessing functions
\code{\link{enw_assign_group}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_retrospective_data}()}
}
\concept{preprocess}
