% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_metadata_delay}
\alias{enw_metadata_delay}
\title{Calculate reporting delay metadata for a given maximum delay}
\usage{
enw_metadata_delay(max_delay = 20, breaks = 4)
}
\arguments{
\item{max_delay}{Numeric, defaults to 20 and needs to be greater than or
equal to 1 and an integer (internally it will be coerced to one using
\code{\link[=as.integer]{as.integer()}}). The maximum number of days to include in the delay
distribution. Observations with delays larger then the maximum delay will be
dropped. If the specified maximum delay is too short, nowcasts can be biased
as important parts of the true delay distribution are cut off. At the same
time, computational cost scales non-linearly with this setting, so you want
the maximum delay to be as long as necessary, but not much longer. Consider
what delays are realistic for your application, and when in doubt, check if
increasing the maximum delay noticeably changes the delay distribution or
nowcasts as estimated by epinowcast. If it does, your maximum delay may still
be too short. Note that delays are zero indexed and so include the reference
date and \code{max_delay - 1} other days (i.e. a \code{max_delay} of 1 corresponds to
no delay). See \code{\link[=enw_metadata_delay]{enw_metadata_delay()}} for checking the coverage of a delay
distribution for a given maximum delay.}

\item{breaks}{Numeric, defaults to 4. The number of breaks to use when
constructing a categorised version of numeric delays.}
}
\value{
A  \code{data.frame}  of delay metadata. This includes:
\itemize{
\item \code{delay}: The numeric delay from reference date to report.
\item \code{delay_cat}: The categorised delay. This may be useful for model building.
\item \code{delay_week}: The numeric week since the delay was reported. This again
may be useful for model building.
\item \code{delay_tail}: A logical variable defining if the delay is in the upper
75\% of the potential delays. This may be particularly useful when building
models that assume a parametric distribution in order to increase the weight
of the tail of the reporting distribution in a pragmatic way.
}
}
\description{
Calculate delay metadata based on the supplied maximum delay and independent
of other metadata or date indexing. These data are meant to be used in
conjunction with metadata on the date of reference. Users can build
additional features with this  \code{data.frame}  or regenerate it using this
function in the output of \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.
}
\examples{
enw_metadata_delay(max_delay = 20, breaks = 4)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
