% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_metadata_maxdelay}
\alias{enw_metadata_maxdelay}
\title{Calculate metadata for the maximum reporting delay}
\usage{
enw_metadata_maxdelay(obs, max_delay = 20)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{max_delay}{Numeric, defaults to 20. The maximum number of days to
include in the delay distribution. Observations with a delay larger than
\code{max_delay} will be dropped during preprocessing. The time required for model
fitting increases non-linearly with this setting, so carefully choose a
maximum that makes sense for your data. Note that delays use zero-based
indexing, i.e. \code{max_delay} includes the reference date and the
\code{max_delay - 1} following days.}
}
\value{
A \code{list} of maximum delays. These include:
\itemize{
\item \code{spec}: The maximum delay specified by the user.
\item \code{obs}: The maximum delay observed in the data.
\item \code{model}: The maximum delay used by the model.
}
}
\description{
The maximum reporting delay is used to make the modeling of
reporting delays tractable by right-truncating the delay distribution at a
reasonable number of days. The maximum delay is specified by the user, and
can be smaller or larger than the maximum delay observed in the data.
Importantly, epinowcast will currently always use the smaller one of the two
maximum delays for modeling. This means that observations with a delay
larger than the specified maximum delay will be dropped from the analysis.
In some settings, for example outbreaks where little data is available
this may not be ideal. If this is an issue for you, please get in touch
with the developers by opening an issue on GitHub.
}
\examples{
obs <- enw_example(type = "preprocessed_observations")$obs[[1]]
enw_metadata_maxdelay(obs, max_delay = 20)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_cumulative_to_incidence}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_incidence_to_cumulative}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
