% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_missing}
\alias{enw_missing}
\title{Missing reference data model module}
\usage{
enw_missing(formula = ~1, data)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link{lme4} syntax (see \code{\link[=re]{re()}}), and random walks
defined using the \code{\link[=rw]{rw()}} helper function.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Missing reference data model module
}
\examples{
enw_missing(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
