% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_missing}
\alias{enw_missing}
\title{Missing reference data model module}
\usage{
enw_missing(formula = ~1, data)
}
\arguments{
\item{formula}{A formula (as implemented in \code{\link[=enw_formula]{enw_formula()}}) describing
the missing data proportion on the logit scale by reference date. This can
use features defined by reference date as defined in \code{metareference} as
produced by \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. "~0" implies no model is required.
Otherwise an intercept is always needed}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list containing the supplied formulas, data passed into a list
describing the models, a \code{data.frame} describing the priors used, and a
function that takes the output data and priors and returns a function that
can be used to sample from a tightened version of the prior distribution.
}
\description{
Missing reference data model module
}
\examples{
# Missing model with a fixed intercept only
enw_missing(data = enw_example("preprocessed"))

# No missingness model specified
enw_missing(~0, data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
