% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_new_reports}
\alias{enw_new_reports}
\title{FUNCTION_TITLE}
\usage{
enw_new_reports(obs)
}
\arguments{
\item{obs}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_obs}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_retrospective_data}()}
}
\concept{preprocess}
