% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_nowcast_samples}
\alias{enw_nowcast_samples}
\title{FUNCTION_TITLE}
\usage{
enw_nowcast_samples(fit, obs)
}
\arguments{
\item{fit}{A \code{cmdstanr} fit object.}

\item{obs}{An observation data frame containing \code{reference_date}
columns of the same length as the number of rows in the posterior and the
most up to date observation for each date. This is used to align the
posterior with the observations. The easiest source of this data is the
output of latest output of \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{enw_add_latest_obs_to_nowcast}()},
\code{\link{enw_nowcast_summary}()},
\code{\link{enw_posterior}()},
\code{\link{enw_pp_summary}()},
\code{\link{enw_quantiles_to_long}()},
\code{\link{enw_summarise_samples}()}
}
\concept{postprocess}
