% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_nowcast_summary}
\alias{enw_nowcast_summary}
\title{FUNCTION_TITLE}
\usage{
enw_nowcast_summary(fit, obs, probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95))
}
\arguments{
\item{fit}{PARAM_DESCRIPTION}

\item{obs}{PARAM_DESCRIPTION}

\item{probs}{PARAM_DESCRIPTION, Default: c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8,
0.95)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{enw_add_latest_obs_to_nowcast}()},
\code{\link{enw_posterior}()}
}
\concept{postprocess}
