% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_obs}
\alias{enw_obs}
\title{Setup observation model and data}
\usage{
enw_obs(family = "negbin", data)
}
\arguments{
\item{family}{A character string describing the observation model to
use in the likelihood. By default this is a negative binomial ("negbin")
with support for additional observation models being planned.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Setup observation model and data
}
\examples{
enw_obs(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
