% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_obs}
\alias{enw_obs}
\title{Setup observation model and data}
\usage{
enw_obs(family = "negbin", data)
}
\arguments{
\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Setup observation model and data
}
\examples{
enw_obs(data = enw_example("preprocessed"))
}
\seealso{
Functions used for modelling
\code{\link{enw_inits}()},
\code{\link{enw_model}()},
\code{\link{enw_priors}()},
\code{\link{enw_sample}()}
}
\concept{model}
