% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_obs}
\alias{enw_obs}
\title{Setup observation model and data}
\usage{
enw_obs(
  family = c("negbin", "negbin1d", "poisson"),
  observation_indicator = NULL,
  data
)
}
\arguments{
\item{family}{Character string, the observation model to use in the
likelihood; enforced by \code{\link[base:match.arg]{base::match.arg()}}. By default this is a
negative binomial with a quadratic mean-variance
relationship ("negbin"). Negative binomial with a linear
mean-variance relationship ("negbin1d") and Poisson ("poisson") are
also available.}

\item{observation_indicator}{A character string, the name of the column in
the data that indicates whether an observation is observed or not (using a
logical variable) and therefore whether or not it should be used in the
likelihood. This variable should be present in the data input to
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. It can be generated using \code{flag_observation} in
\code{\link[=enw_complete_dates]{enw_complete_dates()}} or it can be created directly using
\code{\link[=enw_flag_observed_observations]{enw_flag_observed_observations()}}. If either of these approaches are used
then the variable will be name \code{.observed}. Default is \code{NULL}.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Setup observation model and data
}
\examples{
enw_obs(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_dayofweek_structural_reporting}()},
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()},
\code{\link{enw_structural_reporting_metadata}()}
}
\concept{modelmodules}
