% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_obs}
\alias{enw_obs}
\title{Setup observation model and data}
\usage{
enw_obs(family = c("negbin", "poisson"), data)
}
\arguments{
\item{family}{Character string, the observation model to use in the
likelihood; enforced by \code{\link[base:match.arg]{base::match.arg()}}. By default this is a
negative binomial ("negbin") with Poisson ("poisson") also being
available. Support for additional observation models is planned,
please open an issue with suggestions.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Setup observation model and data
}
\examples{
enw_obs(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_reference}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
