% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_obs}
\alias{enw_obs}
\title{Setup observation model and data}
\usage{
enw_obs(family = "negbin", data)
}
\arguments{
\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Setup observation model and data
}
\examples{
enw_obs(data = enw_example("preprocessed"))
}
\seealso{
Other model: 
\code{\link{enw_model}()}
}
\concept{model}
