% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_pathfinder}
\alias{enw_pathfinder}
\title{Fit a CmdStan model using the pathfinder algorithm}
\usage{
enw_pathfinder(
  data,
  model = epinowcast::enw_model(),
  diagnostics = TRUE,
  init = NULL,
  ...
)
}
\arguments{
\item{data}{A list of data as produced by model modules (for example
\code{\link[=enw_expectation]{enw_expectation()}}, \code{\link[=enw_obs]{enw_obs()}}, etc.) and as required for use the
\code{model} being used.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}} or as
supplied by the user.}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be returned as a \code{data.frame}.}

\item{init}{A list of initial values or a function to generate initial
values. If not provided, the model will attempt to generate initial values}

\item{...}{Additional parameters to be passed to \code{CmdStanModel$pathfinder()}.}
}
\value{
A data.table containing the fit, data, and fit_args.
If diagnostics is TRUE, it also includes the run_time column with the timing
information.
}
\description{
For more information on the pathfinder algorithm see the
\href{https://mc-stan.org/cmdstanr/reference/model-method-pathfinder.html}{CmdStan documentation}. # nolint
}
\details{
Note that the \code{threads_per_chain} argument is renamed to \code{num_threads} to
match the \code{CmdStanModel$pathfinder()} method.

This fitting method is faster but more approximate than the NUTS sampler
used in \code{\link[=enw_sample]{enw_sample()}} and as such is recommended for use in exploratory
analysis and model development.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
pobs <- enw_example("preprocessed")

nowcast <- epinowcast(pobs,
  expectation = enw_expectation(~1, data = pobs),
  fit = enw_fit_opts(enw_pathfinder, pp = TRUE),
  obs = enw_obs(family = "poisson", data = pobs),
)

summary(nowcast)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_model}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_stan_to_r}()},
\code{\link{enw_unset_cache}()},
\code{\link{epinowcast_stan_header}()},
\code{\link{remove_profiling}()},
\code{\link{stan_cpp_adjoint_files}()},
\code{\link{stan_cpp_fallback_body}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
