% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_delay_counts}
\alias{enw_plot_delay_counts}
\title{Plot notifications by delay group}
\usage{
enw_plot_delay_counts(pobs, delay_group_thresh, facet = TRUE)
}
\arguments{
\item{pobs}{A preprocessed data object as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{delay_group_thresh}{A numeric vector defining
left-closed interval thresholds for delay groups.}

\item{facet}{Logical. When \code{TRUE} (the default), plots with
more than one \code{.group} are automatically wrapped by group.
Set to \code{FALSE} to disable and add a custom facet layer.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Stacked bar plot of notifications by reference
date, coloured by delay group.
}
\examples{
pobs <- enw_example("preprocessed_observations")
enw_plot_delay_counts(pobs, c(0, 2, 5, 10, 21))
}
\seealso{
Plotting functions
\code{\link{enw_delay_categories}()},
\code{\link{enw_delay_quantiles}()},
\code{\link{enw_plot_delay_cumulative}()},
\code{\link{enw_plot_delay_fraction}()},
\code{\link{enw_plot_delay_quantiles}()},
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
