% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_delay_cumulative}
\alias{enw_plot_delay_cumulative}
\title{Plot cumulative empirical reporting delay}
\usage{
enw_plot_delay_cumulative(pobs, delay_group_thresh)
}
\arguments{
\item{pobs}{A preprocessed data object as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{delay_group_thresh}{A numeric vector defining
left-closed interval thresholds for delay groups.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Stacked ribbon plot showing the cumulative
fraction reported by delay group over reference dates.
}
\examples{
pobs <- enw_example("preprocessed_observations")
enw_plot_delay_cumulative(pobs, c(0, 2, 5, 10, 21))
}
\seealso{
Plotting functions
\code{\link{enw_delay_categories}()},
\code{\link{enw_delay_quantiles}()},
\code{\link{enw_plot_delay_counts}()},
\code{\link{enw_plot_delay_fraction}()},
\code{\link{enw_plot_delay_quantiles}()},
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
