% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_emp_ts_del}
\alias{enw_plot_emp_ts_del}
\title{Stacked barplot of reported cases by reference date depending on delay}
\usage{
enw_plot_emp_ts_del(pobs, delay_group_thresh, ...)
}
\arguments{
\item{pobs}{Pre-processed observations as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{delay_group_thresh}{A vector defining thresholds for the grouping of
reporting delays based on left-closed intervals. Smallest value should be
zero, largest should correspond to max_delay.}

\item{...}{Additional arguments passed to \code{\link[=enw_plot_emprep_cum]{enw_plot_emprep_cum()}}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Stacked barplot of reported cases by reference date depending on delay
}
\seealso{
Plotting functions
\code{\link{enw_plot_emprep_cum}()},
\code{\link{enw_plot_emprep_frac}()},
\code{\link{enw_plot_emprep_quant}()},
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
