% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_obs}
\alias{enw_plot_obs}
\title{Generic quantile plot}
\usage{
enw_plot_obs(obs, latest_obs = NULL, log = TRUE, ...)
}
\arguments{
\item{obs}{A \code{data.frame} of summarised posterior estimates
containing at least a \code{confirm} count column and a date variable}

\item{latest_obs}{A\code{data.frame}of observed data containing at least a
\code{confirm} count variable and the same date variable as in the main data.frame
used for plotting.}

\item{log}{Logical, defaults to \code{FALSE}. Should counts be plot on the log
scale.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:aes]{ggplot2::aes()}} must at least
specify the x date variable.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Generic quantile plot
}
\examples{
nowcast <- enw_example("nowcast")
obs <- enw_example("obs")

# Plot observed data by reference date
enw_plot_obs(obs, x = reference_date)

# Plot observed data by reference date with more recent data
enw_plot_obs(nowcast$latest[[1]], obs, x = reference_date)
}
\seealso{
Plotting functions
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
