% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_quantiles}
\alias{enw_plot_quantiles}
\title{Generic quantile plot}
\usage{
enw_plot_quantiles(posterior, latest_obs = NULL, log = TRUE, ...)
}
\arguments{
\item{posterior}{A \code{data.frame} of summarised posterior estimates
containing at least a \code{confirm} count column a date variable,
quantile estimates for the 5\%, 20\%, 80\%, and 95\% quantiles and the
mean and median.}

\item{latest_obs}{A\code{data.frame}of observed data containing at least a
\code{confirm} count variable and the same date variable as in the main data.frame
used for plotting.}

\item{log}{Logical, defaults to \code{FALSE}. Should counts be plot on the log
scale.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:aes]{ggplot2::aes()}} must at least
specify the x date variable.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Generic quantile plot
}
\seealso{
Plotting functions
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
