% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_theme}
\alias{enw_plot_theme}
\title{FUNCTION_TITLE}
\usage{
enw_plot_theme(plot)
}
\arguments{
\item{plot}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
Plotting functions
\code{\link{enw_plot}()}
}
\concept{plot}
