% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{enw_plot_theme}
\alias{enw_plot_theme}
\title{Package plot theme}
\usage{
enw_plot_theme(plot)
}
\arguments{
\item{plot}{\code{ggplot2} plot object.}
}
\value{
\code{ggplot2} plot object.
}
\description{
Package plot theme
}
\seealso{
Plotting functions
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{plot.epinowcast}()}
}
\concept{plot}
