% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_posterior}
\alias{enw_posterior}
\title{Summarise the posterior}
\usage{
enw_posterior(fit, variables = NULL, probs = c(0.05, 0.2, 0.8, 0.95), ...)
}
\arguments{
\item{fit}{A \code{cmdstanr} fit object.}

\item{variables}{A character vector of variables to return
posterior summaries for. By default summaries for all parameters
are returned.}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work.}

\item{...}{Additional arguments that may be passed but will not be used.}
}
\value{
A \code{data.frame} summarising the model posterior.
}
\description{
A generic wrapper around \code{\link[posterior:draws_summary]{posterior::summarise_draws()}} with
opinionated defaults.
}
\examples{
fit <- enw_example("nowcast")
enw_posterior(fit$fit[[1]], variables = "expr_beta")
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{build_ord_obs}()},
\code{\link{enw_add_latest_obs_to_nowcast}()},
\code{\link{enw_nowcast_samples}()},
\code{\link{enw_nowcast_summary}()},
\code{\link{enw_pp_summary}()},
\code{\link{enw_quantiles_to_long}()},
\code{\link{enw_summarise_samples}()},
\code{\link{subset_obs}()}
}
\concept{postprocess}
