% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_pp_summary}
\alias{enw_pp_summary}
\title{FUNCTION_TITLE}
\usage{
enw_pp_summary(fit, diff_obs, probs = c(0.05, 0.2, 0.35, 0.5, 0.65, 0.8, 0.95))
}
\arguments{
\item{fit}{A \code{cmdstanr} fit object.}

\item{diff_obs}{PARAM_DESCRIPTION}

\item{probs}{A vector of numeric probabilities to produce
quantile summaries for. By default these are the 5\%, 20\%, 80\%,
and 95\% quantiles which are also the minimum set required for
plotting functions to work.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{enw_add_latest_obs_to_nowcast}()},
\code{\link{enw_nowcast_samples}()},
\code{\link{enw_nowcast_summary}()},
\code{\link{enw_posterior}()},
\code{\link{enw_quantiles_to_long}()},
\code{\link{enw_summarise_samples}()}
}
\concept{postprocess}
