% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_preprocess_data}
\alias{enw_preprocess_data}
\title{Preprocess observations}
\usage{
enw_preprocess_data(
  obs,
  by = c(),
  max_delay = 20,
  max_delay_strat = "add_to_max_delay",
  rep_holidays = c(),
  ref_holidays = c(),
  min_report_date,
  set_negatives_to_zero = TRUE
)
}
\arguments{
\item{obs}{A data frame containing at least the following variables:
\verb{reference date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{max_delay}{Numeric defaults to 20. The maximum number of days to
include in the delay distribution. Computation scales non-linearly with this
setting so consider what maximum makes sense for your data carefully.}

\item{max_delay_strat}{Character string indicating how to handle
reported cases beyond the specified maximum delay. Options include:
excluding ("exclude") and adding to the maximum delay ("add_to_max_delay").
Adding to the maximum delay is the default. Compare \code{confirm}, \code{max_confirm},
\code{prop_reported} columns to understand the impact of this assumption.}

\item{rep_holidays}{DESCRIPTION}

\item{ref_holidays}{DESCRIPTION}

\item{min_report_date}{PARAM_DESCRIPTION}

\item{set_negatives_to_zero}{Logical, defaults to TRUE. Should negative
counts (for calculated incidence of observations) be set to zero. Currently
downstream modelling does not support negative counts and so setting must be
TRUE if intending to use \code{\link[=epinowcast]{epinowcast()}}.}
}
\value{
A data.table containing processed observations as a series of nested
data frames.
}
\description{
This function preprocesses raw observations under the
assumption they are reported as cumulative counts by a reference and
report date and is used to assign groups. It also constructs data objects
used by downstreaming visualisation and modelling functions including the
observed empirical probability of a report on a given day, the cumulative
probability of report, the latest available observations, incidence of
observations, and metadata about the date of reference and report (used to
construct models).
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_obs}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()},
\code{\link{enw_retrospective_data}()}
}
\concept{preprocess}
