% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_preprocess_data}
\alias{enw_preprocess_data}
\title{Preprocess observations}
\usage{
enw_preprocess_data(
  obs,
  by = NULL,
  max_delay = 20,
  timestep = "day",
  set_negatives_to_zero = TRUE,
  ...,
  copy = TRUE
)
}
\arguments{
\item{obs}{A \code{data.frame} containing at least the following variables:
\code{reference_date} (index date of interest), \code{report_date} (report date for
observations), \code{confirm} (cumulative observations by reference and report
date).}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{max_delay}{Numeric defaults to 20 and needs to be greater than or equal
to 1 and an integer (internally it will be coerced to one using
\code{\link[=as.integer]{as.integer()}}). The maximum number of days to include in the delay
distribution. Computation scales non-linearly with this setting so consider
what maximum makes sense for your data carefully. Note that this is zero
indexed and so includes the reference date and \code{max_delay - 1} other days
(i.e. a \code{max_delay} of 1 corresponds with no delay). If a \code{max_delay} greater
than the maximum delay in the data is supplied then \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}
will throw a warning but in some cases this may be appropriate (e.g. if
at the beginning of a time series). In these cases the user should check the
model specification carefully as the model will be extrapolating beyond the
observed data.}

\item{timestep}{The timestep to used in the process model (i.e. the
reference date model). This can be a string ("day", "week", "month") or a
numeric whole number representing the number of days. If your data does not
have this timestep then you may wish to make use of
\code{\link[=enw_aggregate_cumulative]{enw_aggregate_cumulative()}} to aggregate your data to the desired timestep.}

\item{set_negatives_to_zero}{Logical, defaults to TRUE. Should negative
counts (for calculated incidence of observations) be set to zero. Currently
downstream modelling does not support negative counts and so setting must be
TRUE if intending to use \code{\link[=epinowcast]{epinowcast()}}.}

\item{...}{Other arguments to \code{\link[=enw_add_metaobs_features]{enw_add_metaobs_features()}},
e.g. \code{holidays}, which sets commonly used metadata
(e.g. day of week, days since start of time series)}

\item{copy}{A logical; if \code{TRUE} (the default) creates a copy; otherwise,
modifies \code{obs} in place.}
}
\value{
A data.table containing processed observations as a series of nested
data.frames as well as variables containing metadata. These are:
\itemize{
\item \code{obs}: (observations with the addition of empirical reporting proportions
and and restricted to the specified maximum delay).
\item \code{new_confirm}: Incidence of notifications by reference and report date.
Empirical reporting distributions are also added.
\item \code{latest}: The latest available observations.
\item \code{missing_reference}: Observations missing reference dates.
\item \code{reporting_triangle}: Incident observations by report and reference date in
the standard reporting triangle matrix format.
\item \code{metareference}: Metadata reference dates derived from observations.
\item \code{metrareport}: Metadata for report dates.
\item \code{metadelay}: Metadata for reporting delays produced using
\code{\link[=enw_delay_metadata]{enw_delay_metadata()}}.
\item \code{time}: Numeric, number of timepoints in the data.
\item \code{snapshots}: Numeric, number of available data snapshots to use for
nowcasting.
\item \code{groups}: Numeric, Number of groups/strata in the supplied observations
(set using \code{by}).
\item \code{max_delay}: Numeric, the maximum delay in the processed data
\item \code{max_date}: The maximum available report date.
}
}
\description{
This function preprocesses raw observations under the
assumption they are reported as cumulative counts by a reference and
report date and is used to assign groups. It also constructs data objects
used by visualisation and modelling functions including the
observed empirical probability of a report on a given day, the cumulative
probability of report, the latest available observations, incidence of
observations, and metadata about the date of reference and report (used to
construct models). This function wraps other preprocessing functions that may
be instead used individually if required. Note that internally reports
beyond the user specified delay are dropped for modelling purposes with the
\code{cum_prop_reported} and \code{max_confirm} variables allowing the user to check
the impact this may have (if \code{cum_prop_reported} is significantly below 1 a
longer \code{max_delay} may be appropriate). Also note that if missing reference
or report dates are suspected to occur in your data then these need to be
completed with \code{\link[=enw_complete_dates]{enw_complete_dates()}}.
}
\examples{
library(data.table)

# Filter example hospitalisation data to be national and over all ages
nat_germany_hosp <- germany_covid19_hosp[location == "DE"]
nat_germany_hosp <- nat_germany_hosp[age_group \%in\% "00+"]

# Preprocess with default settings
pobs <- enw_preprocess_data(nat_germany_hosp)
pobs
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_flag_observed_observations}()},
\code{\link{enw_impute_na_observations}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
