% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_priors_as_data_list}
\alias{enw_priors_as_data_list}
\title{Convert prior \code{data.frame} to list}
\usage{
enw_priors_as_data_list(priors)
}
\arguments{
\item{priors}{A \code{data.frame} with columns \code{variable},
\code{mean}, and \code{sd} describing normal priors.
Default priors in the appropriate format are returned by
the \verb{$priors} element of \code{\link[=enw_reference]{enw_reference()}} and other model
module functions.}
}
\value{
A named list with each entry specifying a prior as a length
two vector (specifying the mean and standard deviation of the prior).
}
\description{
Converts priors defined in a \code{data.frame} into a list
format for use by stan. In addition it adds "_p" to all
variable names in order too allow them to be distinguished from
their standard usage within modelling code.
}
\examples{
priors <- data.frame(variable = "x", mean = 1, sd = 2)
enw_priors_as_data_list(priors)
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_model}()},
\code{\link{enw_pathfinder}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_stan_to_r}()},
\code{\link{enw_unset_cache}()},
\code{\link{epinowcast_stan_header}()},
\code{\link{remove_profiling}()},
\code{\link{stan_cpp_adjoint_files}()},
\code{\link{stan_cpp_fallback_body}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
