% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_priors_as_data_list}
\alias{enw_priors_as_data_list}
\title{FUNCTION_TITLE}
\usage{
enw_priors_as_data_list(priors)
}
\arguments{
\item{priors}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
priors <- enw_priors()
enw_priors_as_data_list(priors)
}
\concept{priors}
