% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_priors_as_data_list}
\alias{enw_priors_as_data_list}
\title{FUNCTION_TITLE}
\usage{
enw_priors_as_data_list(priors)
}
\arguments{
\item{priors}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
priors <- enw_priors()
enw_priors_as_data_list(priors)
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_fit_opts}()},
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_posterior_as_prior}()},
\code{\link{enw_replace_priors}()},
\code{\link{remove_profiling}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
