% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{enw_quantiles_to_long}
\alias{enw_quantiles_to_long}
\title{Convert summarised quantiles from wide to long format}
\usage{
enw_quantiles_to_long(posterior)
}
\arguments{
\item{posterior}{A \code{data.frame} as output by \code{\link[=enw_posterior]{enw_posterior()}}.}
}
\value{
A \code{data.frame} of quantiles in long format.
}
\description{
Convert summarised quantiles from wide to long format
}
\examples{
fit <- enw_example("nowcast")
posterior <- enw_posterior(fit$fit[[1]], var = "expr_lelatent_int[1,1]")
enw_quantiles_to_long(posterior)
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{enw_add_latest_obs_to_nowcast}()},
\code{\link{enw_nowcast_samples}()},
\code{\link{enw_nowcast_summary}()},
\code{\link{enw_posterior}()},
\code{\link{enw_pp_summary}()},
\code{\link{enw_summarise_samples}()}
}
\concept{postprocess}
