% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_reference}
\alias{enw_reference}
\title{Reference date logit hazard reporting  model module}
\usage{
enw_reference(
  parametric = ~1,
  distribution = "lognormal",
  non_parametric = ~0,
  data
)
}
\arguments{
\item{parametric}{DESCRIPTION}

\item{distribution}{DESCRIPTION}

\item{non_parametric}{DESCRIPTION}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Reference date logit hazard reporting  model module
}
\examples{
enw_reference(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_report}()}
}
\concept{modelmodules}
