% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{enw_reference_by_report}
\alias{enw_reference_by_report}
\title{Construct a lookup of references dates by report}
\usage{
enw_reference_by_report(
  missing_reference,
  reps_with_complete_refs,
  metareference,
  max_delay
)
}
\arguments{
\item{missing_reference}{\code{missing_reference} \code{data.frame} output from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{reps_with_complete_refs}{A \code{data.frame} of report dates with complete
(i.e fully reported) reference dates as produced using
\code{\link[=enw_reps_with_complete_refs]{enw_reps_with_complete_refs()}}.}

\item{metareference}{\code{metareference} \code{data.frame} output from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{max_delay}{The maximum number of days to model in the delay
distribution. Must be an integer greater than or equal to 1. Observations
with delays larger then the maximum delay will be dropped. If the specified
maximum delay is too short, nowcasts can be biased as important parts of the
true delay distribution are cut off. At the same time, computational cost
scales non-linearly with this setting, so you want the maximum delay to be as
long as necessary, but not much longer. Consider what delays are realistic
for your application, and when in doubt, check if increasing the maximum
delay noticeably changes the delay distribution or nowcasts as estimated by
epinowcast. If it does, your maximum delay may still be too short.
Note that delays are zero indexed and so include the reference date and
\code{max_delay - 1} other days (i.e. a \code{max_delay} of 1 corresponds to
no delay). You can use \code{\link[=check_max_delay]{check_max_delay()}} to check the coverage of a delay
distribution for different maximum delays.}
}
\value{
A wide \code{data.frame} with each row being a complete report date and'
the columns being the observation index for each reporting delay
}
\description{
Construct a lookup of references dates by report
}
\seealso{
Helper functions for model modules
\code{\link{add_max_observed_delay}()},
\code{\link{add_pmfs}()},
\code{\link{convolution_matrix}()},
\code{\link{enw_reps_with_complete_refs}()},
\code{\link{extract_obs_metadata}()},
\code{\link{extract_sparse_matrix}()},
\code{\link{latest_obs_as_matrix}()}
}
\concept{modelmodulehelpers}
