% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_replace_priors}
\alias{enw_replace_priors}
\title{Replace default priors with user specified priors}
\usage{
enw_replace_priors(priors, custom_priors)
}
\arguments{
\item{priors}{A \code{data.frame} with columns \code{variable},
\code{mean}, and \code{sd} describing normal priors.
Default priors in the appropriate format are returned by
the \verb{$priors} element of \code{\link[=enw_reference]{enw_reference()}} and other model
module functions.}

\item{custom_priors}{A \code{data.frame} with columns \code{variable},
\code{mean}, and \code{sd} describing normal priors to use as
replacements.
Rows in \code{custom_priors} replace matching rows in \code{priors}
by the \code{variable} column.
Vectorised prior names (i.e. those of the form
\code{variable[n]}) are matched after stripping the index
(treated as \code{variable}).}
}
\value{
A data.table of prior definitions (variable, mean and sd).
}
\description{
Replaces default model priors with user specified ones
(restricted to normal priors with specified mean and standard
deviations).
A common use is extracting the posterior from a previous
\code{\link[=epinowcast]{epinowcast()}} run (using \code{summary(nowcast, type = "fit")})
and using it as a prior for subsequent fits.
}
\details{
Default priors can be obtained from each model module's
\verb{$priors} element, e.g. \code{enw_reference(data = pobs)$priors}.
See the \code{priors} argument of \code{\link[=epinowcast]{epinowcast()}} for a list of
available prior variable names by module.
}
\examples{
# Update priors from a data.frame
priors <- data.frame(variable = c("x", "y"), mean = c(1, 2), sd = c(1, 2))
custom_priors <- data.frame(variable = "x[1]", mean = 10, sd = 2)
enw_replace_priors(priors, custom_priors)

# Update priors from a previous model fit
default_priors <- enw_reference(
  distribution = "lognormal",
  data = enw_example("preprocessed"),
)$priors
print(default_priors)

fit_priors <- summary(
  enw_example("nowcast"),
  type = "fit",
  variables = c("refp_mean_int", "refp_sd_int", "sqrt_phi")
)
fit_priors

enw_replace_priors(default_priors, fit_priors)
}
\seealso{
Functions used to help convert models into the format required for stan:
\code{\link[=enw_formula_as_data_list]{enw_formula_as_data_list()}},
\code{\link[=enw_get_cache]{enw_get_cache()}},
\code{\link[=enw_model]{enw_model()}},
\code{\link[=enw_pathfinder]{enw_pathfinder()}},
\code{\link[=enw_priors_as_data_list]{enw_priors_as_data_list()}},
\code{\link[=enw_sample]{enw_sample()}},
\code{\link[=enw_set_cache]{enw_set_cache()}},
\code{\link[=enw_stan_to_r]{enw_stan_to_r()}},
\code{\link[=enw_unset_cache]{enw_unset_cache()}},
\code{\link[=remove_profiling]{remove_profiling()}},
\code{\link[=write_stan_files_no_profile]{write_stan_files_no_profile()}}
}
\concept{modeltools}
