% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_replace_priors}
\alias{enw_replace_priors}
\title{Replace default priors with user specfied priors}
\usage{
enw_replace_priors(priors, custom_priors)
}
\arguments{
\item{priors}{A data.frame with the following variables:
\code{variable}, \code{mean}, \code{sd} describing normal priors. Priors in the
appropriate format are returned by \code{\link[=enw_reference]{enw_reference()}} as well as by
other similar model specification functions.}

\item{custom_priors}{A data.frame with the following variables:
\code{variable}, \code{mean}, \code{sd} describing normal priors. Priors in the
appropriate format are returned by \code{\link[=enw_reference]{enw_reference()}} as well as by
other similar model specification functions. Priors in this data.frame
replace the default priors.}
}
\value{
A data.table of prior definitions (variable, mean and sd).
}
\description{
This function is used internally by \url{epinowcast} to replace
default model priors with users specified ones (restricted to
normal priors with specified mean and standard deviations). A common
use would be extracting the posterior from a previous \code{\link[=epinowcast]{epinowcast()}}
run (using \code{summary(nowcast, type = fit)}) and using this a prior.
}
\examples{
priors <- data.frame(variable = c("x", "y"), mean = c(1, 2), sd = c(1, 2))
custom_priors <- data.frame(variable = "x", mean = 10, sd = 2)
enw_replace_priors(priors, custom_priors)
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_sample}()},
\code{\link{remove_profiling}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
