% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_replace_priors}
\alias{enw_replace_priors}
\title{Replace default priors with user specified priors}
\usage{
enw_replace_priors(priors, custom_priors)
}
\arguments{
\item{priors}{A \code{data.frame} with the following variables:
\code{variable}, \code{mean}, \code{sd} describing normal priors. Priors in the
appropriate format are returned by \code{\link[=enw_reference]{enw_reference()}} as well as by
other similar model specification functions.}

\item{custom_priors}{A \code{data.frame} with the following variables:
\code{variable}, \code{mean}, \code{sd} describing normal priors. Priors in the
appropriate format are returned by \code{\link[=enw_reference]{enw_reference()}} as well as by
other similar model specification functions. Priors in this data.frame
replace the default priors. Note that currently vectorised prior names
(i.e those of the form \code{variable[n]} will be treated as \code{variable}).}
}
\value{
A data.table of prior definitions (variable, mean and sd).
}
\description{
This function is used internally by \url{epinowcast} to replace
default model priors with users specified ones (restricted to
normal priors with specified mean and standard deviations). A common
use would be extracting the posterior from a previous \code{\link[=epinowcast]{epinowcast()}}
run (using \code{summary(nowcast, type = fit)}) and using this a prior.
}
\examples{
# Update priors from a data.frame
priors <- data.frame(variable = c("x", "y"), mean = c(1, 2), sd = c(1, 2))
custom_priors <- data.frame(variable = "x[1]", mean = 10, sd = 2)
enw_replace_priors(priors, custom_priors)

# Update priors from a previous model fit
default_priors <- enw_reference(
 distribution = "lognormal", data = obs
)$priors
print(default_priors)

fit_priors <- summary(
 enw_example("nowcast"), type = "fit",
 variables = c("refp_mean_int", "refp_sd_int", "sqrt_phi")
)
fit_priors

enw_replace_priors(default_priors, fit_priors)
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_model}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_sample}()},
\code{\link{remove_profiling}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
