% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_report}
\alias{enw_report}
\title{Report date logit hazard reporting  model module}
\usage{
enw_report(formula = ~0, structural = ~0, data)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link{lme4} syntax (see \code{\link[=re]{re()}}), and random walks
defined using the \code{\link[=rw]{rw()}} helper function.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list as required by stan.
}
\description{
Report date logit hazard reporting  model module
}
\examples{
enw_report(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()}
}
\concept{modelmodules}
