% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_report}
\alias{enw_report}
\title{Report date logit hazard reporting  model module}
\usage{
enw_report(non_parametric = ~0, structural = ~0, data)
}
\arguments{
\item{non_parametric}{A formula (as implemented in \code{\link[=enw_formula]{enw_formula()}})
describing the non-parametric logit hazard model for report date effects.
This can use features defined by report date as defined in \code{metareport} as
produced by \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. Note that the intercept for this model
is set to 0 as it should be used for specifying report date related hazards
rather than time-invariant hazards, which should instead be modelled using
the \code{non_parametric} argument of \code{\link[=enw_reference]{enw_reference()}}. Set to \code{~0} to disable
(internally converted to \code{~1} and flagged as inactive). See \code{\link[=enw_formula]{enw_formula()}}
for details on formula syntax.}

\item{structural}{A formula with fixed effects and using only binary
variables, and factors describing the known reporting structure (i.e weekday
only reporting). The base case (i.e the first factor entry) should describe
the dates for which reporting is possible. Internally dates with a non-zero
element in the design matrix have their hazard set to 0. This can use
features defined by report date as defined in \code{metareport} as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. Note that the intercept for this model is set to 0
in order to allow all dates without other structural reasons to not be
reported to be reported. Note that this feature is not yet available to
users.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list containing the supplied formulas, data passed into a list
describing the models, a \code{data.frame} describing the priors used, and a
function that takes the output data and priors and returns a function that
can be used to sample from a tightened version of the prior distribution.
}
\description{
Report date logit hazard reporting  model module
}
\examples{
enw_report(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()}
}
\concept{modelmodules}
