% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-modules.R
\name{enw_report}
\alias{enw_report}
\title{Report date logit hazard reporting  model module}
\usage{
enw_report(non_parametric = ~0, structural = NULL, data)
}
\arguments{
\item{non_parametric}{A formula (as implemented in \code{\link[=enw_formula]{enw_formula()}})
describing the non-parametric logit hazard model. This can use features
defined by report date as defined in \code{metareport} as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}. Note that the intercept for this model is set to 0
as it should be used for specifying report date related hazards vs time
invariant hazards which should instead be modelled using the
\code{non_parametric} argument of \code{\link[=enw_reference]{enw_reference()}}}

\item{structural}{A nested list of matrices by group and reference date
describing the known reporting structure (i.e weekday only reporting).
Each matrix should have dimensions of max_delay x max_delay, where
each column represents the probability of an observation being reported
on a specific day given the delay. This is particularly useful for
modeling fixed reporting cycles, such as weekly reporting on Wednesdays
as seen in the German hospitalization data example.}

\item{data}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}
}
\value{
A list containing the supplied formulas, data passed into a list
describing the models, a \code{data.frame} describing the priors used, and a
function that takes the output data and priors and returns a function that
can be used to sample from a tightened version of the prior distribution.
}
\description{
Report date logit hazard reporting  model module
}
\examples{
enw_report(data = enw_example("preprocessed"))
}
\seealso{
Model modules
\code{\link{enw_expectation}()},
\code{\link{enw_fit_opts}()},
\code{\link{enw_missing}()},
\code{\link{enw_obs}()},
\code{\link{enw_reference}()}
}
\concept{modelmodules}
