% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_reporting_triangle}
\alias{enw_reporting_triangle}
\title{Construct the reporting triangle}
\usage{
enw_reporting_triangle(obs)
}
\arguments{
\item{obs}{A \code{data.frame} as produced by \code{\link[=enw_add_incidence]{enw_add_incidence()}}.
Must contain the following variables: \code{reference_date}, \code{.group}, \code{delay}.}
}
\value{
A \code{data.frame} with each row being a reference date, and columns
being observations by reporting delay.
}
\description{
Constructs the reporting triangle with each row representing a reference date
and columns being observations by report date
}
\examples{
obs <- enw_example("preprocessed")$new_confirm
enw_reporting_triangle(obs)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_flag_observed_observations}()},
\code{\link{enw_impute_na_observations}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_obs_at_delay}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()}
}
\concept{preprocess}
