% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_reporting_triangle_to_long}
\alias{enw_reporting_triangle_to_long}
\title{Recast the reporting triangle from wide to long format}
\usage{
enw_reporting_triangle_to_long(obs)
}
\arguments{
\item{obs}{A \code{data.frame} in the format produced by
\code{\link[=enw_reporting_triangle]{enw_reporting_triangle()}}.}
}
\value{
A long format reporting triangle as a \code{data.frame} with additional
variables \code{new_confirm} and \code{delay}.
}
\description{
Recast the reporting triangle from wide to long format
}
\examples{
obs <- enw_example("preprocessed")$new_confirm
rt <- enw_reporting_triangle(obs)
enw_reporting_triangle_to_long(rt)
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_complete_dates}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_delay}()},
\code{\link{enw_filter_reference_dates}()},
\code{\link{enw_filter_report_dates}()},
\code{\link{enw_flag_observed_observations}()},
\code{\link{enw_impute_na_observations}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata_delay}()},
\code{\link{enw_metadata}()},
\code{\link{enw_missing_reference}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
