% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{enw_reps_with_complete_refs}
\alias{enw_reps_with_complete_refs}
\title{Identify report dates with complete (i.e up to the maximum delay) reference
dates}
\usage{
enw_reps_with_complete_refs(new_confirm, max_delay, by = NULL, copy = TRUE)
}
\arguments{
\item{new_confirm}{\code{new_confirm} \code{data.frame} output from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{max_delay}{Numeric defaults to 20 and needs to be greater than or equal
to 1 and an integer (internally it will be coerced to one using
\code{\link[=as.integer]{as.integer()}}). The maximum number of days to include in the delay
distribution. Observations with delays larger then the maximum delay will be
dropped. If the specified maximum delay is too short, nowcasts can be biased
as important parts of the true delay distribution are cut off. At the same
time, computational cost scales non-linearly with this setting, so you want
the maximum delay to be as long as necessary, but not much longer. Consider
what delays are realistic for your application, and when in doubt, check if
increasing the maximum delay noticeably changes the delay distribution or
nowcasts as estimated by epinowcast. If it does, your maximum delay may still
be too short. Note that delays are zero indexed and so include the reference
date and \code{max_delay - 1} other days (i.e. a \code{max_delay} of 1 corresponds to
no delay). See \code{\link[=enw_metadata_delay]{enw_metadata_delay()}} for checking the coverage of a delay
distribution for a given maximum delay.}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{copy}{A logical; if \code{TRUE} (the default) creates a copy; otherwise,
modifies \code{obs} in place.}
}
\value{
A \code{data.frame} containing a \code{report_date} variable, and grouping
variables specified for report dates that have complete reporting.
}
\description{
Identify report dates with complete (i.e up to the maximum delay) reference
dates
}
\seealso{
Helper functions for model modules
\code{\link{add_pmfs}()},
\code{\link{convolution_matrix}()},
\code{\link{enw_reference_by_report}()},
\code{\link{extract_sparse_matrix}()},
\code{\link{latest_obs_as_matrix}()}
}
\concept{modelmodulehelpers}
