% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{enw_reps_with_complete_refs}
\alias{enw_reps_with_complete_refs}
\title{Identify report dates with complete (i.e up to the maximum delay) reference
dates}
\usage{
enw_reps_with_complete_refs(new_confirm, max_delay, by = NULL, copy = TRUE)
}
\arguments{
\item{new_confirm}{\code{new_confirm} \code{data.frame} output from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{max_delay}{Numeric, defaults to 20. The maximum number of days to
include in the delay distribution. Observations with a delay larger than
\code{max_delay} will be dropped during preprocessing. The time required for model
fitting increases non-linearly with this setting, so carefully choose a
maximum that makes sense for your data. Note that delays use zero-based
indexing, i.e. \code{max_delay} includes the reference date and the
\code{max_delay - 1} following days.}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}

\item{copy}{A logical; if \code{TRUE} (the default) creates a copy; otherwise,
modifies \code{obs} in place.}
}
\value{
A \code{data.frame} containing a \code{report_date} variable, and grouping
variables specified for report dates that have complete reporting.
}
\description{
Identify report dates with complete (i.e up to the maximum delay) reference
dates
}
\seealso{
Helper functions for model modules
\code{\link{add_pmfs}()},
\code{\link{convolution_matrix}()},
\code{\link{enw_reference_by_report}()},
\code{\link{extract_sparse_matrix}()},
\code{\link{latest_obs_as_matrix}()}
}
\concept{modelmodulehelpers}
