% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{enw_reps_with_complete_refs}
\alias{enw_reps_with_complete_refs}
\title{Identify report dates with complete (i.e up to the maximum delay) reference
dates}
\usage{
enw_reps_with_complete_refs(new_confirm, max_delay, by = c())
}
\arguments{
\item{new_confirm}{\code{new_confirm} data.frame output from
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{max_delay}{Numeric defaults to 20. The maximum number of days to
include in the delay distribution. Computation scales non-linearly with this
setting so consider what maximum makes sense for your data carefully. Note
that this is zero indexed and so includes the reference date and
\code{max_delay - 1} other days.}

\item{by}{A character vector describing the stratification of
observations. This defaults to no grouping. This should be used
when modelling multiple time series in order to identify them for
downstream modelling}
}
\value{
A data frame containing a \code{report_date} variable, and grouping
variables specified for report dates that have complete reporting.
}
\description{
Identify report dates with complete (i.e up to the maximum delay) reference
dates
}
\seealso{
Helper functions for model modules
\code{\link{enw_reference_by_report}()}
}
\concept{modelmodulehelpers}
