% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{enw_retrospective_data}
\alias{enw_retrospective_data}
\title{Create retrospective data from a full data set}
\usage{
enw_retrospective_data(
  obs,
  latest_rep_date,
  remove_rep_days,
  earliest_ref_date,
  include_ref_days
)
}
\arguments{
\item{obs}{an observation data frame containing \code{report_date} and
\code{reference_date} columns.}

\item{latest_rep_date}{latest report date to include in the retrospective
data set}

\item{remove_rep_days}{if \code{latest_rep_date} is not given, the number
of report dates to remove, starting from the latest date included}

\item{earliest_ref_date}{earliest reference date to include in the
retrospective data set}

\item{include_ref_days}{if \code{earilest_ref_date} is not given, the number
of reference dates to include, ending with the latest reference
date included once reporting dates have been removed.}
}
\value{
A data.table containing the retrospective data set
}
\description{
This is a helper function which allows users to create
truncated data sets at past time points from a given larger data set.
This is useful when evaluating nowcast performance against fully
observed data.
}
\seealso{
Preprocessing functions
\code{\link{enw_add_delay}()},
\code{\link{enw_add_max_reported}()},
\code{\link{enw_add_metaobs_features}()},
\code{\link{enw_assign_group}()},
\code{\link{enw_construct_data}()},
\code{\link{enw_delay_metadata}()},
\code{\link{enw_extend_date}()},
\code{\link{enw_filter_obs}()},
\code{\link{enw_latest_data}()},
\code{\link{enw_metadata}()},
\code{\link{enw_new_reports}()},
\code{\link{enw_preprocess_data}()},
\code{\link{enw_reporting_triangle_to_long}()},
\code{\link{enw_reporting_triangle}()}
}
\concept{preprocess}
