% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_sample}
\alias{enw_sample}
\title{Fit a CmdStan model using NUTS}
\usage{
enw_sample(
  data,
  model = epinowcast::enw_model(),
  init = NULL,
  init_method = c("prior", "pathfinder"),
  init_method_args = list(),
  diagnostics = TRUE,
  ...
)
}
\arguments{
\item{data}{A list of data as produced by model modules (for example
\code{\link[=enw_expectation]{enw_expectation()}}, \code{\link[=enw_obs]{enw_obs()}}, etc.) and as required for use the
\code{model} being used.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}} or as
supplied by the user.}

\item{init}{A list of initial values or a function to generate initial
values. If not provided, the model will attempt to generate initial values}

\item{init_method}{The method to use for initializing the model. Defaults to
"prior" which samples initial values from the prior. "pathfinder", which uses
the pathfinder algorithm (\code{\link[=enw_pathfinder]{enw_pathfinder()}}) to initialize the model.}

\item{init_method_args}{A list of additional arguments to pass to the
initialization method.}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be returned as a \code{data.frame}.}

\item{...}{Additional parameters passed to the \code{sample} method of \code{cmdstanr}.}
}
\value{
A \code{data.frame} containing the \code{cmdstanr} fit, the input data, the
fitting arguments, and optionally summary diagnostics.
}
\description{
Fit a CmdStan model using NUTS
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
pobs <- enw_example("preprocessed")

nowcast <- epinowcast(pobs,
  expectation = enw_expectation(~1, data = pobs),
  fit = enw_fit_opts(enw_sample, pp = TRUE),
  obs = enw_obs(family = "poisson", data = pobs),
)

summary(nowcast)

# Use pathfinder initialization
nowcast_pathfinder <- epinowcast(pobs,
  expectation = enw_expectation(~1, data = pobs),
  fit = enw_fit_opts(enw_sample, pp = TRUE, init_method = "pathfinder"),
  obs = enw_obs(family = "poisson", data = pobs),
)

summary(nowcast_pathfinder)
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used to help convert models into the format required for stan:
\code{\link[=enw_formula_as_data_list]{enw_formula_as_data_list()}},
\code{\link[=enw_get_cache]{enw_get_cache()}},
\code{\link[=enw_model]{enw_model()}},
\code{\link[=enw_pathfinder]{enw_pathfinder()}},
\code{\link[=enw_priors_as_data_list]{enw_priors_as_data_list()}},
\code{\link[=enw_replace_priors]{enw_replace_priors()}},
\code{\link[=enw_set_cache]{enw_set_cache()}},
\code{\link[=enw_stan_to_r]{enw_stan_to_r()}},
\code{\link[=enw_unset_cache]{enw_unset_cache()}},
\code{\link[=remove_profiling]{remove_profiling()}},
\code{\link[=write_stan_files_no_profile]{write_stan_files_no_profile()}}
}
\concept{modeltools}
