% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{enw_sample}
\alias{enw_sample}
\title{Fit a brancing process strain model}
\usage{
enw_sample(data, model = epinowcast::enw_model(), diagnostics = TRUE, ...)
}
\arguments{
\item{data}{A list of data as produced by \code{\link[=enw_as_data_list]{enw_as_data_list()}}.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}}.}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be returned as a \code{data.frame}.}

\item{...}{Additional parameters passed to the \code{sample} method of \code{cmdstanr}.}
}
\value{
A \code{data.frame} containing the \code{cmdstanr} fit, the input data, the
fitting arguments, and optionally summary diagnostics.
}
\description{
Fit a brancing process strain model
}
\seealso{
Functions used for modelling
\code{\link{enw_as_data_list}()},
\code{\link{enw_inits}()},
\code{\link{enw_model}()}
}
\concept{model}
