% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_sample}
\alias{enw_sample}
\title{Fit a CmdStan model using NUTS}
\usage{
enw_sample(data, model = epinowcast::enw_model(), diagnostics = TRUE, ...)
}
\arguments{
\item{data}{A list of data as produced by model modules (for example
\code{\link[=enw_expectation]{enw_expectation()}}, \code{\link[=enw_obs]{enw_obs()}}, etc.) and as required for use the
\code{model} being used.}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}} or as
supplied by the user.}

\item{diagnostics}{Logical, defaults to \code{TRUE}. Should fitting diagnostics
be returned as a  \code{data.frame}.}

\item{...}{Additional parameters passed to the \code{sample} method of \code{cmdstanr}.}
}
\value{
A  \code{data.frame} containing the \code{cmdstanr} fit, the input data, the
fitting arguments, and optionally summary diagnostics.
}
\description{
Fit a CmdStan model using NUTS
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_model}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{remove_profiling}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
