% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-validation.R
\name{enw_score_nowcast}
\alias{enw_score_nowcast}
\title{FUNCTION_TITLE}
\usage{
enw_score_nowcast(nowcast, latest_obs, round_to = 3, ...)
}
\arguments{
\item{nowcast}{DESCRIPTION.}

\item{latest_obs}{DESCRIPTION.}

\item{round_to}{DESCRIPTION}

\item{...}{Additional arguments passed to \code{\link[scoringutils:eval_forecasts]{scoringutils::eval_forecasts()}}.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\concept{modelvalidation}
