% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{enw_set_cache}
\alias{enw_set_cache}
\title{Set caching location for Stan models}
\usage{
enw_set_cache(path, type = c("session", "persistent", "all"))
}
\arguments{
\item{path}{A valid filepath representing the desired cache location. If
the directory does not exist it will be created.}

\item{type}{A character string specifying the cache type. It can be one of
"session", "persistent", or "all". Default is "session".
"session" sets the cache for the current session, "persistent" writes the
cache location to the user's \code{.Renviron} file,  and "all" does both.}
}
\value{
The string of the filepath set.
}
\description{
This function allows the user to set a cache location for Stan models
rather than a temporary directory. This can reduce the need for model
compilation on every new model run across sessions or within a session.
For R version 4.0.0 and above, it's recommended to use the persistent cache
as shown in the example.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Set to local directory
my_enw_cache <- enw_set_cache(file.path(tempdir(), "test"))
enw_get_cache()
\dontrun{
# Use the package cache in R >= 4.0
if (R.version.string >= "4.0.0") {
  enw_set_cache(
    tools::R_user_dir(package = "epinowcast", "cache"),
    type = "all"
  )
}
}
\dontshow{\}) # examplesIf}
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_model}()},
\code{\link{enw_pathfinder}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{enw_stan_to_r}()},
\code{\link{enw_unset_cache}()},
\code{\link{epinowcast_stan_header}()},
\code{\link{remove_profiling}()},
\code{\link{stage_stan_header}()},
\code{\link{stan_cpp_adjoint_files}()},
\code{\link{stan_cpp_fallback_body}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
