% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_set_cache}
\alias{enw_set_cache}
\title{Set caching location for Stan models}
\usage{
enw_set_cache(path)
}
\arguments{
\item{path}{A valid filepath representing the desired cache location}
}
\value{
The string of the filepath set
}
\description{
This function allows the user to set a cache location for
Stan models rather than a temp directory. This can reduce the
need for model compilation on every new model run across sessions.
}
\examples{
# Set to local directory
my_enw_cache <- file.path(tempdir(), "test")
enw_set_cache(my_enw_cache)
enw_get_cache()
\dontrun{
# Use the package cache in R >= 4.0
if(R.version.string >= 4.0) {
  enw_set_cache(tools::R_user_dir(package = "epinowcast", "cache"))
}
}
}
