% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_unset_cache}
\alias{enw_unset_cache}
\title{Unset Stan cache location}
\usage{
enw_unset_cache(persistent = FALSE)
}
\arguments{
\item{persistent}{A logical representing if the cache location should be
removed from the user's \code{.Renviron} file with default of \code{FALSE}.}
}
\value{
the prior cache location, if it existed
}
\description{
Removes \code{enw_cache_location} environment variable from
the user .Renviron file and removes it from the local
environment.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
enw_unset_cache()

enw_unset_cache(enw_set_cache(file.path(tempdir(), "test")))
\dontshow{\}) # examplesIf}
}
\seealso{
Utility functions
\code{\link{aggregate_rolling_sum}()},
\code{\link{coerce_date}()},
\code{\link{coerce_dt}()},
\code{\link{date_to_numeric_modulus}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_set_cache}()},
\code{\link{get_internal_timestep}()},
\code{\link{is.Date}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
