% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{enw_unset_cache}
\alias{enw_unset_cache}
\title{Unset Stan cache location}
\usage{
enw_unset_cache()
}
\value{
the prior cache location, if it existed
}
\description{
Removes \code{enw_cache_location} environment variable from
the user .Renviron file and removes it from the local
environment.
}
\examples{
enw_unset_cache()

}
\seealso{
Utility functions
\code{\link{aggregate_rolling_sum}()},
\code{\link{coerce_date}()},
\code{\link{coerce_dt}()},
\code{\link{date_to_numeric_modulus}()},
\code{\link{enw_get_cache}()},
\code{\link{get_internal_timestep}()},
\code{\link{is.Date}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
