% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast-package.R
\docType{package}
\name{epinowcast-package}
\alias{epinowcast-package}
\title{epinowcast: A Bayesian Framework for Real-Time Infectious Disease Surveillance}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A modular Bayesian framework for real-time infectious disease surveillance. Provides tools for nowcasting, reproduction number estimation, delay estimation, and forecasting from data subject to reporting delays, right-truncation, missing data, and incomplete ascertainment. Users can build models suited to their setting using a flexible formula interface supporting fixed effects, random effects, random walks, and time-varying parameters, with options including parametric and non-parametric delay distributions with optional modifiers (via discrete-time hazard models), renewal processes, observation models, missing data imputation, and stratified analyses with partial pooling. By jointly estimating disease dynamics and reporting patterns, our framework enables earlier and more reliable detection of trends. While designed with epidemiological applications in mind, the framework can be applied to any right-truncated time series count data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://package.epinowcast.org}
  \item \url{https://github.com/epinowcast/epinowcast/}
  \item Report bugs at \url{https://github.com/epinowcast/epinowcast/issues/}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Adrian Lison \email{adrian.lison@bsse.ethz.ch} (\href{https://orcid.org/0000-0002-6822-8437}{ORCID})
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
  \item Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})
  \item Hugo Gruson \email{hugo.gruson@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID})
  \item Felix Guenther \email{felixguenther1@gmx.de} (\href{https://orcid.org/0000-0001-6582-1174}{ORCID})
  \item Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})
  \item James Mba Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID})
  \item Jessalyn Sebastian \email{jessalynsebastian@gmail.com} (\href{https://orcid.org/0000-0002-1768-3229}{ORCID})
}

Other contributors:
\itemize{
  \item Hannah Choi \email{hannah.choi1@lshtm.ac.uk} [contributor]
  \item Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID}) [contributor]
  \item Joel Hellewell \email{joel@ebi.ac.uk} (\href{https://orcid.org/0000-0003-2683-0849}{ORCID}) [contributor]
  \item Luis Rivas \email{luisnicolasrivas@gmail.com} [contributor]
  \item Sang Woo Park \email{swp2@princeton.edu} (\href{https://orcid.org/0000-0003-2202-3361}{ORCID}) [contributor]
  \item Nathan McIntosh \email{natemcintosh@gmail.com} [contributor]
  \item Kath Sherratt \email{katharine.sherratt@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-2049-3423}{ORCID}) [contributor]
  \item Nikos Bosse \email{nikos.bosse@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-7750-5280}{ORCID}) [contributor]
  \item Adam Howes \email{adamthowes@gmail.com} (\href{https://orcid.org/0000-0003-2386-4031}{ORCID}) [contributor]
  \item Kaitlyn Johnson \email{johnsonkaitlyne9@gmail.com} (\href{https://orcid.org/0000-0001-8011-0012}{ORCID}) [contributor]
  \item Barbora Nemcova \email{barbora.nemcova@kit.edu} (\href{https://orcid.org/0009-0004-7565-4145}{ORCID}) [contributor]
}

}
\keyword{internal}
