% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast-package.R
\docType{package}
\name{epinowcast-package}
\alias{epinowcast-package}
\title{epinowcast: Hierarchical Nowcasting of Right Censored Epidemiological Counts}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to enable flexible and efficient hierarchical nowcasting of right censored epidemiological counts using a semi-mechanistic Bayesian method with support for both day of reference and day of report effects. Nowcasting in this context is the estimation of the total notifications (for example hospitalisations or deaths) that will be reported for a given date based on those currently reported and the pattern of reporting for previous days. This can be useful when tracking the spread of infectious disease in real-time as otherwise changes in trends can be obfuscated by partial reporting or their detection may be delayed due to the use of simpler methods like truncation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiforecasts.io/epinowcast/}
  \item \url{https://github.com/epiforecasts/epinowcast/}
  \item Report bugs at \url{https://github.com/epiforecasts/epinowcast/issues/}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Adrian Lison \email{adrian.lison@bsse.ethz.ch} (\href{https://orcid.org/0000-0002-6822-8437}{ORCID})
}

Other contributors:
\itemize{
  \item Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID}) [contributor]
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk} [contributor]
}

}
\keyword{internal}
