% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast-package.R
\docType{package}
\name{epinowcast-package}
\alias{epinowcast-package}
\title{epinowcast: Flexible Hierarchical Nowcasting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools to enable flexible and efficient hierarchical nowcasting of right-truncated epidemiological time-series using a semi-mechanistic Bayesian model with support for a range of reporting and generative processes. Nowcasting, in this context, is gaining situational awareness using currently available observations and the reporting patterns of historical observations. This can be useful when tracking the spread of infectious disease in real-time: without nowcasting, changes in trends can be obfuscated by partial reporting or their detection may be delayed due to the use of simpler methods like truncation. While the package has been designed with epidemiological applications in mind, it could be applied to any set of right-truncated time-series count data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://epiforecasts.io/epinowcast/}
  \item \url{https://github.com/epiforecasts/epinowcast/}
  \item Report bugs at \url{https://github.com/epiforecasts/epinowcast/issues/}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{sam.abbott@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Adrian Lison \email{adrian.lison@bsse.ethz.ch} (\href{https://orcid.org/0000-0002-6822-8437}{ORCID})
  \item Sebastian Funk \email{sebastian.funk@lshtm.ac.uk}
}

Other contributors:
\itemize{
  \item Michael DeWitt \email{me.dewitt.jr@gmail.com} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID}) [contributor]
  \item Hannah Choi \email{hannah.choi1@lshtm.ac.uk} [contributor]
  \item Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID}) [contributor]
}

}
\keyword{internal}
