% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast.R
\name{epinowcast}
\alias{epinowcast}
\title{Nowcast using partially observed data}
\usage{
epinowcast(
  data,
  model = epinowcast::enw_model(),
  reference = epinowcast::enw_reference(parametric = ~1, distribution = "lognormal",
    non_parametric = ~0, data = data),
  report = epinowcast::enw_report(formula = ~0, structural = ~0, data),
  expectation = epinowcast::enw_expectation(formula = ~rw(day, .group), order = 1, data
    = data),
  missing = epinowcast::enw_missing(formula = ~0, data = data),
  observation = epinowcast::enw_obs(family = "negbin"),
  fit = enw_fit_opts(fit = epinowcast::enw_sample, nowcast = TRUE, pp = FALSE,
    likelihood = TRUE, debug = FALSE, output_loglik = FALSE)
)
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Provides a user friendly interface around package functionality
to produce a nowcast from observed preprocessed data, and a series of user
defined models.
}
\seealso{
Other epinowcast: 
\code{\link{plot.epinowcast}()},
\code{\link{summary.epinowcast}()}
}
\concept{epinowcast}
