% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast.R
\name{epinowcast}
\alias{epinowcast}
\title{FUNCTION_TITLE}
\usage{
epinowcast(
  pobs,
  reference_effects = enw_intercept_model(pobs$metareference[[1]]),
  report_effects = enw_intercept_model(pobs$metareport[[1]]),
  dist = "lognormal",
  probs = c(0.05, 0.35, 0.5, 0.65, 0.95),
  model = NULL,
  nowcast = TRUE,
  pp = FALSE,
  likelihood = TRUE,
  debug = FALSE,
  output_loglik = FALSE,
  ...
)
}
\arguments{
\item{pobs}{PARAM_DESCRIPTION}

\item{reference_effects}{PARAM_DESCRIPTION,
Default: enw_intercept_model(pobs$metareference[\link{1}])}

\item{report_effects}{PARAM_DESCRIPTION,
Default: enw_intercept_model(pobs$metareport[\link{1}])}

\item{dist}{PARAM_DESCRIPTION, Default: 'lognormal'}

\item{probs}{PARAM_DESCRIPTION, Default: c(0.05, 0.35, 0.5, 0.65, 0.95)}

\item{model}{PARAM_DESCRIPTION, Default: NULL}

\item{nowcast}{PARAM_DESCRIPTION, Default: TRUE}

\item{pp}{PARAM_DESCRIPTION, Default: FALSE}

\item{likelihood}{PARAM_DESCRIPTION, Default: TRUE}

\item{debug}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\concept{nowcast}
