% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast.R
\name{epinowcast}
\alias{epinowcast}
\title{Nowcast right censored data}
\usage{
epinowcast(
  pobs,
  reference_effects = enw_intercept_model(pobs$metareference[[1]]),
  report_effects = enw_intercept_model(pobs$metareport[[1]]),
  dist = "lognormal",
  probs = c(0.05, 0.35, 0.5, 0.65, 0.95),
  model = epinowcast::enw_model(),
  nowcast = TRUE,
  pp = FALSE,
  likelihood = TRUE,
  debug = FALSE,
  output_loglik = FALSE,
  ...
)
}
\arguments{
\item{pobs}{Output from \code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{reference_effects}{A list of fixed and random design matrices
defining the date of reference model. Defaults to \code{\link[=enw_intercept_model]{enw_intercept_model()}}
which is an intercept only model.}

\item{report_effects}{A list of fixed and random design matrices
defining the date of reports model. Defaults to \code{\link[=enw_intercept_model]{enw_intercept_model()}}
which is an intercept only model.}

\item{dist}{Character string indicating the type of distribution to use for
reference date effects. The default is to use a lognormal but other options
available include: gamma distributed ("gamma").}

\item{probs}{PARAM_DESCRIPTION, Default: c(0.05, 0.35, 0.5, 0.65, 0.95)}

\item{model}{A \code{cmdstanr} model object as loaded by \code{\link[=enw_model]{enw_model()}}.}

\item{likelihood}{Logical, defaults to \code{TRUE}. Should the likelihood be
included in the model}

\item{debug}{Logical, defaults to \code{FALSE}. Should within model debug
information be returned.}

\item{output_loglik}{Logical, defaults to \code{FALSE}. Should the
log-likelihood be output. Disabling this will speed up fitting
if evaluating the model fit is not required.}

\item{...}{Additional arguments passed to \code{\link[=enw_sample]{enw_sample()}}.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Provides a user friendly interface around package functionality
to produce a nowcast from observed preprocessed data, a reference model, and
a report model.
}
\seealso{
Other epinowcast: 
\code{\link{plot.epinowcast}()},
\code{\link{summary.epinowcast}()}
}
\concept{epinowcast}
