% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{extract_obs_metadata}
\alias{extract_obs_metadata}
\title{Extract observation metadata}
\usage{
extract_obs_metadata(new_confirm, observation_indicator = NULL)
}
\arguments{
\item{new_confirm}{A data.table containing the columns: "reference_date",
"delay", ".group", "new_confirm", and "max_obs_delay".
As produced by \code{\link[=enw_preprocess_data]{enw_preprocess_data()}} in the \code{new_confirm} output with the
addition of the "max_obs_delay" column as produced by
\code{\link[=add_max_observed_delay]{add_max_observed_delay()}}.}

\item{observation_indicator}{A character string specifying the column name
in \code{new_confirm} that indicates whether an observation is observed or not.
This column should be a logical vector. If NULL (default), all observations
are considered observed.}
}
\value{
A list containing:
\itemize{
\item \code{st}: time index of each snapshot (snapshot time).
\item \code{ts}: snapshot index by time and group.
\item \code{sl}: number of reported observations per snapshot (snapshot
length).
\item \code{csl}: cumulative version of sl.
\item \code{lsl}: number of consecutive reported observations per
snapshot accounting for missing data.
\item \code{clsl}: cumulative version of lsl.
\item \code{nsl}: number of observed observations per snapshot (snapshot
length).
\item \code{cnsl}: cumulative version of nsl.
\item \code{sg}: group index of each snapshot (snapshot group).
}
}
\description{
This function extracts metadata from the provided dataset to be used in the
observation model.
}
\seealso{
Helper functions for model modules
\code{\link{add_max_observed_delay}()},
\code{\link{add_pmfs}()},
\code{\link{convolution_matrix}()},
\code{\link{enw_reference_by_report}()},
\code{\link{enw_reps_with_complete_refs}()},
\code{\link{extract_sparse_matrix}()},
\code{\link{latest_obs_as_matrix}()},
\code{\link{simulate_double_censored_pmf}()}
}
\concept{modelmodulehelpers}
