% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-module-helpers.R
\name{extract_sparse_matrix}
\alias{extract_sparse_matrix}
\title{Extract sparse matrix elements}
\usage{
extract_sparse_matrix(mat, prefix = "")
}
\arguments{
\item{mat}{A matrix to extract the sparse matrix from.}

\item{prefix}{A character string to prefix the names of the returned list.}
}
\value{
Return a list that describes the sparse matrix this includes:
\itemize{
\item \code{nw} the number of non-zero elements in the matrix.
\item \code{w} the non-zero elements of the matrix.
\item \code{nv} the number of non-zero row identifiers in the matrix.
\item \code{v} the non-zero row identifiers of the matrix.
\item \code{nu} the number of non-zero column identifiers in the matrix.
\item \code{u} the non-zero column identifiers of the matrix.
}
}
\description{
This helper function allows the extraction of a sparse matrix from a matrix
using \code{rstan::extract_sparse_parsts()} and returns these elements in a named
list for use in stan.
}
\examples{
mat <- matrix(1:9, nrow = 3)
extract_sparse_matrix(mat)
}
\seealso{
Helper functions for model modules
\code{\link{add_pmfs}()},
\code{\link{convolution_matrix}()},
\code{\link{enw_reference_by_report}()},
\code{\link{enw_reps_with_complete_refs}()},
\code{\link{extract_obs_metadata}()},
\code{\link{latest_obs_as_matrix}()},
\code{\link{simulate_double_censored_pmf}()}
}
\concept{modelmodulehelpers}
