% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{germany_covid19_hosp}
\alias{germany_covid19_hosp}
\title{Hospitalisations in Germany by date of report and reference}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1536885 rows and 5 columns.
}
\usage{
germany_covid19_hosp
}
\value{
A \code{data.table}
}
\description{
Hospitalisations in Germany by date of report and reference
}
\seealso{
Package data sets
\code{\link{enw_example}()}
}
\concept{data}
\keyword{datasets}
