% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{gp_terms}
\alias{gp_terms}
\title{Finds Gaussian process terms in a formula object}
\usage{
gp_terms(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link[lme4]{lme4} syntax (see \code{\link[=re]{re()}}), and
random walks defined using the \code{\link[=rw]{rw()}} helper function. See the Details
section below for a comprehensive explanation of the supported syntax.}
}
\value{
A character vector containing the Gaussian process terms
identified in the supplied formula.
}
\description{
This function extracts Gaussian process terms denoted
using \code{\link[=gp]{gp()}} from a formula so that they can be processed on their
own.
}
\examples{
epinowcast:::gp_terms(~ 1 + age_group + gp(week))
epinowcast:::gp_terms(~ 1 + gp(week, kernel = "se") + gp(day))
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{ar}()},
\code{\link{arima}()},
\code{\link{arima_terms}()},
\code{\link{arma}()},
\code{\link{as_string_formula}()},
\code{\link{construct_arima}()},
\code{\link{construct_gp}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{construct_secondary}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()},
\code{\link{gp}()},
\code{\link{ma}()},
\code{\link{parse_formula}()},
\code{\link{re}()},
\code{\link{remove_arima_terms}()},
\code{\link{remove_gp_terms}()},
\code{\link{remove_rw_terms}()},
\code{\link{remove_secondary_terms}()},
\code{\link{rw}()},
\code{\link{rw_terms}()},
\code{\link{secondary}()},
\code{\link{secondary_terms}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
