% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{is.Date}
\alias{is.Date}
\title{Check a vector is a Date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical
}
\description{
Checks that a vector is a date
}
\seealso{
Utility functions
\code{\link{coerce_date}()},
\code{\link{convert_cmdstan_to_rstan}()},
\code{\link{expose_stan_fns}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
