% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.Date}
\alias{is.Date}
\title{Check an object is a Date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical
}
\description{
Checks that an object is a date
}
\seealso{
Utility functions
\code{\link{aggregate_rolling_sum}()},
\code{\link{coerce_date}()},
\code{\link{coerce_dt}()},
\code{\link{date_to_numeric_modulus}()},
\code{\link{enw_get_cache}()},
\code{\link{enw_set_cache}()},
\code{\link{enw_unset_cache}()},
\code{\link{get_internal_timestep}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
