% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-design-tools.R
\name{is.Date}
\alias{is.Date}
\title{FUNCTION_TITLE}
\usage{
is.Date(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Utility functions
\code{\link{coerce_date}()},
\code{\link{convert_cmdstan_to_rstan}()},
\code{\link{expose_stan_fns}()},
\code{\link{stan_fns_as_string}()}
}
\concept{utils}
