% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse a formula into components}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link[lme4]{lme4} syntax (see \code{\link[=re]{re()}}), and
random walks defined using the \code{\link[=rw]{rw()}} helper function. See the Details
section below for a comprehensive explanation of the supported syntax.}
}
\value{
A list of formula components. These currently include:
\itemize{
\item \code{fixed}: A character vector of fixed effect terms
\item \code{random}: A list of of \link[lme4]{lme4} style random effects
\item \code{rw}: A character vector of \code{\link[=rw]{rw()}} random walk terms.
\item \code{arima}: A character vector of \code{\link[=arima]{arima()}} ARIMA(p, d, q) terms.
\item \code{gp}: A character vector of \code{\link[=gp]{gp()}} Gaussian process terms.
\item \code{secondary}: A character vector of \code{\link[=secondary]{secondary()}} stratum terms.
}
}
\description{
This function uses a series internal functions
to break an input formula into its component parts each of which
can then be handled separately. Currently supported components are
fixed effects, \link[lme4]{lme4} style random effects, and random walks
using the \code{\link[=rw]{rw()}} helper function.
}
\section{Reference}{

The random walk functions used internally by this function were
adapted from code written by J Scott (under an MIT license) as part of
the \code{epidemia} package (https://github.com/ImperialCollegeLondon/epidemia/).
}

\examples{
epinowcast:::parse_formula(~ 1 + age_group + location)

epinowcast:::parse_formula(~ 1 + age_group + (1 | location))

epinowcast:::parse_formula(~ 1 + (age_group | location))

epinowcast:::parse_formula(~ 1 + (1 | location) + rw(week, location))

epinowcast:::parse_formula(~ secondary(cases))
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{ar}()},
\code{\link{arima}()},
\code{\link{arima_terms}()},
\code{\link{arma}()},
\code{\link{as_string_formula}()},
\code{\link{construct_arima}()},
\code{\link{construct_gp}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{construct_secondary}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()},
\code{\link{gp}()},
\code{\link{gp_terms}()},
\code{\link{ma}()},
\code{\link{re}()},
\code{\link{remove_arima_terms}()},
\code{\link{remove_gp_terms}()},
\code{\link{remove_rw_terms}()},
\code{\link{remove_secondary_terms}()},
\code{\link{rw}()},
\code{\link{rw_terms}()},
\code{\link{secondary}()},
\code{\link{secondary_terms}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
