% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse a formula into components}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link{lme4} syntax (see \code{\link[=re]{re()}}), and random walks
defined using the \code{\link[=rw]{rw()}} helper function.}
}
\value{
A list of formula components. These currently include:
\itemize{
\item \code{fixed}: A character vector of fixed effect terms
\item \code{random}: A list of of \link{lme4} style random effects
\item \code{rw}: A character vector of \code{\link[=rw]{rw()}} random walk terms.
}
}
\description{
This function uses a series internal functions
to break an input formula into its component parts each of which
can then be handled separately. Currently supported components are
fixed effects, \link{lme4} style random effects, and random walks using the
\code{\link[=rw]{rw()}} helper function.
}
\section{Reference}{

The random walk functions used internally by this function were
adapted from code written by J Scott (under an MIT license) as part of
the \code{epidemia} package (https://github.com/ImperialCollegeLondon/epidemia/).
}

\examples{
epinowcast:::parse_formula(~ 1 + age_group + location)

epinowcast:::parse_formula(~ 1 + age_group + (1 | location))

epinowcast:::parse_formula(~ 1 + (age_group | location))

epinowcast:::parse_formula(~ 1 + (1 | location) + rw(week, location))
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()},
\code{\link{re}()},
\code{\link{remove_rw_terms}()},
\code{\link{rw}()},
\code{\link{rw_terms}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
