% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{parse_formula}
\alias{parse_formula}
\title{FUNCTION_TITLE}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
