% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.enw_preprocess_data}
\alias{plot.enw_preprocess_data}
\title{Plot method for enw_preprocess_data}
\usage{
\method{plot}{enw_preprocess_data}(x, type = "emp_rep_cum", delay_group_thresh = NULL, quantiles = NULL, ...)
}
\arguments{
\item{x}{A \code{data.table} of preprocessed data as produced by
\code{\link[=enw_preprocess_data]{enw_preprocess_data()}}.}

\item{type}{A character string indicating the type of the desired plot.
Currently supported options are "emp_rep_cum" which plots the empirical
cumulative reporting delay distribution for each stratum (using
\code{\link[=enw_plot_emprep_cum]{enw_plot_emprep_cum()}}), "emp_rep_frac" (using \code{\link[=enw_plot_emprep_frac]{enw_plot_emprep_frac()}}),
"emp_rep_quant" model (using \code{\link[=enw_plot_emprep_quant]{enw_plot_emprep_quant()}}), and "emp_ts_del"
(using \code{\link[=enw_plot_emp_ts_del]{enw_plot_emp_ts_del()}}).}

\item{delay_group_thresh}{A vector defining thresholds for the grouping of
reporting delays based on left-closed intervals. Smallest value should be
zero, largest should correspond to max_delay.}

\item{quantiles}{A vector of quantiles used for the "emp_rep_quant" plot.
Default is NULL, corresponding to the 0.10, 0.50, and 0.90 quantiles.}

\item{...}{Pass additional arguments to lower level plot functions.}
}
\value{
\code{ggplot2} object
}
\description{
\code{plot} method for preprocessed data of class
"enw_preprocess_data".
Creates different descriptive plots of the empirical reporting delay
distribution and time-series of reported cases per reference day.
}
\seealso{
Other epinowcast: 
\code{\link{epinowcast}()},
\code{\link{plot.epinowcast}()},
\code{\link{summary.epinowcast}()}

Plotting functions
\code{\link{enw_plot_emp_ts_del}()},
\code{\link{enw_plot_emprep_cum}()},
\code{\link{enw_plot_emprep_frac}()},
\code{\link{enw_plot_emprep_quant}()},
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_obs}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()},
\code{\link{plot.epinowcast}()}
}
\concept{epinowcast}
\concept{plot}
