% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast-methods.R
\name{plot.epinowcast}
\alias{plot.epinowcast}
\title{Plot method for epinowcast}
\usage{
\method{plot}{epinowcast}(x, type = "nowcast", ...)
}
\arguments{
\item{x}{A list of output as produced by \code{\link[=epinowcast]{epinowcast()}}.}

\item{type}{A character string indicating the type of plot required.
Currently supported options are "nowcast" which plots the nowcast
for each dataset along with the latest available observed data, and
"posterior" which plots observations reported at the time against
simulated observations from  the model.}

\item{...}{Pass additional arguments to plot functions.}
}
\value{
\code{ggplot2} object
}
\description{
\code{plot} method for class "epinowcast".
}
\seealso{
plot epinowcast

Other epinowcast: 
\code{\link{epinowcast}()},
\code{\link{summary.epinowcast}()}
}
\concept{epinowcast}
