% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast-methods.R
\name{plot.epinowcast}
\alias{plot.epinowcast}
\title{Plot method for epinowcast}
\usage{
\method{plot}{epinowcast}(x, obs = NULL, type = "nowcast", log = FALSE, ...)
}
\arguments{
\item{x}{A \code{data.table} of output as produced by \code{\link[=epinowcast]{epinowcast()}}.}

\item{obs}{A \code{data.frame} of observed data which may be passed to lower level
methods.}

\item{type}{A character string indicating the type of plot required.
Currently supported options are "nowcast" which plots the nowcast
for each dataset along with the latest available observed data (using
\code{\link[=enw_plot_nowcast_quantiles]{enw_plot_nowcast_quantiles()}}), and "posterior_prediction" which plots
observations reported at the time against simulated observations from  the
model (using \code{\link[=enw_plot_pp_quantiles]{enw_plot_pp_quantiles()}}).}

\item{log}{Logical, defaults to \code{FALSE}. Should counts be plot on the log
scale.}

\item{...}{Pass additional arguments to plot functions.}
}
\value{
\code{ggplot2} object
}
\description{
\code{plot} method for class "epinowcast".
}
\seealso{
Other epinowcast: 
\code{\link{epinowcast}()},
\code{\link{summary.epinowcast}()}

Plotting functions
\code{\link{enw_plot_nowcast_quantiles}()},
\code{\link{enw_plot_pp_quantiles}()},
\code{\link{enw_plot_quantiles}()},
\code{\link{enw_plot_theme}()}
}
\concept{epinowcast}
\concept{plot}
