% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.epinowcast}
\alias{print.epinowcast}
\title{Print method for epinowcast}
\usage{
\method{print}{epinowcast}(x, ...)
}
\arguments{
\item{x}{A \code{data.table} output from \code{\link[=epinowcast]{epinowcast()}}.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
\code{print} method for class \code{"epinowcast"}.
}
\examples{
nowcast <- enw_example("nowcast")
nowcast
}
\seealso{
Other epinowcast: 
\code{\link{epinowcast}()},
\code{\link{plot.enw_preprocess_data}()},
\code{\link{plot.epinowcast}()},
\code{\link{print.enw_preprocess_data}()},
\code{\link{print.summary.enw_preprocess_data}()},
\code{\link{summary.enw_preprocess_data}()},
\code{\link{summary.epinowcast}()}
}
\concept{epinowcast}
