% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.enw_preprocess_data}
\alias{print.summary.enw_preprocess_data}
\title{Print method for summary.enw_preprocess_data}
\usage{
\method{print}{summary.enw_preprocess_data}(x, ...)
}
\arguments{
\item{x}{A \code{summary.enw_preprocess_data} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns \code{x}.
}
\description{
\code{print} method for the output of
\code{\link[=summary.enw_preprocess_data]{summary.enw_preprocess_data()}}.
}
\seealso{
Other epinowcast: 
\code{\link{epinowcast}()},
\code{\link{plot.epinowcast}()},
\code{\link{print.enw_preprocess_data}()},
\code{\link{print.epinowcast}()},
\code{\link{summary.enw_preprocess_data}()},
\code{\link{summary.epinowcast}()}
}
\concept{epinowcast}
