% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{re}
\alias{re}
\title{Defines random effect terms using the lme4 syntax}
\usage{
re(formula)
}
\arguments{
\item{formula}{A formula in the format used by \link{lme4} to define random
effects}
}
\value{
A list to be parsed internally.
}
\description{
Defines random effect terms using the lme4 syntax
}
