% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{re}
\alias{re}
\title{Defines random effect terms using the lme4 syntax}
\usage{
re(formula)
}
\arguments{
\item{formula}{A formula in the format used by \link{lme4} to define random
effects}
}
\value{
A list to be parsed internally.
}
\description{
Defines random effect terms using the lme4 syntax
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
