% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{re}
\alias{re}
\title{Defines random effect terms using the lme4 syntax}
\usage{
re(formula)
}
\arguments{
\item{formula}{A random effect as returned by \code{\link[lme4:findbars]{lme4::findbars()}}
when a random effect is defined using the \link{lme4} syntax in
formula. Currently only simplified random effecs (i.e
LHS | RHS) are supported.}
}
\value{
A list defining the fixed and random effects of the specified
random effect
}
\description{
Defines random effect terms using the lme4 syntax
}
\examples{
form <- parse_formula(~ 1 + (1 | age_group))
re(form$random[[1]])

form <- parse_formula(~ 1 + (location | age_group))
re(form$random[[1]])
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{rw_terms}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
