% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{remove_gp_terms}
\alias{remove_gp_terms}
\title{Remove Gaussian process terms from a formula object}
\usage{
remove_gp_terms(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link[lme4]{lme4} syntax (see \code{\link[=re]{re()}}), and
random walks defined using the \code{\link[=rw]{rw()}} helper function. See the Details
section below for a comprehensive explanation of the supported syntax.}
}
\value{
A formula object with the Gaussian process terms removed.
}
\description{
This function removes Gaussian process terms denoted
using \code{\link[=gp]{gp()}} from a formula so they can be processed on their own.
}
\examples{
epinowcast:::remove_gp_terms(~ 1 + age_group + gp(week))
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{ar}()},
\code{\link{arima}()},
\code{\link{arima_terms}()},
\code{\link{arma}()},
\code{\link{as_string_formula}()},
\code{\link{construct_arima}()},
\code{\link{construct_gp}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()},
\code{\link{gp}()},
\code{\link{gp_terms}()},
\code{\link{ma}()},
\code{\link{parse_formula}()},
\code{\link{re}()},
\code{\link{remove_arima_terms}()},
\code{\link{remove_rw_terms}()},
\code{\link{rw}()},
\code{\link{rw_terms}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
