% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{remove_profiling}
\alias{remove_profiling}
\title{Remove profiling statements from a character vector representing stan code}
\usage{
remove_profiling(s)
}
\arguments{
\item{s}{Character vector representing stan code}
}
\value{
A \code{character} vector of the stan code without profiling statements
}
\description{
Remove profiling statements from a character vector representing stan code
}
