% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{remove_profiling}
\alias{remove_profiling}
\title{Remove profiling statements from a character vector representing stan code}
\usage{
remove_profiling(s)
}
\arguments{
\item{s}{Character vector representing stan code}
}
\value{
A \code{character} vector of the stan code without profiling statements
}
\description{
Remove profiling statements from a character vector representing stan code
}
\seealso{
Functions used to help convert models into the format required for stan
\code{\link{enw_formula_as_data_list}()},
\code{\link{enw_priors_as_data_list}()},
\code{\link{enw_replace_priors}()},
\code{\link{enw_sample}()},
\code{\link{write_stan_files_no_profile}()}
}
\concept{modeltools}
