% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{remove_rw_terms}
\alias{remove_rw_terms}
\title{Remove random walk terms}
\usage{
remove_rw_terms(x)
}
\arguments{
\item{x}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
This function was adapted from code written
by J Scott (under an MIT license) as part of
the \code{epidemia} package (https://github.com/ImperialCollegeLondon/epidemia/).
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{re}()},
\code{\link{rw}()},
\code{\link{split_formula_to_terms}()},
\code{\link{terms_rw}()}
}
\concept{formulatools}
