% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{rw}
\alias{rw}
\title{Adds random walks with Gaussian steps to the model.}
\usage{
rw(time, by, type = "independent")
}
\arguments{
\item{time}{Defines the random walk time period.}

\item{by}{Defines the grouping parameter used for the random walk.
If not specified no grouping is used.}

\item{type}{Character string, defaults to "independent". How should the
standard deviation of byed random walks be estimated. Currently this can
be set to be independent or dependent across groups.}
}
\value{
A list defining the time frame, group, and type with class
"enw_rw_term" that can be interpreted by \code{\link[=construct_rw]{construct_rw()}}.
}
\description{
A call to \code{rw()} can be used in the 'formula' argument of model
construction functions in the \code{epinowcast} package such as \code{\link[=enw_formula]{enw_formula()}}.
Does not evaluate arguments but instead simply passes information for use in
model construction.
}
\examples{
rw(time)

rw(time, location)

rw(time, location, type = "dependent")
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{split_formula_to_terms}()}
}
\concept{formulatools}
