% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{rw}
\alias{rw}
\title{Adds random walks with Gaussian steps to the model.}
\usage{
rw(time, by, type = "independent")
}
\arguments{
\item{time}{Defines the random walk time period.}

\item{by}{Defines the bying parameter used for the random walk.
If not specified no bying is used.}

\item{type}{Character string, defaults to "independent". How should the
standard deviation of byed random walks be estimated. Currently this can
be set to be independent by by or dependent across bys.}
}
\value{
A list to be parsed internally.
}
\description{
A call to \code{rw()} can be used in the 'formula' argument of model
construction functions in the \code{epinowcast} package. Does not evaluate
arguments but instead simply passes information for use in model
construction.
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{split_formula_to_terms}()},
\code{\link{terms_rw}()}
}
\concept{formulatools}
