% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{rw}
\alias{rw}
\title{Adds random walks with Gaussian steps to the model.}
\usage{
rw(time, group, type = "independent")
}
\arguments{
\item{time}{Defines the random walk time period.}

\item{group}{Defines the grouping parameter used for the random walk.
If not specified no grouping is used.}

\item{type}{Character string, defaults to "independent". How should the
standard deviation of grouped random walks be estimated. Currently this can
be set to be independent by group or dependent across groups.}
}
\value{
A list to be parsed internally.
}
\description{
A call to \code{rw()} can be used in the 'formula' argument of model
construction functions in the \code{epinowcast} package. Does not evaluate
arguments but instead simply passes information for use in model
construction.
}
