% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{split_formula_to_terms}
\alias{split_formula_to_terms}
\title{Split formula into individual terms}
\usage{
split_formula_to_terms(formula)
}
\arguments{
\item{formula}{DESCRIPTION}
}
\value{
A character vector of formula terms
}
\description{
Split formula into individual terms
}
\examples{
epinowcast:::split_formula_to_terms(~ 1 + age_group + location)
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw_terms}()},
\code{\link{rw}()}
}
\concept{formulatools}
