% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{split_formula_to_terms}
\alias{split_formula_to_terms}
\title{Split formula into individual terms}
\usage{
split_formula_to_terms(formula)
}
\arguments{
\item{formula}{A model formula that may use standard fixed
effects, random effects using \link{lme4} syntax (see \code{\link[=re]{re()}}), and random walks
defined using the \code{\link[=rw]{rw()}} helper function.}
}
\value{
A character vector of formula terms
}
\description{
Split formula into individual terms
}
\examples{
epinowcast:::split_formula_to_terms(~ 1 + age_group + location)
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{enw_formula}()},
\code{\link{enw_manual_formula}()},
\code{\link{parse_formula}()},
\code{\link{re}()},
\code{\link{remove_rw_terms}()},
\code{\link{rw}()},
\code{\link{rw_terms}()}
}
\concept{formulatools}
