% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-tools.R
\name{split_formula_to_terms}
\alias{split_formula_to_terms}
\title{FUNCTION_TITLE}
\usage{
split_formula_to_terms(formula)
}
\arguments{
\item{formula}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\seealso{
Functions used to help convert formulas into model designs
\code{\link{as_string_formula}()},
\code{\link{construct_re}()},
\code{\link{construct_rw}()},
\code{\link{parse_formula}()},
\code{\link{remove_rw_terms}()},
\code{\link{re}()},
\code{\link{rw}()},
\code{\link{terms_rw}()}
}
\concept{formulatools}
