% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stan_fns_as_string}
\alias{stan_fns_as_string}
\title{Read in a stan function file as a character string}
\usage{
stan_fns_as_string(files, target_dir)
}
\arguments{
\item{files}{A character vector of file names}

\item{target_dir}{A character string giving the directory in which
files can be found.}
}
\value{
A character string in the of stan functions.
}
\description{
Read in a stan function file as a character string
}
\seealso{
Utility functions
\code{\link{aggregate_rolling_sum}()},
\code{\link{coerce_date}()},
\code{\link{coerce_dt}()},
\code{\link{convert_cmdstan_to_rstan}()},
\code{\link{date_to_numeric_modulus}()},
\code{\link{expose_stan_fns}()},
\code{\link{get_internal_timestep}()},
\code{\link{is.Date}()}
}
\concept{utils}
