% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stan_fns_as_string}
\alias{stan_fns_as_string}
\title{Read in a stan function file as a character string}
\usage{
stan_fns_as_string(files, include)
}
\arguments{
\item{files}{A character vector specifying the names of Stan files to be
exposed. These must be in the \code{include} directory. Defaults to all Stan
files in the \code{include} directory. Note that the following files contain
overloaded functions and cannot be exposed: "delay_lpmf.stan",
"allocate_observed_obs.stan", "obs_lpmf.stan", and "effects_priors_lp.stan".}

\item{include}{A character string specifying the directory containing Stan
files. Defaults to the 'stan/functions' directory of the \code{\link[=epinowcast]{epinowcast()}}
package.}
}
\value{
A character string in the of stan functions.
}
\description{
Read in a stan function file as a character string
}
\seealso{
Utility functions
\code{\link{aggregate_rolling_sum}()},
\code{\link{coerce_date}()},
\code{\link{coerce_dt}()},
\code{\link{date_to_numeric_modulus}()},
\code{\link{get_internal_timestep}()},
\code{\link{is.Date}()}
}
\concept{utils}
