% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{subset_obs}
\alias{subset_obs}
\title{Subset observations data table for either modelled dates
or not-modelled earlier dates.}
\usage{
subset_obs(ord_obs, max_delay, internal_timestep, select)
}
\arguments{
\item{ord_obs}{The observations \code{data.table} to be subset,
as pulled from the result of calling epinowcast() and
coerced to a data table.}

\item{max_delay}{Whole number representing the maximum delay
in units of the timestep.}

\item{internal_timestep}{A numeric value representing the number
of days in the timestep, e.g. 7 when the timesteps are weeks.}

\item{select}{String, select reference dates from "before" the
max_delay or "after"?}
}
\value{
A \code{data.frame} subset for the desired observations
}
\description{
Subset observations data table for either modelled dates
or not-modelled earlier dates.
}
\seealso{
Functions used for postprocessing of model fits
\code{\link{build_ord_obs}()},
\code{\link{enw_add_latest_obs_to_nowcast}()},
\code{\link{enw_nowcast_samples}()},
\code{\link{enw_nowcast_summary}()},
\code{\link{enw_posterior}()},
\code{\link{enw_pp_summary}()},
\code{\link{enw_quantiles_to_long}()},
\code{\link{enw_summarise_samples}()}
}
\concept{postprocess}
