% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.epinowcast}
\alias{summary.epinowcast}
\title{Summary method for epinowcast}
\usage{
\method{summary}{epinowcast}(
  object,
  type = c("nowcast", "nowcast_samples", "fit", "posterior_prediction"),
  ...
)
}
\arguments{
\item{object}{A \code{data.table} output from \code{\link[=epinowcast]{epinowcast()}}.}

\item{type}{Character string indicating the summary to return; enforced by
\code{\link[base:match.arg]{base::match.arg()}}. Supported options are:
\itemize{
\item "nowcast" which summarises nowcast posterior with \code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}},
\item "nowcast_samples" which samples latest with \code{\link[=enw_nowcast_samples]{enw_nowcast_samples()}},
\item "fit" which returns the summarised \code{cmdstanr} fit with \code{\link[=enw_posterior]{enw_posterior()}},
\item "posterior_prediction" which returns summarised posterior predictions for
the observations after fitting using \code{\link[=enw_pp_summary]{enw_pp_summary()}}.
}}

\item{...}{Additional arguments passed to summary specified by \code{type}.}
}
\value{
A summary data.frame
}
\description{
\code{summary} method for class "epinowcast".
}
\examples{
nowcast <- enw_example("nowcast")

# Summarise nowcast posterior
summary(nowcast, type = "nowcast")

# Nowcast posterior samples
summary(nowcast, type = "nowcast_samples")

# Nowcast model fit
summary(nowcast, type = "fit")

# Posterior predictions
summary(nowcast, type = "posterior_prediction")
}
\seealso{
summary epinowcast

Other epinowcast: 
\code{\link{epinowcast}()},
\code{\link{plot.epinowcast}()}
}
\concept{epinowcast}
