% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epinowcast-methods.R
\name{summary.epinowcast}
\alias{summary.epinowcast}
\title{Summary method for epinowcast}
\usage{
\method{summary}{epinowcast}(object, type = "nowcast", ...)
}
\arguments{
\item{object}{A \code{data.table} output from \code{\link[=epinowcast]{epinowcast()}}.}

\item{type}{A character string indicating the type of summary to return.
Currently supported options are "nowcast" which summaries the nowcast
posterior using \code{\link[=enw_nowcast_summary]{enw_nowcast_summary()}}, and "fit" which returns the
summarised \code{cmdstanr} fit.}

\item{...}{Pass additional arguments to summary functions.}
}
\value{
\code{ggplot2} object
}
\description{
\code{summary} method for class "epinowcast".
}
\seealso{
summary epinowcast

Other epinowcast: 
\code{\link{epinowcast}()},
\code{\link{plot.epinowcast}()}
}
\concept{epinowcast}
