% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{write_stan_files_no_profile}
\alias{write_stan_files_no_profile}
\title{Write copies of the .stan files of a Stan model and its #include files
with all profiling statements removed.}
\usage{
write_stan_files_no_profile(
  stan_file,
  include_paths = NULL,
  target_dir = epinowcast::enw_get_cache()
)
}
\arguments{
\item{stan_file}{The path to a .stan file containing a Stan program.}

\item{include_paths}{Paths to directories where Stan should look for files
specified in #include directives in the Stan program.}

\item{target_dir}{The path to a directory in which the manipulated .stan
files without profiling statements should be stored. To avoid overriding of
the original .stan files, this should be different from the directory of the
original model and the \code{include_paths}.}
}
\value{
A \code{list} containing the path to the .stan file without profiling
statements and the include_paths for the included .stan files without
profiling statements
}
\description{
Write copies of the .stan files of a Stan model and its #include files
with all profiling statements removed.
}
\seealso{
Functions used to help convert models into the format required for stan:
\code{\link[=enw_formula_as_data_list]{enw_formula_as_data_list()}},
\code{\link[=enw_get_cache]{enw_get_cache()}},
\code{\link[=enw_model]{enw_model()}},
\code{\link[=enw_pathfinder]{enw_pathfinder()}},
\code{\link[=enw_priors_as_data_list]{enw_priors_as_data_list()}},
\code{\link[=enw_replace_priors]{enw_replace_priors()}},
\code{\link[=enw_sample]{enw_sample()}},
\code{\link[=enw_set_cache]{enw_set_cache()}},
\code{\link[=enw_stan_to_r]{enw_stan_to_r()}},
\code{\link[=enw_unset_cache]{enw_unset_cache()}},
\code{\link[=remove_profiling]{remove_profiling()}}
}
\concept{modeltools}
